Appending Axes

AxisIndices.append_axis — Function

append_axis(x, y)

Returns the appended axes x and y. New subtypes of AbstractAxis must implement a unique append_axis method.

Examples

julia> using AxisIndices

julia> append_axis(Axis(UnitMRange(1, 10)), SimpleAxis(UnitMRange(1, 10)))
Axis(UnitMRange(1:20) => UnitMRange(1:20))

julia> append_axis(SimpleAxis(UnitMRange(1, 10)), SimpleAxis(UnitMRange(1, 10)))
SimpleAxis(UnitMRange(1:20))

AxisIndices.append_axis! — Function

append_axis!(x, y)

Returns the appended axes x and y. New subtypes of AbstractAxis must implement a unique append_axis! method.

Examples

julia> using AxisIndices

julia> x, y = Axis(UnitMRange(1, 10)), SimpleAxis(UnitMRange(1, 10));

julia> append_axis!(x, y);

julia> length(x)
20

julia> append_axis!(y, x);

julia> length(y)
30

AxisIndices.append_keys — Function

append_keys(x, y)

Returns the appropriate keys of and index within the operation append_axis(x, y)

See also: append_axis

AxisIndices.append_values — Function

append_values(x, y)

Returns the appropriate values of and index within the operation append_axis(x, y)

See also: append_axis