Array Interface

Construction

Take a standard array and attach custom keys along the indices of each dimension.

julia> using AxisIndices

julia> A_base = [1 2; 3 4];

julia> A_axis = AxisArray(A_base, ["a", "b"], [:one, :two])
2×2 AxisArray{Int64,2}
 • dim_1 - ["a", "b"]
 • dim_2 - [:one, :two]
      one   two  
  a     1     2  
  b     3     4  

Note that the keys provided are converted to a subtype of AbstractAxis.

julia> axes(A_axis, 1)
Axis(["a", "b"] => Base.OneTo(2))

An AxisArray may also be initiliazed using similar syntax as Array{T}(undef, dims).

julia> A_axis = AxisArray{Int}(undef, ["a", "b"], [:one, :two]);

julia> A_axis[:,:] = A_base;

julia> A_axis
2×2 AxisArray{Int64,2}
 • dim_1 - ["a", "b"]
 • dim_2 - [:one, :two]
      one   two  
  a     1     2  
  b     3     4  

Names can be attached to each dimension/axis using NamedAxisArray.

julia> A_named_axis = NamedAxisArray{(:xdim, :ydim)}(A_axis)
2×2 NamedAxisArray{Int64,2}
 • xdim - ["a", "b"]
 • ydim - [:one, :two]
      one   two  
  a     1     2  
  b     3     4  

julia> A_named_axis == NamedAxisArray{(:xdim, :ydim)}(A_base, ["a", "b"], [:one, :two])
true

We can also attach metadata to an an array.

julia> A_meta_axis = MetaAxisArray(A_base, (["a", "b"], [:one, :two]), metadata = "a life well Steved")
2×2 MetaAxisArray{Int64,2}
 • dim_1 - ["a", "b"]
 • dim_2 - [:one, :two]
metadata: String
 • a life well Steved
      one   two
  a     1     2
  b     3     4

julia> A_meta_axis == MetaAxisArray(A_axis, metadata = "a life well Steved")
true

julia> NamedMetaAxisArray{(:xdim, :ydim)}(A_base, ["a", "b"], [:one, :two], metadata = "a life well Steved")
2×2 MetaAxisArray{(:xdim, :ydim),Int64}
 • xdim - ["a", "b"]
 • ydim - [:one, :two]
metadata: String
 • a life well Steved
      one   two
  a     1     2
  b     3     4

Indexing

Behavior of an AxisArray is similar to that of the common Array type.

julia> using AxisIndices

julia> import Unitful: s

julia> A_base = reshape(1:9, 3,3);

julia> A_axis = AxisArray(A_base, ((.1:.1:.3)s, ["a", "b", "c"]))
3×3 AxisArray{Int64,2}
 • dim_1 - 0.1 s:0.1 s:0.3 s
 • dim_2 - ["a", "b", "c"]
          a   b   c
  0.1 s   1   4   7
  0.2 s   2   5   8
  0.3 s   3   6   9

julia> A_axis[1,1] == A_base[1,1]
true

julia> A_axis[1] == A_base[1] # linear indexing works too
true

julia> A_axis[1,:]
3-element AxisArray{Int64,1}
 • dim_1 - ["a", "b", "c"]

  a   1
  b   4
  c   7

julia> A_axis[1:2, 1:2]
2×2 AxisArray{Int64,2}
 • dim_1 - 0.1 s:0.1 s:0.2 s
 • dim_2 - ["a", "b"]
          a   b
  0.1 s   1   4
  0.2 s   2   5

julia> A_axis[1:3]
3-element AxisArray{Int64,1}
 • dim_1 - 0.1 s:0.1 s:0.3 s

  0.1 s   1
  0.2 s   2
  0.3 s   3

In addition to standard indexing, an AxisArray can be indexed by its keys...

julia> A_axis[.1s, "a"]
1

julia> A_axis[0.1s..0.3s, ["a", "b"]]
3×2 AxisArray{Int64,2}
 • dim_1 - 0.1 s:0.1 s:0.3 s
 • dim_2 - ["a", "b"]
          a   b
  0.1 s   1   4
  0.2 s   2   5
  0.3 s   3   6

...or functions that filter the keys.

julia> A_axis[!=(.2s), in(["a", "c"])]
2×2 AxisArray{Int64,2}
 • dim_1 - Unitful.Quantity{Float64,𝐓,Unitful.FreeUnits{(s,),𝐓,nothing}}[0.1 s, 0.3 s]
 • dim_2 - ["a", "c"]
          a   c
  0.1 s   1   7
  0.3 s   3   9

Combining Different Axes TODO

Reference

AxisIndices.Arrays.AbstractAxisArray

AxisIndices.Arrays.AxisArray

AxisIndices.Arrays.AxisArray

AxisIndices.Arrays.AxisArray

AxisIndices.Arrays.AxisArray

AxisIndices.Arrays.AxisArray

AxisIndices.Arrays.AxisArray

AxisIndices.Arrays.AxisVector

AxisIndices.Arrays.MetaAxisArray

AxisIndices.Arrays.NamedAxisArray

AxisIndices.Arrays.NamedMetaAxisArray

AxisIndices.Arrays.OffsetArray

AxisIndices.Arrays.OffsetVector

AxisIndices.Arrays.AbstractAxisArray — Type

AbstractAxisArray

AbstractAxisArray is a subtype of AbstractArray that offers integration with the AbstractAxis interface. The only methods that absolutely needs to be defined for a subtype of AbstractAxisArray are axes, parent, similar_type, and similar. Most users should find the provided AxisArray subtype is sufficient for the majority of use cases. Although custom behavior may be accomplished through a new subtype of AbstractAxisArray, customizing the behavior of many methods described herein can be accomplished through a unique subtype of AbstractAxis.

This implementation is meant to be basic, well documented, and have sane defaults that can be overridden as necessary. In other words, default methods for manipulating arrays that return an AxisArray should not cause unexpected downstream behavior for users; and developers should be able to freely customize the behavior of AbstractAxisArray subtypes with minimal effort.

AxisIndices.Arrays.AxisArray — Type

AxisArray{T,N,P,AI}

An array struct that wraps any parent array and assigns it an AbstractAxis for each dimension. The first argument is the parent array and the second argument is a tuple of subtypes to AbstractAxis or keys that will be converted to subtypes of AbstractAxis with the provided keys.

AxisIndices.Arrays.AxisArray — Method

AxisArray(parent::AbstractArray, args...) -> AxisArray(parent, tuple(args))

Passes args to a tuple for constructing an AxisArray.

Examples

julia> using AxisIndices

julia> A = AxisArray(reshape(1:9, 3,3), 2:4, 3.0:5.0);

julia> A[1, 1]
1

julia> A[==(2), ==(3.0)]
1

julia> A[1:2, 1:2] == [1 4; 2 5]
true

julia> A[<(4), <(5.0)] == [1 4; 2 5]
true

AxisIndices.Arrays.AxisArray — Method

AxisArray(parent::AbstractArray, keys::Tuple[, values=axes(parent), check_length=true])

Given an the some array parent and a tuple of vectors keys corresponding to each dimension of parent constructs an AxisArray. Each element of keys is paired with an element of values to compose a subtype of AbstractAxis. values map the keys to the indices of parent.

Examples

julia> using AxisIndices

julia> AxisArray(ones(2,2), (["a", "b"], ["one", "two"]))
2×2 AxisArray{Float64,2}
 • dim_1 - ["a", "b"]
 • dim_2 - ["one", "two"]
      one   two
  a   1.0   1.0
  b   1.0   1.0

AxisIndices.Arrays.AxisArray — Method

AxisArray{T,N}(undef, dims::NTuple{N,Integer})
AxisArray{T,N}(undef, keys::NTuple{N,AbstractVector})

Construct an uninitialized N-dimensional array containing elements of type T were the size of each dimension is equal to the corresponding integer in dims.

Construct an uninitialized N-dimensional array containing elements of type T were the size of each dimension is determined by the length of the corresponding collection in `keys.

Examples

```jldoctest julia> using AxisIndices

julia> size(AxisArray{Int,2}(undef, (2,2))) (2, 2)

julia> size(AxisArray{Int,2}(undef, (["a", "b"], [:one, :two]))) (2, 2)

AxisIndices.Arrays.AxisArray — Method

AxisArray(parent::AbstractArray, axes::Tuple{Vararg{AbstractAxis}}[, check_length=true])

Construct an AxisArray using parent and explicit subtypes of AbstractAxis. If check_length is true then each dimension of parent's length is checked to match the length of the corresponding axis (e.g., size(parent 1) == length(axes[1]).

Examples

julia> using AxisIndices

julia> AxisArray(ones(2,2), (SimpleAxis(2), SimpleAxis(2)))
2×2 AxisArray{Float64,2}
 • dim_1 - 1:2
 • dim_2 - 1:2
        1     2
  1   1.0   1.0
  2   1.0   1.0

AxisIndices.Arrays.AxisArray — Method

AxisArray{T}(undef, keys::NTuple{N,AbstractVector})

Construct an uninitialized N-dimensional array containing elements of type T were the size of each dimension is determined by the length of the corresponding collection in `keys.

Examples

julia> using AxisIndices

julia> size(AxisArray{Int}(undef, (["a", "b"], [:one, :two])))
(2, 2)

AxisIndices.Arrays.AxisVector — Type

AxisVector

A vector whose indices have keys.

Examples

julia> using AxisIndices

julia> AxisVector([1, 2], [:a, :b])
2-element AxisArray{Int64,1}
 • dim_1 - [:a, :b]

  a   1
  b   2

AxisIndices.Arrays.MetaAxisArray — Type

MetaAxisArray

An AxisArray with metadata.

Examples

julia> using AxisIndices

julia> MetaAxisArray([1 2; 3 4], (["a", "b"], [:one, :two]), metadata = "some metadata")
2×2 MetaAxisArray{Int64,2}
 • dim_1 - ["a", "b"]
 • dim_2 - [:one, :two]
metadata: String
 • some metadata
      one   two
  a     1     2
  b     3     4

AxisIndices.Arrays.NamedAxisArray — Type

NamedAxisArray(parent::AbstractArray; kwargs...) = NamedAxisArray(parent, kwargs)
NamedAxisArray(parent::AbstractArray, axes::NamedTuple{L,AbstractAxes})

Type alias for NamedDimsArray whose parent array is a subtype of AxisArray. An abbreviated alias and constructor for NamedAxisArray. If key word arguments are provided then each key word becomes the name of a dimension and its assigned value is sent to the corresponding axis when constructing the underlying AxisArray.

Examples

julia> using AxisIndices

julia> A = NamedAxisArray{(:x, :y, :z)}(reshape(1:24, 2, 3, 4), ["a", "b"], ["one", "two", "three"], 2:5)
2×3×4 NamedAxisArray{Int64,3}
 • x - ["a", "b"]
 • y - ["one", "two", "three"]
 • z - 2:5
[x, y, z[2]] =
      one   two   three
  a     1     3       5
  b     2     4       6

[x, y, z[3]] =
      one   two   three
  a     7     9      11
  b     8    10      12

[x, y, z[4]] =
      one   two   three
  a    13    15      17
  b    14    16      18

[x, y, z[5]] =
      one   two   three
  a    19    21      23
  b    20    22      24

julia> dimnames(A)
(:x, :y, :z)

julia> axes_keys(A)
(["a", "b"], ["one", "two", "three"], 2:5)

julia> B = A["a", :, :]
3×4 NamedAxisArray{Int64,2}
 • y - ["one", "two", "three"]
 • z - 2:5
          2    3    4    5
    one   1    7   13   19
    two   3    9   15   21
  three   5   11   17   23

julia> C = B["one",:]
4-element NamedAxisArray{Int64,1}
 • z - 2:5

  2    1
  3    7
  4   13
  5   19

AxisIndices.Arrays.NamedMetaAxisArray — Type

NamedMetaAxisArray

An AxisArray with metadata and named dimensions.

AxisIndices.Arrays.OffsetArray — Type

OffsetArray

An array whose axes are all OffsetAxis

AxisIndices.Arrays.OffsetVector — Type

OffsetVector

A vector whose axis is OffsetAxis

Base.deleteat! — Method

deleteat!(a::AbstractAxisVector, arg)

Remove the items corresponding to A[arg], and return the modified a. Subsequent items are shifted to fill the resulting gap. If the axis of a is an AbstractSimpleAxis then it is shortened to match the length of a. If the

Examples

julia> using AxisIndices

julia> x = AxisArray([1, 2, 3, 4]);

julia> axes_keys(deleteat!(x, 3))
(OneToMRange(3),)

julia> x = AxisArray([1, 2, 3, 4], ["a", "b", "c", "d"]);

julia> axes_keys(deleteat!(x, "c"))
(["a", "b", "d"],)

Base.inv — Method

inv(M::AbstractAxisMatrix)

Computes the inverse of an AbstractAxisMatrixMatrix

julia> using AxisIndices, LinearAlgebra

julia> M = AxisArray([2 5; 1 3], ["a", "b"], [:one, :two]);

julia> axes_keys(inv(M))
([:one, :two], ["a", "b"])

Base.reshape — Method

reshape(A::AbstractAxisArray, shape)

Reshape the array and axes of A.

Examples

julia> using AxisIndices

julia> A = reshape(AxisArray(Vector(1:8), [:a, :b, :c, :d, :e, :f, :g, :h]), 4, 2);

julia> axes(A)
(Axis([:a, :b, :c, :d] => Base.OneTo(4)), SimpleAxis(Base.OneTo(2)))

julia> axes(reshape(A, 2, :))
(Axis([:a, :b] => Base.OneTo(2)), SimpleAxis(Base.OneTo(4)))

Base.rot180 — Method

rot180(A::AbstractAxisMatrix)

Rotate A 180 degrees, along with its axes keys.

Examples

julia> using AxisIndices

julia> a = AxisArray([1 2; 3 4], ["a", "b"], ["one", "two"]);

julia> b = rot180(a);

julia> axes_keys(b)
(["b", "a"], ["two", "one"])

julia> c = rotr90(rotr90(a));

julia> axes_keys(c)
(["b", "a"], ["two", "one"])

julia> a["a", "one"] == b["a", "one"] == c["a", "one"]
true

Base.rotl90 — Method

rotl90(A::AbstractAxisMatrix)

Rotate A left 90 degrees, along with its axes keys.

Examples

julia> using AxisIndices

julia> a = AxisArray([1 2; 3 4], ["a", "b"], ["one", "two"]);

julia> b = rotl90(a);

julia> axes_keys(b)
(["two", "one"], ["a", "b"])

julia> a["a", "one"] == b["one", "a"]
true

Base.rotr90 — Method

rotr90(A::AbstractAxisMatrix)

Rotate A right 90 degrees, along with its axes keys.

Examples

julia> using AxisIndices

julia> a = AxisArray([1 2; 3 4], ["a", "b"], ["one", "two"]);

julia> b = rotr90(a);

julia> axes_keys(b)
(["one", "two"], ["b", "a"])

julia> a["a", "one"] == b["one", "a"]
true

LinearAlgebra.diag — Method

diag(M::AbstractAxisMatrix, k::Integer=0; dim::Val=Val(1))

The kth diagonal of an AbstractAxisMatrixMatrix, M. The keyword argument dim specifies which which dimension's axis to preserve, with the default being the first dimension. This can be change by specifying dim=Val(2) instead.

julia> using AxisIndices, LinearAlgebra

julia> A = AxisArray([1 2 3; 4 5 6; 7 8 9], ["a", "b", "c"], [:one, :two, :three]);

julia> axes_keys(diag(A))
(["a", "b", "c"],)

julia> axes_keys(diag(A, 1; dim=Val(2)))
([:one, :two],)

LinearAlgebra.lq — Method

lq(A::AbstractAxisArray, args...; kwargs...)

Compute the LQ factorization of an AbstractAxisArray A.

Examples

julia> using AxisIndices, LinearAlgebra

julia> m = AxisArray([1.0 2; 3 4], (Axis(2:3 => Base.OneTo(2)), Axis(3:4 => Base.OneTo(2))));

julia> F = lq(m);

julia> keys.(axes(F.L))
(2:3, Base.OneTo(2))

julia> keys.(axes(F.Q))
(Base.OneTo(2), 3:4)

julia> keys.(axes(F.L * F.Q))
(2:3, 3:4)

LinearAlgebra.lu — Function

lu(A::AbstractAxisArray, args...; kwargs...)

Compute the LU factorization of an AbstractAxisArray A.

Examples

julia> using AxisIndices, LinearAlgebra

julia> m = AxisArray([1.0 2; 3 4], (Axis(2:3 => Base.OneTo(2)), Axis(3:4 => Base.OneTo(2))));

julia> F = lu(m);

julia> keys.(axes(F.L))
(2:3, Base.OneTo(2))

julia> keys.(axes(F.U))
(Base.OneTo(2), 3:4)

julia> keys.(axes(F.p))
(2:3,)

julia> keys.(axes(F.P))
(2:3, 2:3)

julia> keys.(axes(F.P * m))
(2:3, 3:4)

julia> keys.(axes(F.L * F.U))
(2:3, 3:4)

LinearAlgebra.qr — Method

qr(F::AbstractAxisArray, args...; kwargs...)

Compute the QR factorization of an AbstractAxisArray A.

Examples

julia> using AxisIndices, LinearAlgebra

julia> m = AxisArray([1.0 2; 3 4], (Axis(2:3 => Base.OneTo(2)), Axis(3:4 => Base.OneTo(2))));

julia> F = qr(m, Val(true));

julia> keys.(axes(F.Q))
(2:3, Base.OneTo(2))

julia> keys.(axes(F.R))
(Base.OneTo(2), 3:4)

julia> keys.(axes(F.Q * F.R))
(2:3, 3:4)

julia> keys.(axes(F.p))
(2:3,)

julia> keys.(axes(F.P))
(2:3, 2:3)

julia> keys.(axes(F.P * AxisArray([1.0 2; 3 4], (2:3, 3:4))))
(2:3, 3:4)

LinearAlgebra.svd — Method

svd(F::AbstractAxisArray, args...; kwargs...)

Compute the singular value decomposition (SVD) of an AbstractAxisArray A.

Examples

julia> using AxisIndices, LinearAlgebra

julia> m = AxisArray([1.0 2; 3 4], (Axis(2:3 => Base.OneTo(2)), Axis(3:4 => Base.OneTo(2))));

julia> F = svd(m);

julia> axes(F.U)
(Axis(2:3 => Base.OneTo(2)), SimpleAxis(Base.OneTo(2)))

julia> axes(F.V)
(Axis(3:4 => Base.OneTo(2)), SimpleAxis(Base.OneTo(2)))

julia> axes(F.Vt)
(SimpleAxis(Base.OneTo(2)), Axis(3:4 => Base.OneTo(2)))

julia> axes(F.U * Diagonal(F.S) * F.Vt)
(Axis(2:3 => Base.OneTo(2)), Axis(3:4 => Base.OneTo(2)))

Statistics.cor — Method

cor(x::AbstractAxisMatrix; dims=1, kwargs...)

Performs cor on the parent matrix of x and reconstructs a similar type with the appropriate axes.

Examples

julia> using AxisIndices, Statistics

julia> A = AxisArray([1 2 3; 4 5 6; 7 8 9], ["a", "b", "c"], [:one, :two, :three]);

julia> axes_keys(cor(A, dims = 2))
(["a", "b", "c"], ["a", "b", "c"])

julia> axes_keys(cor(A, dims = 1))
([:one, :two, :three], [:one, :two, :three])

Statistics.cov — Method

cov(x::AbstractAxisMatrix; dims=1, kwargs...)

Performs cov on the parent matrix of x and reconstructs a similar type with the appropriate axes.

Examples

julia> using AxisIndices, Statistics

julia> A = AxisArray([1 2 3; 4 5 6; 7 8 9], ["a", "b", "c"], [:one, :two, :three]);

julia> axes_keys(cov(A, dims = 2))
(["a", "b", "c"], ["a", "b", "c"])

julia> axes_keys(cov(A, dims = 1))
([:one, :two, :three], [:one, :two, :three])