The Axis Interface

The following describes the components necessary to construct and manipulate existing and new subtypes of AbstractAxis.

Introduction

The supertype to all axis types herein is the AbstractAxis, which is a subtype of AbstractUnitRange{<:Integer}.

The vocabulary can get a bit a tricky here and documentation tries to stick to these functional definitions:

axis: maps a set of keys to a set of indices.

indices: a set of integers (e.g., <:Integer) that locate the in memory locations of a elements.

keys: maps a set of any type to a set of indices

If we have a set of keys a b c and a set of indices 1 2 3 then the key a maps to the index 1. Given these definitions, the AbstractAxis differs from the classic dictionary in the following two ways:

The valtype of AbstractAxis is always an integer.

The values are always unique and continuous.

The two main axis types defined here are Axis and SimpleAxis. The standard syntax for indexing doesn't change at all for these types.

julia> using AxisIndices, Dates

julia> sa = SimpleAxis(1:10)
SimpleAxis(1:10)

julia> sa[2]
2

julia> sa[>(2)]
SimpleAxis(3:10)

julia> a = Axis(1:10)
Axis(1:10 => Base.OneTo(10))

julia> a[2]
2

julia> a[2:3]
Axis(2:3 => 2:3)

But now we can also use functions to index by the keys of an AbstractAxis.

julia> a = Axis(2.0:11.0, 1:10)
Axis(2.0:1.0:11.0 => 1:10)

julia> a[1]
1

julia> a[isequal(2.0)]
1

julia> a[>(2)]
Axis(3.0:1.0:11.0 => 2:10)

julia> a[>(2.0)]
Axis(3.0:1.0:11.0 => 2:10)

julia> a[and(>(2.0), <(8.0))]
Axis(3.0:1.0:7.0 => 2:6)

julia> sa[in(3:5)]
SimpleAxis(3:5)

This also allows certain syntax special treatment because they are obviously not referring to traditional integer based indexing.

julia> x, y, z = Axis((:one, :two, :three)), Axis(["one", "two", "three"]), Axis(Second(1):Second(1):Second(3));

julia> x[:one]
1

julia> x[:one] == y["one"] == z[Second(1)]
true

julia> x[[:one, :two]]
2-element Array{Int64,1}:
 1
 2

Note in the last example that a vector was returned instead of an AbstractAxis. An AbstractAxis is a subtype of AbstractUnitRange and therefore cannot be reformed after any operation that does not guarantee the return of another unit range. This is similar to the behavior of UnitRange in base.

How Does It Work?

This package makes a small alteration in the indexing pipeline for arrays from base. Where an array typically passes indices along a similar path to:

getindex(A::Array, i1, i2, i3) -> to_indices(A, axes(), (i1, i2, i3))
to_indices(A, (ax1, ax2, ax3), (i1, i2, i3)) ->
(to_index(A, i1), to_index(A, i2), to_index(A, i3)) -> I
_getindex(A, I)

This package does:

getindex(A::Array, i1, i2, i3) -> to_indices(A, axes(), (i1, i2, i3))
to_indices(A, (ax1, ax2, ax3), (i1, i2, i3)) ->
(to_index(ax1, i1), to_index(ax2, i2), to_index(ax3, i3)) -> I
_getindex(A, I)

This allows customizing the initial arguments given to getindex per each axis instead of by the array type or the type of i1/2/3.

Performance

Indexing CartesianAxes is comparable to that of CartesianIndices.

julia> using AxisIndices, BenchmarkTools

julia> cartaxes = CartesianAxes((Axis(2.0:5.0), Axis(1:4)));

julia> cartinds = CartesianIndices((1:4, 1:4));

julia> @btime getindex(cartaxes, 2, 2)
  20.848 ns (1 allocation: 32 bytes)
CartesianIndex(2, 2)

julia> @btime getindex(cartinds, 2, 2)
  22.317 ns (1 allocation: 32 bytes)
CartesianIndex(2, 2)

julia> @btime getindex(cartaxes, ==(3.0), 2)
  444.374 ns (7 allocations: 416 bytes)
CartesianIndex(2, 2)

Indexing LinearAxes is comparable to that of LinearIndices

julia> using AxisIndices, BenchmarkTools

julia> linaxes = LinearAxes((Axis(1.0:4.0), Axis(1:4)));

julia> lininds = LinearIndices((1:4, 1:4));

julia> @btime getindex(linaxes, 2, 2)
  18.275 ns (0 allocations: 0 bytes)
6

julia> @btime getindex(lininds, 2, 2)
  18.849 ns (0 allocations: 0 bytes)
6

julia> @btime getindex(linaxes, ==(3.0), 2)
  381.098 ns (6 allocations: 384 bytes)
7

You may notice there's significant overhead for using the filtering syntax. However, the filtering syntax takes advantage of a special type in base, Fix2. This means that we can take advantage of filtering methods that have been optimized for specific types of keys. Here we do the same thing as above but we create a function that knows it's going to perform filtering.

julia> getindex_filter(a, i1, i2) = a[==(i1), ==(i2)]
getindex_filter (generic function with 1 method)

julia> @btime getindex_filter(linaxes, 3.0, 2)
  57.216 ns (0 allocations: 0 bytes)
7

julia> linaxes2 = LinearAxes((Axis(Base.OneTo(4)), Axis(Base.OneTo(4))));

julia> @btime getindex_filter(linaxes2, 3, 2)
  22.070 ns (0 allocations: 0 bytes)
7

Indexing linaxes is much faster now that it can be optimized inside of a function call. However, it's still a little over twice as slow as normal indexing. That's largely because of the cost of searching 1.0:4.0 (which is a StepRangeLen type in this case). The second benchmark demonstrates how close we really are to standard indexing given similar range types.

Combining Axes

These methods are responsible for assisting in broadcasting operations.

AxisIndices.combine_axis — Function

combine_axis(x, y) -> collection

Returns the combination of axes x and y. This method relies on combine_values and combine_keys to form a new axis.

Examples

julia> using AxisIndices

julia> a, b = 1:10, Base.OneTo(10);

julia> combine_axis(a, SimpleAxis(b))
SimpleAxis(1:10)

julia> combine_axis(a, b)
1:10

julia> combine_axis(b, a)
1:10

julia> c = combine_axis(a, SimpleAxis(b))
SimpleAxis(1:10)

julia> d = combine_axis(SimpleAxis(b), a)
SimpleAxis(1:10)

julia> e = combine_axis(c, d)
SimpleAxis(1:10)

julia> f = combine_axis(Axis(a), a)
Axis(1:10 => 1:10)

julia> g = combine_axis(a, Axis(b))
Axis(Base.OneTo(10) => 1:10)

julia> combine_axis(f, g)
Axis(1:10 => 1:10)

AxisIndices.combine_values — Function

combine_values(x, y)

Returns the combination of the values of x and y, creating a new index. New subtypes of AbstractAxis may implement a unique combine_values method if needed. Default behavior is to use the return of promote_rule(x, y) for the type of the combined values.

AxisIndices.combine_keys — Function

combine_keys(x, y)

Returns the combination of x and y, assuming they are keys to another structure. Default behavior is to use promote_rule for converting x. This method should be overloaded to accomodate combining keys where a traditional promotion isn't appropriate.

Examples

The default behavior is to promote the first argument.

julia> using AxisIndices

julia> combine_keys(Base.OneTo(10), Base.OneTo(10))
Base.OneTo(10)

julia> combine_keys(1:10, Base.OneTo(10))
1:10

Using combine_keys allows combining axes that aren't appropriate for conventional promotion.

julia> combine_keys(1:2, string.(1:2))
2-element Array{String,1}:
 "1"
 "2"

julia> combine_keys(1:2, Symbol.(1:2))
2-element Array{Symbol,1}:
 Symbol("1")
 Symbol("2")

Reindexing Axes

AxisIndices.reindex — Function

reindex(a::AbstractAxis, inds::AbstractVector{Integer}) -> AbstractAxis

Returns an AbstractAxis of the same type as a where the keys of the axis are constructed by indexing into the keys of a with inds (keys(a)[inds]) and the values have the same first element as first(values(a)) but a length matching inds.

Examples

julia> using AxisIndices

julia> x, y, z = Axis(1:10, 2:11), Axis(1:10), SimpleAxis(1:10);

julia>  reindex(x, collect(1:2:10))
Axis([1, 3, 5, 7, 9] => 2:6)

julia> reindex(y, collect(1:2:10))
Axis([1, 3, 5, 7, 9] => Base.OneTo(5))

julia> reindex(z, collect(1:2:10))
SimpleAxis(1:5)

AxisIndices.unsafe_reindex — Function

unsafe_reindex(a::AbstractAxis, inds::AbstractVector) -> AbstractAxis

Similar to reindex this function returns an index of the same type as a but doesn't check that inds is inbounds. New subtypes of AbstractAxis must implement a unique unsafe_reindex method.

See also: reindex

Examples

julia> using AxisIndices

julia> AxisIndices.unsafe_reindex(SimpleAxis(OneToMRange(10)), 1:5)
SimpleAxis(OneToMRange(5))

julia> AxisIndices.unsafe_reindex(SimpleAxis(OneToSRange(10)), 1:5)
SimpleAxis(OneToSRange(5))

Reducing Axes

AxisIndices.reduce_axes — Function

reduce_axes(a, dims)

Returns the appropriate axes for a measure that reduces dimensions along the dimensions dims.

Example

julia> using AxisIndices

julia> reduce_axes(rand(2, 4), 2)
(Base.OneTo(2), Base.OneTo(1))

julia> reduce_axes(rand(2, 4), (1,2))
(Base.OneTo(1), Base.OneTo(1))

julia> reduce_axes(rand(2, 4), :)
()

julia> reduce_axes((Axis(1:4), Axis(1:4)), 2)
(Axis(1:4 => Base.OneTo(4)), Axis(1:1 => Base.OneTo(1)))

AxisIndices.reduce_axis — Function

reduce_axis(a)

Reduces axis a to single value. Allows custom index types to have custom behavior throughout reduction methods (e.g., sum, prod, etc.)

See also: reduce_axes

Example

julia> using AxisIndices

julia> reduce_axis(Axis(1:4))
Axis(1:1 => Base.OneTo(1))

julia> reduce_axis(1:4)
1:1

Resizing Axes

These methods help with operations that need to resize axes, either dynamically or by creating a new instance of an axis. In addition to helping with operations related to array resizing, these may be useful for managing the axis of a vector throughout a push!, pushfirst!, pop, and popfirst! operation.

AxisIndices.resize_first — Function

resize_first(x, n::Integer)

Returns a collection similar to x that grows or shrinks from the first index to be of size n.

Examples

julia> using AxisIndices

julia> x = collect(1:5);

julia> resize_first(x, 2)
2-element Array{Int64,1}:
 4
 5

julia> resize_first(x, 7)
7-element Array{Int64,1}:
 -1
  0
  1
  2
  3
  4
  5

julia> resize_first(x, 5)
5-element Array{Int64,1}:
 1
 2
 3
 4
 5

AxisIndices.resize_first! — Function

resize_first!(x, n::Integer)

Returns the collection x after growing or shrinking the first index to be of size n.

Examples

julia> using AxisIndices

julia> x = collect(1:5);

julia> resize_first!(x, 2);

julia> x
2-element Array{Int64,1}:
 4
 5

julia> resize_first!(x, 6);

julia> x
6-element Array{Int64,1}:
 0
 1
 2
 3
 4
 5

AxisIndices.resize_last — Function

resize_last(x, n::Integer)

Returns a collection similar to x that grows or shrinks from the last index to be of size n.

Examples

julia> using AxisIndices

julia> x = collect(1:5);

julia> resize_last(x, 2)
2-element Array{Int64,1}:
 1
 2

julia> resize_last(x, 7)
7-element Array{Int64,1}:
 1
 2
 3
 4
 5
 6
 7

julia>  resize_last(x, 5)
5-element Array{Int64,1}:
 1
 2
 3
 4
 5

AxisIndices.resize_last! — Function

resize_last!(x, n::Integer)

Returns the collection x after growing or shrinking the last index to be of size n.

Examples

julia> using AxisIndices

julia> x = collect(1:5);

julia> resize_last!(x, 2);

julia> x
2-element Array{Int64,1}:
 1
 2

julia> resize_last!(x, 5);

julia> x
5-element Array{Int64,1}:
 1
 2
 3
 4
 5

AxisIndices.grow_first — Function

grow_first(x, n)

Returns a collection similar to x that grows by n elements from the first index.

Examples

julia> using AxisIndices

julia> mr = UnitMRange(1, 10)
UnitMRange(1:10)

julia> grow_first(mr, 2)
UnitMRange(-1:10)

AxisIndices.grow_first! — Function

grow_first!(x, n)

Returns the collection x after growing from the first index by n elements.

Examples

julia> using AxisIndices

julia> mr = UnitMRange(1, 10)
UnitMRange(1:10)

julia> grow_first!(mr, 2);

julia> mr
UnitMRange(-1:10)

AxisIndices.grow_last — Function

grow_last(x, n)

Returns a collection similar to x that grows by n elements from the last index.

Examples

julia> using AxisIndices

julia> mr = UnitMRange(1, 10)
UnitMRange(1:10)

julia> grow_last(mr, 2)
UnitMRange(1:12)

AxisIndices.grow_last! — Function

grow_last!(x, n)

Returns the collection x after growing from the last index by n elements.

Examples

julia> using AxisIndices

julia> mr = UnitMRange(1, 10)
UnitMRange(1:10)

julia> grow_last!(mr, 2);

julia> mr
UnitMRange(1:12)

AxisIndices.shrink_first — Function

shrink_first(x, n)

Returns a collection similar to x that shrinks by n elements from the first index.

Examples

julia> using AxisIndices

julia> mr = UnitMRange(1, 10)
UnitMRange(1:10)

julia> shrink_first(mr, 2)
UnitMRange(3:10)

AxisIndices.shrink_first! — Function

shrink_first!(x, n)

Returns the collection x after shrinking from the first index by n elements.

AxisIndices.shrink_last — Function

shrink_last(x, n)

Returns a collection similar to x that shrinks by n elements from the last index.

Examples

julia> using AxisIndices

julia> mr = UnitMRange(1, 10)
UnitMRange(1:10)

julia> shrink_last(mr, 2)
UnitMRange(1:8)

AxisIndices.shrink_last! — Function

shrink_last!(x, n)

Returns the collection x after shrinking from the last index by n elements.

Examples

julia> using AxisIndices

julia> mr = UnitMRange(1, 10)
UnitMRange(1:10)

julia> shrink_last!(mr, 2);

julia> mr
UnitMRange(1:8)

AxisIndices.next_type — Function

next_type(x::T)

Returns the immediately greater value of type T.

Examples

julia> using AxisIndices

julia> AxisIndices.next_type("b")
"c"

julia> AxisIndices.next_type(:b)
:c

julia> AxisIndices.next_type('a')
'b': ASCII/Unicode U+0062 (category Ll: Letter, lowercase)

julia> AxisIndices.next_type(1)
2

julia> AxisIndices.next_type(2.0)
2.0000000000000004

julia> AxisIndices.next_type("")
""

AxisIndices.prev_type — Function

prev_type(x::T)

Returns the immediately lesser value of type T.

Examples

julia> using AxisIndices

julia> AxisIndices.prev_type("b")
"a"

julia> AxisIndices.prev_type(:b)
:a

julia> AxisIndices.prev_type('b')
'a': ASCII/Unicode U+0061 (category Ll: Letter, lowercase)

julia> AxisIndices.prev_type(1)
0

julia> AxisIndices.prev_type(1.0)
0.9999999999999999

julia> AxisIndices.prev_type("")
""

Axis Types

AxisIndices.AbstractAxis — Type

AbstractAxis

An AbstractVector subtype optimized for indexing.

AxisIndices.values_type — Function

values_type(x)

Retrieves the type of the values of x. This should be functionally equivalent to typeof(values(x)).

Examples

julia> using AxisIndices

julia>  values_type(Axis(1:2))
Base.OneTo{Int64}

julia> values_type(typeof(Axis(1:2)))
Base.OneTo{Int64}

julia> values_type(typeof(1:2))
UnitRange{Int64}

AxisIndices.keys_type — Function

keys_type(x)

Retrieves the type of the keys of x.

Examples

julia> using AxisIndices

julia>  keys_type(Axis(1:2))
UnitRange{Int64}

julia> keys_type(typeof(Axis(1:2)))
UnitRange{Int64}

julia> keys_type(UnitRange{Int})
Base.OneTo{Int64}

AxisIndices.Axis — Type

Axis(k[, v=OneTo(length(k))])

Subtypes of AbstractAxis that maps keys to values. The first argument specifies the keys and the second specifies the values. If only one argument is specified then the values span from 1 to the length of k.

Examples

The value for all of these is the same.

julia> using AxisIndices

julia> x = Axis(2.0:11.0, 1:10)
Axis(2.0:1.0:11.0 => 1:10)

julia> y = Axis(2.0:11.0)  # when only one argument is specified assume it's the keys
Axis(2.0:1.0:11.0 => Base.OneTo(10))

julia> z = Axis(1:10)
Axis(1:10 => Base.OneTo(10))

Standard indexing returns the same values

julia> x[2]
2

julia> x[2] == y[2] == z[2]
true

julia> x[1:2]
Axis(2.0:1.0:3.0 => 1:2)

julia> y[1:2]
Axis(2.0:1.0:3.0 => 1:2)

julia> z[1:2]
Axis(1:2 => 1:2)

julia> x[1:2] == y[1:2] == z[1:2]
true

Functions that return true or false may be used to search the keys for their corresponding index. The following is equivalent to the previous example.

julia> x[==(3.0)]
2

julia> x[==(3.0)] ==       # 3.0 is the 2nd key of x
       y[isequal(3.0)] ==  # 3.0 is the 2nd key of y
       z[==(2)]            # 2 is the 2nd key of z
true

julia> x[<(4.0)]  # all keys less than 4.0 are 2.0:3.0 which correspond to values 1:2
Axis(2.0:1.0:3.0 => 1:2)

julia> y[<=(3.0)]  # all keys less than or equal to 3.0 are 2.0:3.0 which correspond to values 1:2
Axis(2.0:1.0:3.0 => 1:2)

julia> z[<(3)]  # all keys less than or equal to 3 are 1:2 which correspond to values 1:2
Axis(1:2 => 1:2)

julia> x[<(4.0)] == y[<=(3.0)] == z[<(3)]
true

Notice that == returns a single value instead of a collection of all elements where the key was found to be true. This is because all keys must be unique so there can only ever be one element returned.

AxisIndices.SimpleAxis — Type

SimpleAxis(v)

Povides an AbstractAxis interface for any AbstractUnitRange, v. v will be considered both the values and keys of the return instance.

Examples

A SimpleAxis is useful for giving a standard set of indices the ability to use the filtering syntax for indexing.

julia> using AxisIndices

julia> x = SimpleAxis(2:10)
SimpleAxis(2:10)

julia> x[2]
3

julia> x[==(2)]
2

julia> x[2] == x[==(3)]
true

julia> x[>(2)]
SimpleAxis(3:10)

julia> x[>(2)]
SimpleAxis(3:10)

AxisIndices.CartesianAxes — Type

CartesianAxes

Alias for LinearIndices where indices are subtypes of AbstractAxis.

julia> using AxisIndices

julia> cartaxes = CartesianAxes((Axis(2.0:5.0), Axis(1:4)));

julia> cartinds = CartesianIndices((1:4, 1:4));

julia> cartaxes[2, 2]
CartesianIndex(2, 2)

julia> cartinds[2, 2]
CartesianIndex(2, 2)

AxisIndices.LinearAxes — Type

LinearAxes

Alias for LinearIndices where indices are subtypes of AbstractAxis.

julia> using AxisIndices

julia> linaxes = LinearAxes((Axis(2.0:5.0), Axis(1:4)));

julia> lininds = LinearIndices((1:4, 1:4));

julia> linaxes[2, 2]
6

julia> lininds[2, 2]
6