AxisArrays

References

AxisIndices.Interface.indices — Function

indices(x::AbstractUnitRange)

Returns the indices x.

Examples

julia> using AxisIndices

julia> indices(Axis(["a"], 1:1))
1:1

julia> indices(CartesianIndex(1,1))
(1, 1)

indices(x, i)

Returns the indices corresponding to the i axis

Examples

julia> using AxisIndices

julia> indices(AxisArray(ones(2,2), (2:3, 3:4)), 1)
Base.OneTo(2)

indices(x) -> Tuple

Returns the indices corresponding to all axes of x.

Examples

julia> using AxisIndices

julia> indices(AxisArray(ones(2,2), (2:3, 3:4)))
(Base.OneTo(2), Base.OneTo(2))

julia> indices(Axis(["a"], 1:1))
1:1

julia> indices(CartesianIndex(1,1))
(1, 1)

AxisIndices.Interface.indices_type — Function

indices_type(x, i)

Retrieves axis values of the ith dimension of x.

Examples

julia> using AxisIndices

julia> indices_type([1], 1)
Base.OneTo{Int64}

julia> indices_type(typeof([1]), 1)
Base.OneTo{Int64}

indices_type(x)

Retrieves the type of the values of x. This should be functionally equivalent to typeof(values(x)).

Examples

julia> using AxisIndices

julia>  indices_type(Axis(1:2))
Base.OneTo{Int64}

julia> indices_type(typeof(Axis(1:2)))
Base.OneTo{Int64}

julia> indices_type(typeof(1:2))
UnitRange{Int64}

AxisIndices.Interface.keys_type — Function

keys_type(x, i)

Retrieves axis keys of the ith dimension of x.

Examples

julia> using AxisIndices

julia> keys_type(AxisArray([1], ["a"]), 1)
Array{String,1}

keys_type(x)

Retrieves the type of the keys of x.

Examples

julia> using AxisIndices

julia> keys_type(Axis(1:2))
UnitRange{Int64}

julia> keys_type(typeof(Axis(1:2)))
UnitRange{Int64}

julia> keys_type(UnitRange{Int})
Base.OneTo{Int64}

AxisIndices.Interface.row_axis — Function

row_axis(x) -> axis

Returns the axis corresponding to the first dimension of x.

Examples

julia> using AxisIndices

julia> row_axis(AxisArray(ones(2,2), ["a", "b"], [:one, :two]))
Axis(["a", "b"] => Base.OneTo(2))

AxisIndices.Interface.row_keys — Function

row_keys(x) -> axis

Returns the keys corresponding to the first dimension of x.

Examples

julia> using AxisIndices

julia> row_keys(AxisArray(ones(2,2), ["a", "b"], [:one, :two]))
2-element Array{String,1}:
 "a"
 "b"

AxisIndices.Interface.row_type — Function

row_type(x)

Returns the type of the axis corresponding to the first dimension of x.

Examples

julia> using AxisIndices

julia> row_type(AxisArray(ones(2,2), ["a", "b"], [:one, :two]))
Axis{String,Int64,Array{String,1},Base.OneTo{Int64}}

AxisIndices.Interface.col_axis — Function

col_axis(x) -> axis

Returns the axis corresponding to the second dimension of x.

Examples

julia> using AxisIndices

julia> col_axis(AxisArray(ones(2,2), ["a", "b"], [:one, :two]))
Axis([:one, :two] => Base.OneTo(2))

AxisIndices.Interface.col_keys — Function

col_keys(x) -> axis

Returns the keys corresponding to the second dimension of x.

Examples

julia> using AxisIndices

julia> col_keys(AxisArray(ones(2,2), ["a", "b"], [:one, :two]))
2-element Array{Symbol,1}:
 :one
 :two

AxisIndices.Interface.col_type — Function

col_type(x)

Returns the type of the axis corresponding to the second dimension of x.

Examples

julia> using AxisIndices

julia> col_type(AxisArray(ones(2,2), ["a", "b"], [:one, :two]))
Axis{Symbol,Int64,Array{Symbol,1},Base.OneTo{Int64}}

AxisIndices.Axes.AbstractAxis — Type

AbstractAxis

An AbstractVector subtype optimized for indexing.

AxisIndices.Axes.Axis — Type

Axis(k[, v=OneTo(length(k))])

Subtypes of AbstractAxis that maps keys to values. The first argument specifies the keys and the second specifies the values. If only one argument is specified then the values span from 1 to the length of k.

Examples

The value for all of these is the same.

julia> using AxisIndices

julia> x = Axis(2.0:11.0, 1:10)
Axis(2.0:1.0:11.0 => 1:10)

julia> y = Axis(2.0:11.0)  # when only one argument is specified assume it's the keys
Axis(2.0:1.0:11.0 => Base.OneTo(10))

julia> z = Axis(1:10)
Axis(1:10 => Base.OneTo(10))

Standard indexing returns the same values

julia> x[2]
2

julia> x[2] == y[2] == z[2]
true

julia> x[1:2]
Axis(2.0:1.0:3.0 => 1:2)

julia> y[1:2]
Axis(2.0:1.0:3.0 => 1:2)

julia> z[1:2]
Axis(1:2 => 1:2)

julia> x[1:2] == y[1:2] == z[1:2]
true

Functions that return true or false may be used to search the keys for their corresponding index. The following is equivalent to the previous example.

julia> x[==(3.0)]
2

julia> x[==(3.0)] ==       # 3.0 is the 2nd key of x
       y[isequal(3.0)] ==  # 3.0 is the 2nd key of y
       z[==(2)]            # 2 is the 2nd key of z
true

julia> x[<(4.0)]  # all keys less than 4.0 are 2.0:3.0 which correspond to values 1:2
Axis(2.0:1.0:3.0 => 1:2)

julia> y[<=(3.0)]  # all keys less than or equal to 3.0 are 2.0:3.0 which correspond to values 1:2
Axis(2.0:1.0:3.0 => 1:2)

julia> z[<(3)]  # all keys less than or equal to 3 are 1:2 which correspond to values 1:2
Axis(1:2 => 1:2)

julia> x[<(4.0)] == y[<=(3.0)] == z[<(3)]
true

Notice that == returns a single value instead of a collection of all elements where the key was found to be true. This is because all keys must be unique so there can only ever be one element returned.

AxisIndices.Axes.SimpleAxis — Type

SimpleAxis(v)

Povides an AbstractAxis interface for any AbstractUnitRange, v. v will be considered both the values and keys of the return instance.

Examples

A SimpleAxis is useful for giving a standard set of indices the ability to use the filtering syntax for indexing.

julia> using AxisIndices, StaticRanges

julia> x = SimpleAxis(2:10)
SimpleAxis(2:10)

julia> x[2]
2

julia> x[==(2)]
2

julia> x[2] == x[==(2)]  # keys and values are same
true

julia> x[>(2)]
SimpleAxis(3:10)

julia> x[>(2)]
SimpleAxis(3:10)

julia> x[1]
ERROR: BoundsError: attempt to access 9-element SimpleAxis(2:10 => 2:10) at index [1]
[...]

AxisIndices.Axes.StructAxis — Type

StructAxis{T}

An axis that uses a structure T to form its keys. the field names of

AxisIndices.Axes.struct_view — Function

struct_view(A)

Creates a MappedArray using the StructAxis of A to identify the dimension that needs to be collapsed into a series of SubArrays as views that composed the MappedArray

AxisIndices.Axes.CartesianAxes — Type

CartesianAxes

Alias for LinearIndices where indices are subtypes of AbstractAxis.

Examples

julia> using AxisIndices

julia> cartaxes = CartesianAxes((Axis(2.0:5.0), Axis(1:4)));

julia> cartinds = CartesianIndices((1:4, 1:4));

julia> cartaxes[2, 2]
CartesianIndex(2, 2)

julia> cartinds[2, 2]
CartesianIndex(2, 2)

AxisIndices.Axes.LinearAxes — Type

LinearAxes

Alias for LinearIndices where indices are subtypes of AbstractAxis.

Examples

julia> using AxisIndices

julia> linaxes = LinearAxes((Axis(2.0:5.0), Axis(1:4)));

julia> lininds = LinearIndices((1:4, 1:4));

julia> linaxes[2, 2]
6

julia> lininds[2, 2]
6

AxisIndices.Arrays.AbstractAxisArray — Type

AbstractAxisArray

AbstractAxisArray is a subtype of AbstractArray that offers integration with the AbstractAxis interface. The only methods that absolutely needs to be defined for a subtype of AbstractAxisArray are axes, parent, similar_type, and similar. Most users should find the provided AxisArray subtype is sufficient for the majority of use cases. Although custom behavior may be accomplished through a new subtype of AbstractAxisArray, customizing the behavior of many methods described herein can be accomplished through a unique subtype of AbstractAxis.

This implementation is meant to be basic, well documented, and have sane defaults that can be overridden as necessary. In other words, default methods for manipulating arrays that return an AxisArray should not cause unexpected downstream behavior for users; and developers should be able to freely customize the behavior of AbstractAxisArray subtypes with minimal effort.

AxisIndices.Arrays.AxisArray — Type

AxisArray{T,N,P,AI}

An array struct that wraps any parent array and assigns it an AbstractAxis for each dimension. The first argument is the parent array and the second argument is a tuple of subtypes to AbstractAxis or keys that will be converted to subtypes of AbstractAxis with the provided keys.