Arrays With Axes

The following describes many of the available methods for accommodating multidimensional manipulation of types that have axes.

AbstractAxisIndices Interface

AxisIndices.AbstractAxisIndices — Type

AbstractAxisIndices

AbstractAxisIndices is a subtype of AbstractArray that offers integration with the AbstractAxis interface. The only methods that absolutely needs to be defined for a subtype of AbstractAxisIndices are axes, parent, similar_type, and similar. Most users should find the provided AxisIndicesArray subtype is sufficient for the majority of use cases. Although custom behavior may be accomplished through a new subtype of AbstractAxisIndices, customizing the behavior of many methods described herein can be accomplished through a unique subtype of AbstractAxis.

This implementation is meant to be basic, well documented, and have sane defaults that can be overridden as necessary. In other words, default methods for manipulating arrays that return an AxisIndicesArray should not cause unexpected downstream behavior for users; and developers should be able to freely customize the behavior of AbstractAxisIndices subtypes with minimal effort.

AxisIndices.AxisIndicesArray — Type

AxisIndicesArray(parent_array, tuple_of_keys) -> AxisIndicesArray(parent_array, Axis.(tuple_of_keys))
AxisIndicesArray(parent_array, tuple_of_axis) -> AxisIndicesArray

An array struct that wraps any parent array and assigns it an AbstractAxis for each dimension. The first argument is the parent array and the second argument is a tuple of subtypes to AbstractAxis or keys that will be converted to subtypes of AbstractAxis with the provided keys.

Examples

julia> using AxisIndices

julia> A = AxisIndicesArray(reshape(1:9, 3,3), (2:4, 3.0:5.0));

julia> A[1, 1]
1

julia> A[==(2), ==(3.0)]
1

julia> A[1:2, 1:2] == [1 4; 2 5]
true

julia> A[<(4), <(5.0)] == [1 4; 2 5]
true

Concatenating Arrays With Axes

AxisIndices.cat_axes — Function

cat_axes(x, y, dims) -> Tuple

Returns the appropriate axes for cat(x, y; dims). If any of dims are names then they should refer to the dimensions of x.

Examples

julia> using AxisIndices

julia> cat_axes(LinearAxes((2,3)), LinearAxes((2,3)), dims=(1,2))
(SimpleAxis(Base.OneTo(4)), SimpleAxis(Base.OneTo(6)))

AxisIndices.hcat_axes — Function

hcat_axes(x, y) -> Tuple

Returns the appropriate axes for hcat(x, y).

Examples

julia> using AxisIndices

julia> hcat_axes((Axis(1:4), Axis(1:2)), (Axis(1:4), Axis(1:2)))
(Axis(1:4 => Base.OneTo(4)), Axis(1:4 => Base.OneTo(4)))

julia> a, b = [1; 2; 3; 4; 5], [6 7; 8 9; 10 11; 12 13; 14 15];

julia> hcat_axes(a, b) == axes(hcat(a, b))
true

julia> c, d = CartesianAxes((Axis(1:5),)), CartesianAxes((Axis(1:5), Axis(1:2)));

julia> length.(hcat_axes(c, d)) == length.(hcat_axes(a, b))
true

AxisIndices.vcat_axes — Function

vcat_axes(x, y) -> Tuple

Returns the appropriate axes for vcat(x, y).

Examples

julia> using AxisIndices

julia> vcat_axes((Axis(1:2), Axis(1:4)), (Axis(1:2), Axis(1:4)))
(Axis(1:4 => Base.OneTo(4)), Axis(1:4 => Base.OneTo(4)))

julia> a, b = [1 2 3 4 5], [6 7 8 9 10; 11 12 13 14 15];

julia> vcat_axes(a, b) == axes(vcat(a, b))
true

julia> c, d = LinearAxes((1:1, 1:5,)), LinearAxes((1:2, 1:5));

julia> length.(vcat_axes(c, d)) == length.(vcat_axes(a, b))
true

AxisIndices.cat_axis — Function

cat_axis(x, y)

Returns the concatenation of the axes x and y. New subtypes of AbstractAxis must implement a unique cat_axis method.

AxisIndices.cat_values — Function

cat_values(x, y)

Returns the appropriate values of the x and y index within the operation cat_axis(x, y)

See also: cat_axis

AxisIndices.cat_keys — Function

cat_keys(x, y)

Returns the appropriate keys of the x and y index within the operation cat_axis(x, y)

See also: cat_axis

Mutating Methods

AxisIndices.append_axis — Function

append_axis(x, y)

Returns the appended axes x and y. New subtypes of AbstractAxis must implement a unique append_axis method.

Examples

julia> using AxisIndices

julia> append_axis(Axis(UnitMRange(1, 10)), SimpleAxis(UnitMRange(1, 10)))
Axis(UnitMRange(1:20) => UnitMRange(1:20))

julia> append_axis(SimpleAxis(UnitMRange(1, 10)), SimpleAxis(UnitMRange(1, 10)))
SimpleAxis(UnitMRange(1:20))

AxisIndices.append_axis! — Function

append_axis!(x, y)

Returns the appended axes x and y. New subtypes of AbstractAxis must implement a unique append_axis! method.

Examples

julia> using AxisIndices

julia> x, y = Axis(UnitMRange(1, 10)), SimpleAxis(UnitMRange(1, 10));

julia> append_axis!(x, y);

julia> length(x)
20

julia> append_axis!(y, x);

julia> length(y)
30

AxisIndices.append_keys — Function

append_keys(x, y)

Returns the appropriate keys of and index within the operation append_axis(x, y)

See also: append_axis

AxisIndices.append_values — Function

append_values(x, y)

Returns the appropriate values of and index within the operation append_axis(x, y)

See also: append_axis

Swapping Dimensions

Drop Dimensions

AxisIndices.drop_axes — Function

drop_axes(x, dims)

Returns all axes of x except for those identified by dims. Elements of dims must be unique integers or symbols corresponding to the dimensions or names of dimensions of x.

Examples

julia> using AxisIndices

julia> axs = (Axis(1:5), Axis(1:10));

julia> drop_axes(axs, 1)
(Axis(1:10 => Base.OneTo(10)),)

julia> drop_axes(axs, 2)
(Axis(1:5 => Base.OneTo(5)),)

julia> drop_axes(rand(2, 4), 2)
(Base.OneTo(2),)

Permute Dimensions

AxisIndices.permute_axes — Function

permute_axes(x::AbstractArray, p::Tuple) = permute_axes(axes(x), p)
permute_axes(x::NTuple{N}, p::NTuple{N}) -> NTuple{N}

Returns axes of x in the order of p.

Examples

julia> using AxisIndices

julia> permute_axes(rand(2, 4, 6), (1, 3, 2))
(Base.OneTo(2), Base.OneTo(6), Base.OneTo(4))

julia> permute_axes((Axis(1:2), Axis(1:4), Axis(1:6)), (1, 3, 2))
(Axis(1:2 => Base.OneTo(2)), Axis(1:6 => Base.OneTo(6)), Axis(1:4 => Base.OneTo(4)))

permute_axes(x::AbstractVector)

Returns the permuted axes of x as axes of size 1 × length(x)

Examples

julia> using AxisIndices

julia> length.(permute_axes(rand(4))) == (1, 4)
true

julia> permute_axes((Axis(1:4),))
(SimpleAxis(Base.OneTo(1)), Axis(1:4 => Base.OneTo(4)))

julia> permute_axes((Axis(mrange(1, 4)),))
(SimpleAxis(OneToMRange(1)), Axis(UnitMRange(1:4) => OneToMRange(4)))

julia> permute_axes((Axis(srange(1, 4)),))
(SimpleAxis(OneToSRange(1)), Axis(UnitSRange(1:4) => OneToSRange(4)))

permute_axes(m::AbstractMatrix) -> NTuple{2}

Permute the axes of the matrix m, by flipping the elements across the diagonal of the matrix. Differs from LinearAlgebra's transpose in that the operation is not recursive.

Examples

julia> using AxisIndices

julia> permute_axes(rand(4, 2))
(Base.OneTo(2), Base.OneTo(4))

julia> permute_axes((Axis(1:4), Axis(1:2)))
(Axis(1:2 => Base.OneTo(2)), Axis(1:4 => Base.OneTo(4)))

Covariance and Correlation

AxisIndices.covcor_axes — Function

covcor_axes(x, dim) -> NTuple{2}

Returns appropriate axes for a cov or var method on array x.

Examples

julia> using AxisIndices

julia> covcor_axes(rand(2,4), 1)
(Base.OneTo(4), Base.OneTo(4))

julia> covcor_axes((Axis(1:4), Axis(1:6)), 2)
(Axis(1:4 => Base.OneTo(4)), Axis(1:4 => Base.OneTo(4)))

julia> covcor_axes((Axis(1:4), Axis(1:4)), 1)
(Axis(1:4 => Base.OneTo(4)), Axis(1:4 => Base.OneTo(4)))

Each axis is resized to equal to the smallest sized dimension if given a dimensional argument greater than 2.

julia> covcor_axes((Axis(2:4), Axis(3:4)), 3)
(Axis(2:3 => Base.OneTo(2)), Axis(3:4 => Base.OneTo(2)))

Linear Algebra

Inverse Array

AxisIndices.inverse_axes — Function

inverse_axes(a::AbstractMatrix) = inverse_axes(axes(a))
inverse_axes(a::Tuple{I1,I2}) -> Tuple{I2,I1}

Returns the inverted axes of a, corresponding to the inv method from the LinearAlgebra package in the standard library.

Examples

julia> using AxisIndices

julia> inverse_axes(rand(2,4))
(Base.OneTo(4), Base.OneTo(2))

julia> inverse_axes((Axis(1:2), Axis(1:4)))
(Axis(1:4 => Base.OneTo(4)), Axis(1:2 => Base.OneTo(2)))

Matrix Multiplication

AxisIndices.matmul_axes — Function

matmul_axes(a, b) -> Tuple

Returns the appropriate axes for the return of a * b where a and b are a vector or matrix.

Examples

julia> using AxisIndices

julia> axs2, axs1 = (Axis(1:2), Axis(1:4)), (Axis(1:6),);

julia> matmul_axes(axs2, axs2)
(Axis(1:2 => Base.OneTo(2)), Axis(1:4 => Base.OneTo(4)))

julia> matmul_axes(axs1, axs2)
(Axis(1:6 => Base.OneTo(6)), Axis(1:4 => Base.OneTo(4)))

julia> matmul_axes(axs2, axs1)
(Axis(1:2 => Base.OneTo(2)),)

julia> matmul_axes(axs1, axs1)
()

julia> matmul_axes(rand(2, 4), rand(4, 2))
(Base.OneTo(2), Base.OneTo(2))

julia> matmul_axes(CartesianAxes((2,4)), CartesianAxes((4, 2))) == matmul_axes(rand(2, 4), rand(4, 2))
true

Diagonal

AxisIndices.diagonal_axes — Function

diagonal_axes(x::Tuple{<:AbstractAxis,<:AbstractAxis}) -> collection

Determines the appropriate axis for the resulting vector from a call to diag(::AxisIndicesMatrix). The default behavior is to place the smallest axis at the beginning of a call to combine_axis (e.g., combine_axis(small_axis, big_axis)).

Examples

julia> using AxisIndices

julia> AxisIndices.diagonal_axes((Axis(string.(2:5)), SimpleAxis(1:2)))
Axis(["1", "2"] => UnitMRange(1:2))

julia> AxisIndices.diagonal_axes((SimpleAxis(1:3), Axis(string.(2:5))))
Axis(["1", "2", "3"] => UnitMRange(1:3))

Factorizations

AxisIndices.get_factorization — Function

get_factorization(F::Factorization, A::AbstractAxisIndices, d::Symbol)

Used internally to compose an AxisIndicesArray for each component of a factor decomposition.

QR Factorization

julia> using AxisIndices, LinearAlgebra

julia> m = AxisIndicesArray([1.0 2; 3 4], (Axis(2:3 => Base.OneTo(2)), Axis(3:4 => Base.OneTo(2))));

julia> F = qr(m, Val(true));

julia> keys.(axes(F.Q))
(2:3, Base.OneTo(2))

julia> keys.(axes(F.R))
(Base.OneTo(2), 3:4)

julia> keys.(axes(F.Q * F.R))
(2:3, 3:4)

julia> keys.(axes(F.p))
(2:3,)

julia> keys.(axes(F.P))
(2:3, 2:3)

julia> keys.(axes(F.P * AxisIndicesArray([1.0 2; 3 4], (2:3, 3:4))))
(2:3, 3:4)

LU Factorization

julia> F = lu(m);

julia> keys.(axes(F.L))
(2:3, Base.OneTo(2))

julia> keys.(axes(F.U))
(Base.OneTo(2), 3:4)

julia> keys.(axes(F.p))
(2:3,)

julia> keys.(axes(F.P))
(2:3, 2:3)

julia> keys.(axes(F.P * m))
(2:3, 3:4)

julia> keys.(axes(F.L * F.U))
(2:3, 3:4)

LQ Factorization

julia> F = lq(m);

julia> keys.(axes(F.L))
(2:3, Base.OneTo(2))

julia> keys.(axes(F.Q))
(Base.OneTo(2), 3:4)

julia> keys.(axes(F.L * F.Q))
(2:3, 3:4)

SVD Factorization

julia> F = svd(m);

julia> axes(F.U)
(Axis(2:3 => Base.OneTo(2)), SimpleAxis(Base.OneTo(2)))

julia> axes(F.V)
(Axis(3:4 => Base.OneTo(2)), SimpleAxis(Base.OneTo(2)))

julia> axes(F.Vt)
(SimpleAxis(Base.OneTo(2)), Axis(3:4 => Base.OneTo(2)))

julia> axes(F.U * Diagonal(F.S) * F.Vt)
(Axis(2:3 => Base.OneTo(2)), Axis(3:4 => Base.OneTo(2)))