Broadcasting Axes

There are three things that determine the output of a broadcasting axes.

Standard promotion

Key specific promotion (or promotion over Real)

Order of arguments promotion

Take the following vectors.

julia> using AxisIndices

julia> a = ones(3);

julia> b = AxisIndicesArray(a, 2:4);

julia> c = AxisIndicesArray(a, ["1", "2", "3"]);

We assume the lack of a formal axis defined in a indicates that it's keys are unimportant. Therefore, we freely choose any keys that are formally defined elsewhere to belong to the new result of a broadcast statement. This is the kind of behavior one gets from standard type promotion.

julia> axes_keys(a .+ b)
(UnitMRange(2:4),)

julia> axes_keys(b .+ a)
(UnitMRange(2:4),)

julia> axes_keys(a .+ c)
(["1", "2", "3"],)

julia> axes_keys(c .+ a)
(["1", "2", "3"],)

However, when we do the same with b and c they both have formally defined keys.

julia> axes_keys(b .+ c)
(["2", "3", "4"],)

julia> axes_keys(c .+ b)
(["1", "2", "3"],)

Notice that all outputs become elements of String and a collection that are Vector. We can't depend upon the default promotion in base (e.g., promote(x, y)) because some keys won't have promote_rule defined for two key types. It may not be appropriate to define a promotion rule between something like Int and String in the base Julia library because there is not universally meaningful way to promote it without context. In this instance we clearly want to apply some sort of label along an axis and that label may or may not be intended to parse as an Int, so we always default to broadcasting the key type that is not a subtype of Real.

julia> d = AxisIndicesArray(a, [:a, :b, :c]);

julia> axes_keys(a .+ d, 1) == [:a, :b, :c]
true

The term key specific promotion is used because it is only sensible in the context of keys.

Finally, if both key types are non Real then order of arguments determines promotion.

julia> axes_keys(c .+ d)
(["1", "2", "3"],)

julia> axes_keys(d .+ c, 1) == [:a, :b, :c]
true

Customizing broadcasting behavior should be accomplished through either AxisIndices.CombineStyle or AxisIndices.broadcast_axis.

AxisIndices.broadcast_axis — Function

broadcast_axis(x, y) -> collection
broadcast_axis(::PromoteStyle, ::CombineStyle, x, y)

Returns an axis given the axes x and y that are appropriate for a broadcasting operation.

Examples

julia> using AxisIndices

julia> a, b = 1:10, Base.OneTo(10);

julia> AxisIndices.broadcast_axis(a, SimpleAxis(b))
SimpleAxis(1:10)

julia> AxisIndices.broadcast_axis(a, b)
1:10

julia> AxisIndices.broadcast_axis(b, a)
1:10

julia> c = AxisIndices.broadcast_axis(a, SimpleAxis(b))
SimpleAxis(1:10)

julia> d = AxisIndices.broadcast_axis(SimpleAxis(b), a)
SimpleAxis(1:10)

julia> e = AxisIndices.broadcast_axis(c, d)
SimpleAxis(1:10)

julia> f = AxisIndices.broadcast_axis(Axis(a), a)
Axis(1:10 => 1:10)

julia> g = AxisIndices.broadcast_axis(a, Axis(b))
Axis(Base.OneTo(10) => 1:10)

julia> AxisIndices.broadcast_axis(f, g)
Axis(1:10 => 1:10)

julia> AxisIndices.broadcast_axis(Base.OneTo(10), Base.OneTo(10))
Base.OneTo(10)

julia> AxisIndices.broadcast_axis(1:10, Base.OneTo(10))
1:10

Using combine allows combining axes that aren't appropriate for conventional promotion.

julia> AxisIndices.broadcast_axis(1:2, string.(1:2))
2-element Array{String,1}:
 "1"
 "2"

julia> AxisIndices.broadcast_axis(1:2, Symbol.(1:2))
2-element Array{Symbol,1}:
 Symbol("1")
 Symbol("2")