Combining Axes

These methods are responsible for assisting in broadcasting operations.

AxisIndices.combine_axis — Function

combine_axis(x, y) -> collection

Returns the combination of axes x and y. This method relies on combine_values and combine_keys to form a new axis.

Examples

julia> using AxisIndices

julia> a, b = 1:10, Base.OneTo(10);

julia> combine_axis(a, SimpleAxis(b))
SimpleAxis(1:10)

julia> combine_axis(a, b)
1:10

julia> combine_axis(b, a)
1:10

julia> c = combine_axis(a, SimpleAxis(b))
SimpleAxis(1:10)

julia> d = combine_axis(SimpleAxis(b), a)
SimpleAxis(1:10)

julia> e = combine_axis(c, d)
SimpleAxis(1:10)

julia> f = combine_axis(Axis(a), a)
Axis(1:10 => 1:10)

julia> g = combine_axis(a, Axis(b))
Axis(Base.OneTo(10) => 1:10)

julia> combine_axis(f, g)
Axis(1:10 => 1:10)

AxisIndices.combine_values — Function

combine_values(x, y)

Returns the combination of the values of x and y, creating a new index. New subtypes of AbstractAxis may implement a unique combine_values method if needed. Default behavior is to use the return of promote_rule(x, y) for the type of the combined values.

AxisIndices.combine_keys — Function

combine_keys(x, y)

Returns the combination of x and y, assuming they are keys to another structure. Default behavior is to use promote_rule for converting x. This method should be overloaded to accomodate combining keys where a traditional promotion isn't appropriate.

Examples

The default behavior is to promote the first argument.

julia> using AxisIndices

julia> combine_keys(Base.OneTo(10), Base.OneTo(10))
Base.OneTo(10)

julia> combine_keys(1:10, Base.OneTo(10))
1:10

Using combine_keys allows combining axes that aren't appropriate for conventional promotion.

julia> combine_keys(1:2, string.(1:2))
2-element Array{String,1}:
 "1"
 "2"

julia> combine_keys(1:2, Symbol.(1:2))
2-element Array{Symbol,1}:
 Symbol("1")
 Symbol("2")