Concatenating Axes

AxisIndices.cat_axes — Function

cat_axes(x, y, dims) -> Tuple

Returns the appropriate axes for cat(x, y; dims). If any of dims are names then they should refer to the dimensions of x.

Examples

julia> using AxisIndices

julia> cat_axes(LinearAxes((2,3)), LinearAxes((2,3)), dims=(1,2))
(SimpleAxis(Base.OneTo(4)), SimpleAxis(Base.OneTo(6)))

AxisIndices.hcat_axes — Function

hcat_axes(x, y) -> Tuple

Returns the appropriate axes for hcat(x, y).

Examples

julia> using AxisIndices

julia> hcat_axes((Axis(1:4), Axis(1:2)), (Axis(1:4), Axis(1:2)))
(Axis(1:4 => Base.OneTo(4)), Axis(1:4 => Base.OneTo(4)))

julia> a, b = [1; 2; 3; 4; 5], [6 7; 8 9; 10 11; 12 13; 14 15];

julia> hcat_axes(a, b) == axes(hcat(a, b))
true

julia> c, d = CartesianAxes((Axis(1:5),)), CartesianAxes((Axis(1:5), Axis(1:2)));

julia> length.(hcat_axes(c, d)) == length.(hcat_axes(a, b))
true

AxisIndices.vcat_axes — Function

vcat_axes(x, y) -> Tuple

Returns the appropriate axes for vcat(x, y).

Examples

julia> using AxisIndices

julia> vcat_axes((Axis(1:2), Axis(1:4)), (Axis(1:2), Axis(1:4)))
(Axis(1:4 => Base.OneTo(4)), Axis(1:4 => Base.OneTo(4)))

julia> a, b = [1 2 3 4 5], [6 7 8 9 10; 11 12 13 14 15];

julia> vcat_axes(a, b) == axes(vcat(a, b))
true

julia> c, d = LinearAxes((1:1, 1:5,)), LinearAxes((1:2, 1:5));

julia> length.(vcat_axes(c, d)) == length.(vcat_axes(a, b))
true

AxisIndices.cat_axis — Function

cat_axis(x, y)

Returns the concatenation of the axes x and y. New subtypes of AbstractAxis must implement a unique cat_axis method.

AxisIndices.cat_values — Function

cat_values(x, y)

Returns the appropriate values of the x and y index within the operation cat_axis(x, y)

See also: cat_axis

AxisIndices.cat_keys — Function

cat_keys(x, y)

Returns the appropriate keys of the x and y index within the operation cat_axis(x, y)

See also: cat_axis