Concatenating Axes

Concatenating axes can happen by one of the following:

Stacking two collections of completely unique elements

Resizing a range of values

The second only happens when the two collections provided are subtypes of AbstractRange.

Customizing concatenating axes should be accomplished through either AxisIndices.CombineStyle or AxisIndices.cat_axis!.

AxisIndices.cat_axis — Function

cat_axis(x, y) -> cat_axis(CombineStyle(x, y), x, y)
cat_axis(::CombineStyle, x, y) -> collection

Returns the concatenated axes x and y.

Examples

julia> using AxisIndices

julia> AxisIndices.cat_axis(Axis(UnitMRange(1, 10)), SimpleAxis(UnitMRange(1, 10)))
Axis(UnitMRange(1:20) => UnitMRange(1:20))

julia> AxisIndices.cat_axis(SimpleAxis(UnitMRange(1, 10)), SimpleAxis(UnitMRange(1, 10)))
SimpleAxis(UnitMRange(1:20))

AxisIndices.hcat_axes — Function

hcat_axes(x, y) -> Tuple

Returns the appropriate axes for hcat(x, y).

Examples

julia> using AxisIndices

julia> AxisIndices.hcat_axes((Axis(1:4), Axis(1:2)), (Axis(1:4), Axis(1:2)))
(Axis(1:4 => Base.OneTo(4)), Axis(1:4 => Base.OneTo(4)))

julia> a, b = [1; 2; 3; 4; 5], [6 7; 8 9; 10 11; 12 13; 14 15];

julia> AxisIndices.hcat_axes(a, b) == axes(hcat(a, b))
true

julia> c, d = CartesianAxes((Axis(1:5),)), CartesianAxes((Axis(1:5), Axis(1:2)));

julia> length.(AxisIndices.hcat_axes(c, d)) == length.(AxisIndices.hcat_axes(a, b))
true

AxisIndices.vcat_axes — Function

vcat_axes(x, y) -> Tuple

Returns the appropriate axes for vcat(x, y).

Examples

julia> using AxisIndices

julia> AxisIndices.vcat_axes((Axis(1:2), Axis(1:4)), (Axis(1:2), Axis(1:4)))
(Axis(1:4 => Base.OneTo(4)), Axis(1:4 => Base.OneTo(4)))

julia> a, b = [1 2 3 4 5], [6 7 8 9 10; 11 12 13 14 15];

julia> AxisIndices.vcat_axes(a, b) == axes(vcat(a, b))
true

julia> c, d = LinearAxes((1:1, 1:5,)), LinearAxes((1:2, 1:5));

julia> length.(AxisIndices.vcat_axes(c, d)) == length.(AxisIndices.vcat_axes(a, b))
true