Experimental

Currently there are several experimental modules included. Methods from these modules are not exported by default but may be accessed by using AxisIndices.<NameOfExperimentalModule> These are considered experimental because the are either undergoing rapid change, are likely to end up in a different package entirely, or both.

ObservationDims

AxisIndices.ObservationDims.has_obsdim — Method

has_obsdim(x) -> Bool

Returns true if x has a dimension corresponding to obs.

AxisIndices.ObservationDims.nobs — Method

nobs(x) -> Int

Returns the size along the dimension corresponding to the obs.

AxisIndices.ObservationDims.obs_axis — Method

obs_axis(x)

Returns the axis corresponding to the obs dimension.

AxisIndices.ObservationDims.obs_axis_type — Method

obs_axis_type(x)

Returns the key type corresponding to the obs axis.

AxisIndices.ObservationDims.obs_indices — Method

obs_indices(x)

Returns the indices corresponding to the obs axis

AxisIndices.ObservationDims.obs_keys — Method

obs_keys(x)

Returns the keys corresponding to the obs axis

AxisIndices.ObservationDims.obsdim — Method

obsdim(x) -> Int

Returns the dimension corresponding to the obs.

AxisIndices.ObservationDims.select_obsdim — Method

select_obsdim(x, i)

Return a view of all the data of x where the index for the obs dimension equals i.

AxisIndices.ObservationDims.each_obs — Method

each_obs(x)

Create a generator that iterates over the obs dimensions A, returning views that select all the data from the other dimensions in A.