Internal API

AxisIndices.AxisCore.unsafe_reconstruct — Function

unsafe_reconstruct(axis::AbstractAxis, keys::Ks, values::Vs)

Reconstructs an AbstractAxis of the same type as axis but with keys of type Ks and values of type Vs. This method is considered unsafe because it bypasses checks to ensure that keys and values have the same length and the all keys are unique.

unsafe_reconstruct(axis::AbstractSimpleAxis, values::Vs)

Reconstructs an AbstractSimpleAxis of the same type as axis but values of type Vs.

unsafe_reconstruct(A::AbstractAxisIndices, parent, axes)

Reconstructs an AbstractAxisIndices of the same type as A but with the parent array parent and axes axes. This method depends on an underlying call to similar_types. It is considered unsafe because it bypasses safety checks to ensure the keys of each axis are unique and match the length of each dimension of parent. Therefore, this is not intended for interactive use and should only be used when it is clear all arguments are composed correctly.

AxisIndices.AxisCore.step_key — Function

step_key(x)

Returns the step size of the keys of x.

Examples

julia> using AxisIndices

julia> AxisIndices.step_key(Axis(1:2:10))
2

julia> AxisIndices.step_key(rand(2))
1

julia> AxisIndices.step_key([1])  # LinearIndices are treate like unit ranges
1

AxisIndices.AxisCore.values_type — Function

values_type(x)

Retrieves the type of the values of x. This should be functionally equivalent to typeof(values(x)).

Examples

julia> using AxisIndices

julia>  values_type(Axis(1:2))
Base.OneTo{Int64}

julia> values_type(typeof(Axis(1:2)))
Base.OneTo{Int64}

julia> values_type(typeof(1:2))
UnitRange{Int64}

values_type(x, i)

Retrieves axis values of the ith dimension of x.

Examples

julia> using AxisIndices

julia> values_type([1], 1)
Base.OneTo{Int64}

julia> values_type(typeof([1]), 1)
Base.OneTo{Int64}

AxisIndices.Basics.permute_axes — Function

permute_axes(x::AbstractArray, perms::Tuple) = permute_axes(axes(x), p)
permute_axes(x::NTuple{N}, perms::NTuple{N}) -> NTuple{N}

Returns axes of x in the order of p.

Examples

julia> using AxisIndices

julia> AxisIndices.permute_axes(rand(2, 4, 6), (1, 3, 2))
(Base.OneTo(2), Base.OneTo(6), Base.OneTo(4))

julia> AxisIndices.permute_axes((Axis(1:2), Axis(1:4), Axis(1:6)), (1, 3, 2))
(Axis(1:2 => Base.OneTo(2)), Axis(1:6 => Base.OneTo(6)), Axis(1:4 => Base.OneTo(4)))

permute_axes(x::AbstractVector)

Returns the permuted axes of x as axes of size 1 × length(x)

Examples

julia> using AxisIndices

julia> length.(AxisIndices.permute_axes(rand(4))) == (1, 4)
true

julia> AxisIndices.permute_axes((Axis(1:4),))
(SimpleAxis(Base.OneTo(1)), Axis(1:4 => Base.OneTo(4)))

julia> AxisIndices.permute_axes((Axis(mrange(1, 4)),))
(SimpleAxis(OneToMRange(1)), Axis(UnitMRange(1:4) => OneToMRange(4)))

julia> AxisIndices.permute_axes((Axis(srange(1, 4)),))
(SimpleAxis(OneToSRange(1)), Axis(UnitSRange(1:4) => OneToSRange(4)))

permute_axes(m::AbstractMatrix) -> NTuple{2}

Permute the axes of the matrix m, by flipping the elements across the diagonal of the matrix. Differs from LinearAlgebra's transpose in that the operation is not recursive.

Examples

julia> using AxisIndices

julia> AxisIndices.permute_axes(rand(4, 2))
(Base.OneTo(2), Base.OneTo(4))

julia> AxisIndices.permute_axes((Axis(1:4), Axis(1:2)))
(Axis(1:2 => Base.OneTo(2)), Axis(1:4 => Base.OneTo(4)))

AxisIndices.Basics.append_axis! — Function

append_axis!(x, y)
append_axis!(::CombineStyle, x, y)

Returns the appended axes x and y.

Examples

julia> using AxisIndices

julia> x, y = Axis(UnitMRange(1, 10)), SimpleAxis(UnitMRange(1, 10));

julia> AxisIndices.append_axis!(x, y);

julia> length(x)
20

julia> AxisIndices.append_axis!(y, x);

julia> length(y)
30

AxisIndices.Basics.broadcast_axis — Function

broadcast_axis(x, y) -> collection
broadcast_axis(::PromoteStyle, ::CombineStyle, x, y)

Returns an axis given the axes x and y that are appropriate for a broadcasting operation.

Examples

julia> using AxisIndices

julia> a, b = 1:10, Base.OneTo(10);

julia> AxisIndices.broadcast_axis(a, SimpleAxis(b))
SimpleAxis(1:10)

julia> AxisIndices.broadcast_axis(a, b)
1:10

julia> AxisIndices.broadcast_axis(b, a)
1:10

julia> c = AxisIndices.broadcast_axis(a, SimpleAxis(b))
SimpleAxis(1:10)

julia> d = AxisIndices.broadcast_axis(SimpleAxis(b), a)
SimpleAxis(1:10)

julia> e = AxisIndices.broadcast_axis(c, d)
SimpleAxis(1:10)

julia> f = AxisIndices.broadcast_axis(Axis(a), a)
Axis(1:10 => 1:10)

julia> g = AxisIndices.broadcast_axis(a, Axis(b))
Axis(Base.OneTo(10) => 1:10)

julia> AxisIndices.broadcast_axis(f, g)
Axis(1:10 => 1:10)

julia> AxisIndices.broadcast_axis(Base.OneTo(10), Base.OneTo(10))
Base.OneTo(10)

julia> AxisIndices.broadcast_axis(1:10, Base.OneTo(10))
1:10

Using combine allows combining axes that aren't appropriate for conventional promotion.

julia> AxisIndices.broadcast_axis(1:2, string.(1:2))
2-element Array{String,1}:
 "1"
 "2"

julia> AxisIndices.broadcast_axis(1:2, Symbol.(1:2))
2-element Array{Symbol,1}:
 Symbol("1")
 Symbol("2")

AxisIndices.Math.matmul_axes — Function

matmul_axes(a, b) -> Tuple

Returns the appropriate axes for the return of a * b where a and b are a vector or matrix.

Examples

julia> using AxisIndices

julia> axs2, axs1 = (Axis(1:2), Axis(1:4)), (Axis(1:6),);

julia> AxisIndices.matmul_axes(axs2, axs2)
(Axis(1:2 => Base.OneTo(2)), Axis(1:4 => Base.OneTo(4)))

julia> AxisIndices.matmul_axes(axs1, axs2)
(Axis(1:6 => Base.OneTo(6)), Axis(1:4 => Base.OneTo(4)))

julia> AxisIndices.matmul_axes(axs2, axs1)
(Axis(1:2 => Base.OneTo(2)),)

julia> AxisIndices.matmul_axes(axs1, axs1)
()

julia> AxisIndices.matmul_axes(rand(2, 4), rand(4, 2))
(SimpleAxis(Base.OneTo(2)), SimpleAxis(Base.OneTo(2)))

julia> AxisIndices.matmul_axes(CartesianAxes((2,4)), CartesianAxes((4, 2))) == AxisIndices.matmul_axes(rand(2, 4), rand(4, 2))
true

AxisIndices.Basics.drop_axes — Function

drop_axes(x, dims)

Returns all axes of x except for those identified by dims. Elements of dims must be unique integers or symbols corresponding to the dimensions or names of dimensions of x.

Examples

julia> using AxisIndices

julia> axs = (Axis(1:5), Axis(1:10));

julia> AxisIndices.drop_axes(axs, 1)
(Axis(1:10 => Base.OneTo(10)),)

julia> AxisIndices.drop_axes(axs, 2)
(Axis(1:5 => Base.OneTo(5)),)

julia> AxisIndices.drop_axes(rand(2, 4), 2)
(Base.OneTo(2),)

AxisIndices.Basics.cat_axis — Function

cat_axis(x, y) -> cat_axis(CombineStyle(x, y), x, y)
cat_axis(::CombineStyle, x, y) -> collection

Returns the concatenated axes x and y.

Examples

julia> using AxisIndices

julia> AxisIndices.cat_axis(Axis(UnitMRange(1, 10)), SimpleAxis(UnitMRange(1, 10)))
Axis(UnitMRange(1:20) => UnitMRange(1:20))

julia> AxisIndices.cat_axis(SimpleAxis(UnitMRange(1, 10)), SimpleAxis(UnitMRange(1, 10)))
SimpleAxis(UnitMRange(1:20))

AxisIndices.Basics.cat_axes — Function

cat_axes(x::AbstractArray, y::AbstractArray, xy::AbstractArray, dims)

Produces the appropriate set of axes where x and y are the arrays that were concatenated over dims to produce xy. The appropriate indices of each axis are derived from from xy.

AxisIndices.PrettyArrays.get_formatters — Function

get_formatters(x)

Returns an argument for the formatters argument of a pretty_table method.

AxisIndices.Basics.CombineStyle — Type

CombineStyle

Abstract type that determines the behavior of broadcast_axis, cat_axis, append_axis!.

AxisIndices.Basics.CombineAxis — Type

CombineAxis

Subtype of CombineStyle that informs relevant methods to produce a subtype of AbstractAxis.

AxisIndices.Basics.CombineSimpleAxis — Type

CombineSimpleAxis

Subtype of CombineStyle that informs relevant methods to produce a subtype of AbstractSimpleAxis.

AxisIndices.Basics.CombineResize — Type

CombineResize

Subtype of CombineStyle that informs relevant methods that axes should be combined by resizing a collection (as opposed to by concatenation or appending).

AxisIndices.Basics.CombineStack — Type

CombineStack

Subtype of CombineStyle that informs relevant methods that axes should be combined by stacking elements in some whay (as opposed to resizing a collection).

AxisIndices.AxisIndicesStyles.AxisIndicesStyle — Type

AxisIndicesStyle

Supertype for traits that control the behavior of AxisIndices.to_index and AxisIndices.to_key.

AxisIndices.AxisIndicesStyles.KeyElement — Type

KeyElement

A subtype of AxisIndicesStyle for mapping an argument that refers to a single key to a single index.

AxisIndices.AxisIndicesStyles.IndexElement — Type

IndexElement

A subtype of AxisIndicesStyle for propagating an argument to a single index.

AxisIndices.AxisIndicesStyles.BoolElement — Type

BoolElement

A subtype of AxisIndicesStyle for mapping an argument that refers to a single Bool to a single index.

AxisIndices.AxisIndicesStyles.CartesianElement — Type

CartesianElement

A subtype of AxisIndicesStyle for mapping an argument that refers to a CartesianIndex to a single index.

AxisIndices.AxisIndicesStyles.KeysCollection — Type

KeysCollection

A subtype of AxisIndicesStyle for mapping a collection of keys a collection of indices.

AxisIndices.AxisIndicesStyles.IndicesCollection — Type

IndicesCollection

A subtype of AxisIndicesStyle for propagating an argument to a collection of indices.

AxisIndices.AxisIndicesStyles.IntervalCollection — Type

IntervalCollection

A subtype of AxisIndicesStyle for mapping an interval argument (start..stop) from keys within said interval to a collection of indices.

AxisIndices.AxisIndicesStyles.BoolsCollection — Type

BoolsCollection

A subtype of AxisIndicesStyle for mapping a collection of Bools to a collection of indices.

AxisIndices.AxisIndicesStyles.KeysIn — Type

KeysIn

A subtype of AxisIndicesStyle for mapping all keys given in(keys) to a collection of indices.

AxisIndices.AxisIndicesStyles.KeyEquals — Type

KeyEquals

A subtype of AxisIndicesStyle for mapping a single key in a isequal(key) argument to a single index.

AxisIndices.AxisIndicesStyles.KeysFix2 — Type

KeysFix2

A subtype of AxisIndicesStyle for mapping all keys from fixed argument (e.g., >(key)) to the corresponding collection of indices.

AxisIndices.AxisIndicesStyles.SliceCollection — Type

SliceCollection

A subtype of AxisIndicesStyle indicating that the entire axis should be propagated.

AxisIndices.AxisIndicesStyles.is_element — Function

is_element(x) -> Bool

Whether x is an AxisIndicesStyle, returns true if it's used to index a single element.

AxisIndices.AxisIndicesStyles.is_index — Function

is_index(x) -> Bool

Whether x is an AxisIndicesStyle, returns true if it's used to search the indices space.

AxisIndices.AxisIndicesStyles.is_collection — Function

is_collection(x) -> Bool

Whether x is an AxisIndicesStyle, returns true if it's used to index a collection of indices.

AxisIndices.AxisIndicesStyles.is_key — Function

is_key(x) -> Bool

Whether x is an AxisIndicesStyle, returns true if it's used to search the keys space.

AxisIndices.AxisIndicesStyles.to_index — Function

to_index(axis, arg) -> to_index(AxisIndicesStyle(axis, arg), axis, arg)

Unique implementation of to_index for the AxisIndices package that specializes based on each axis and indexing argument (as opposed to the array and indexing argument).

AxisIndices.AxisIndicesStyles.to_keys — Function

to_keys([::AxisIndicesStyle,] axis, arg, index)

This method is the reverse of AxisIndices.to_index. arg refers to an argument originally passed to AxisIndices.to_index and index refers to the index produced by that same call to AxisIndices.to_index.

This method assumes to all arguments have passed through AxisIndices.to_index and have been checked to be in bounds. Therefore, this is unsafe and intended only for internal use.

AxisIndices.Math.get_factorization — Function

get_factorization(F::Factorization, A::AbstractArray, d::Symbol)

Used internally to compose an AxisIndicesArray for each component of a factor decomposition. F is the result of decomposition, A is an arry (likely a subtype of AbstractAxisIndices), and d is a symbol referring to a component of the factorization.