Metadata

AxisIndices.Metadata.MetadataArray — Type

MetadataArray(parent::AbstractArray, metadata)

Custom AbstractArray object to store an AbstractArray parent as well as some metadata.

Examples

julia> using AxisIndices

julia> v = ["John", "John", "Jane", "Louise"];

julia> s = MetadataArray(v, Dict("John" => "Treatment", "Louise" => "Placebo", "Jane" => "Placebo"))
4-element MetadataArray{String,1,Dict{String,String},Array{String,1}}:
 "John"
 "John"
 "Jane"
 "Louise"

julia> metadata(s)
Dict{String,String} with 3 entries:
  "John"   => "Treatment"
  "Jane"   => "Placebo"
  "Louise" => "Placebo"

AxisIndices.Metadata.MetaAxis — Type

MetaAxis

An axis type that allows storage of arbitraty metadata.

AxisIndices.Metadata.MetaAxisArray — Type

MetaAxisArray

An AxisArray with metadata.

Examples

julia> using AxisIndices

julia> MetaAxisArray([1 2; 3 4], (["a", "b"], [:one, :two]), metadata = "some metadata")
2×2 MetaAxisArray{Int64,2}
 • dim_1 - ["a", "b"]
 • dim_2 - [:one, :two]
metadata: String
 • some metadata
      one   two
  a     1     2
  b     3     4

AxisIndices.Metadata.MetaCartesianAxes — Type

MetaCartesianAxes

Conveniently construct a CartesianAxes that has metadata.

Examples

julia> using AxisIndices

julia> MetaCartesianAxes(([:a, :b], ["one", "two"]); metadata="some metadata")
2×2 MetaCartesianAxes{CartesianIndex{2},2}
 • dim_1 - [:a, :b]
 • dim_2 - ["one", "two"]
metadata: String
 • some metadata
                       one                    two
  a   CartesianIndex(1, 1)   CartesianIndex(1, 2)
  b   CartesianIndex(2, 1)   CartesianIndex(2, 2)

AxisIndices.Metadata.MetaLinearAxes — Type

MetaLinearAxes

Conveniently construct a LinearAxes that has metadata.

Examples

julia> using AxisIndices

julia> MetaLinearAxes(([:a, :b], ["one", "two"]); metadata="some metadata")
2×2 MetaLinearAxes{Int64,2}
 • dim_1 - [:a, :b]
 • dim_2 - ["one", "two"]
metadata: String
 • some metadata
      one   two
  a     1     3
  b     2     4

AxisIndices.Metadata.NamedMetaAxisArray — Type

NamedMetaAxisArray

An AxisArray with metadata and named dimensions.

AxisIndices.Metadata.NamedMetaCartesianAxes — Type

NamedMetaCartesianAxes

Conveniently construct a CartesianAxes that has metadata and each dimension has a name.

Examples

julia> using AxisIndices

julia> x = NamedMetaCartesianAxes{(:dimx,:dimy)}(([:a, :b], ["one", "two"]); metadata="some metadata")
2×2 NamedMetaCartesianAxes{CartesianIndex{2},2}
 • dimx - [:a, :b]
 • dimy - ["one", "two"]
metadata: String
 • some metadata
                       one                    two
  a   CartesianIndex(1, 1)   CartesianIndex(1, 2)
  b   CartesianIndex(2, 1)   CartesianIndex(2, 2)

julia> x == NamedMetaCartesianAxes((dimx = [:a, :b], dimy = ["one", "two"]); metadata="some metadata")
true

AxisIndices.Metadata.NamedMetaLinearAxes — Type

NamedMetaLinearAxes

Conveniently construct a LinearAxes that has metadata and each dimension has a name.

Examples

julia> using AxisIndices

julia> x = NamedMetaLinearAxes{(:dimx,:dimy)}(([:a, :b], ["one", "two"]); metadata="some metadata")
2×2 NamedMetaLinearAxes{Int64,2}
 • dimx - [:a, :b]
 • dimy - ["one", "two"]
metadata: String
 • some metadata
      one   two
  a     1     3
  b     2     4

julia> x == NamedMetaLinearAxes((dimx = [:a, :b], dimy = ["one", "two"]); metadata="some metadata")
true

AxisIndices.Metadata.has_metadata — Function

has_metadata(x) -> Bool

Returns true if x contains additional fields besides those for keys or indices

AxisIndices.Metadata.has_metaproperty — Function

has_metaproperty(x, meta_key) -> Bool

Returns true if x has a property in its metadata structure paired to meta_key.

AxisIndices.Metadata.axis_meta — Function

axis_meta(x)

Returns metadata (i.e. not keys or indices) associated with each axis of the array x.

axis_meta(x, i)

Returns metadata (i.e. not keys or indices) associated with the ith axis of the array x.

AxisIndices.Metadata.axis_metaproperty — Function

axis_metaproperty(x, i, meta_key)

Return the metadata of x paired to meta_key at axis i.

AxisIndices.Metadata.axis_metaproperty! — Function

axis_metaproperty!(x, meta_key, val)

Set the metadata of x paired to meta_key at axis i.

AxisIndices.Metadata.meta — Function

meta(x)

Shortcut for creating MetaAxis.

Examples

julia> using AxisIndices

julia> A = AxisArray(ones(3), meta("a life well, Steved"));

julia> axis_meta(A)
("a life well, Steved",)

AxisIndices.Metadata.metadata — Function

metadata(x)

Returns metadata for x.

AxisIndices.Metadata.metaproperty — Function

metaproperty(x, meta_key)

Return the metadata of x paired to meta_key.

AxisIndices.Metadata.metaproperty! — Function

metadata!(x, meta_key, val)

Set the metadata of x paired to meta_key.

AxisIndices.Metadata.metadata_type — Function

metadata_type(x)

Returns the type of the metadata of x.