Observations

The AxisIndices.ObservationDims module is experimental and demonstrates the convenience of generating methods using the @defdim macro.

Example usage

Let's create some data where each observation represents a unique set of subject measurements.

julia> using AxisIndices

julia> data = NamedAxisArray(reshape(1:6, 2, 3), x = 2:3, observations = [:subject_1, :subject_2, :subject_3])
2×3 NamedAxisArray{Int64,2}
 • x - 2:3
 • observations - [:subject_1, :subject_2, :subject_3]
      subject_1   subject_2   subject_3
  2           1           3           5
  3           2           4           6

The ObservationDims module isn't necessary to access observations or subject specific data...

julia> data[observations = :subject_1]
2-element NamedAxisArray{Int64,1}
 • x - 2:3

  2   1
  3   2

...but there are a number of convenient methods for accessing observation data with this module.

julia> using AxisIndices.ObservationDims

julia> obsdim(data)  # the dimension along which observations are enumerated
2

julia> nobs(data)  # the number of distinct observations
3

julia> obs_keys(data)  # the key for each observation
3-element Array{Symbol,1}:
 :subject_1
 :subject_2
 :subject_3

There are several other useful methods such as each_obs which supports iterating along the observation dimension. Such methods may be helpful in development of an API. For example, multiple dispatch could be used to return an observation iterator with each_obs on non array types also.

Reference

AxisIndices.ObservationDims.each_obs

AxisIndices.ObservationDims.has_obsdim

AxisIndices.ObservationDims.nobs

AxisIndices.ObservationDims.obs_axis

AxisIndices.ObservationDims.obs_axis

AxisIndices.ObservationDims.obs_axis_type

AxisIndices.ObservationDims.obs_indices

AxisIndices.ObservationDims.obs_keys

AxisIndices.ObservationDims.obsdim

AxisIndices.ObservationDims.select_obs

AxisIndices.ObservationDims.each_obs — Method

each_obs(x)

Create a generator that iterates over the obs dimensions A, returning views that select all the data from the other dimensions in A.

AxisIndices.ObservationDims.has_obsdim — Method

has_obsdim(x) -> Bool

Returns true if x has a dimension corresponding to obs.

AxisIndices.ObservationDims.nobs — Method

nobs(x) -> Int

Returns the size along the dimension corresponding to the obs.

AxisIndices.ObservationDims.obs_axis — Method

obs_axis(x, size[; first_pad=nothing, last_pad=nothing, stride=nothing, dilation=nothing])

Returns an AxisIterator along the obs axis.

AxisIndices.ObservationDims.obs_axis — Method

obs_axis(x)

Returns the axis corresponding to the obs dimension.

AxisIndices.ObservationDims.obs_axis_type — Method

obs_axis_type(x)

Returns the key type corresponding to the obs axis.

AxisIndices.ObservationDims.obs_indices — Method

obs_indices(x)

Returns the indices corresponding to the obs axis

AxisIndices.ObservationDims.obs_keys — Method

obs_keys(x)

Returns the keys corresponding to the obs axis

AxisIndices.ObservationDims.obsdim — Method

obsdim(x) -> Int

Returns the dimension corresponding to obs.

AxisIndices.ObservationDims.select_obs — Method

select_obs(x, i)

Return a view of all the data of x where the index for the obs dimension equals i.