OffsetAxes

References

AxisIndices.OffsetAxes — Module

OffsetAxes

The OffsetAxes module provides the CenteredAxis, [IdentityAxis], and OffsetAxis types, with the respective array types CenteredArray, IdentityArray, OffsetArray. All of these array types are aliases for the AxisArray type with the type of the axes specified for a certain type of offset. Alternatively, the center, idaxis, and offset can be used as shorthand for specifying that a specific axis of an AxisArray has ceretain type of offset indexing behavior.

AxisIndices.OffsetAxes.AbstractOffsetAxis — Type

AbstractOffsetAxis{I,Ks,Inds}

Supertype for axes that begin indexing offset from one. All subtypes of AbstractOffsetAxis use the keys for indexing and only convert to the underlying indices when to_index(::OffsetAxis, ::Integer) is called (i.e. when indexing the an array with an AbstractOffsetAxis. See OffsetAxis, CenteredAxis, and IdentityAxis for more details and examples.

AxisIndices.OffsetAxes.CenteredArray — Type

CenteredArray(A::AbstractArray)

Provides centered axes for indexing A.

Examples

julia> using AxisIndices

julia> CenteredArray(ones(3,3))
3×3 AxisArray{Float64,2}
 • dim_1 - -1:1
 • dim_2 - -1:1
        -1     0     1
  -1   1.0   1.0   1.0
   0   1.0   1.0   1.0
   1   1.0   1.0   1.0

AxisIndices.OffsetAxes.CenteredAxis — Type

CenteredAxis(indices)

A CenteredAxis takes indices and provides a user facing set of keys centered around zero. The CenteredAxis is a subtype of AbstractOffsetAxis and its keys are treated as the predominant indexing style. Note that the element type of a CenteredAxis cannot be unsigned because any instance with a length greater than 1 will begin at a negative value.

Examples

A CenteredAxis sends all indexing arguments to the keys and only maps to the indices when to_index is called.

julia> using AxisIndices

julia> axis = CenteredAxis(1:10)
CenteredAxis(-5:4 => 1:10)

julia> axis[10]  # the indexing goes straight to keys and is centered around zero
ERROR: BoundsError: attempt to access 10-element CenteredAxis(-5:4 => 1:10) at index [10]
[...]

julia> axis[-5]
-5

julia> AxisIndices.to_index(axis, -5)
1

AxisIndices.OffsetAxes.CenteredVector — Type

CenteredVector(v::AbstractVector)

Provides a centered axis for indexing v.

Examples

julia> using AxisIndices

julia> CenteredVector(ones(3))
3-element AxisArray{Float64,1}
 • dim_1 - -1:1

  -1   1.0
   0   1.0
   1   1.0

AxisIndices.OffsetAxes.CenteredVector — Method

CenteredVector{T}(init::ArrayInitializer, sz::Integer)

Creates a vector with elements of type T of size sz and a centered axis.

Examples

julia> using AxisIndices

julia> CenteredVector{Union{Missing, Int}}(missing, 3)
3-element AxisArray{Union{Missing, Int64},1}
 • dim_1 - -1:1

  -1   missing
   0   missing
   1   missing

AxisIndices.OffsetAxes.IdentityArray — Type

IdentityArray(A::AbstractArray)

Provides IdentityAxiss for indexing A.

Examples

julia> using AxisIndices

julia> IdentityArray(ones(3,3))[2:3, 2:3]
2×2 AxisArray{Float64,2}
 • dim_1 - 2:3
 • dim_2 - 2:3
        2     3
  2   1.0   1.0
  3   1.0   1.0

AxisIndices.OffsetAxes.IdentityAxis — Type

IdentityAxis(start, stop) -> axis
IdentityAxis(keys::AbstractUnitRange) -> axis
IdentityAxis(keys::AbstractUnitRange, indices::AbstractUnitRange) -> axis

These are particularly useful for creating views of arrays that preserve the supplied axes:

julia> a = rand(8);

julia> v1 = view(a, 3:5);

julia> axes(v1, 1)
Base.OneTo(3)

julia> idr = IdentityAxis(3:5)
IdentityAxis(3:5 => Base.OneTo(3))

julia> v2 = view(a, idr);

julia> axes(v2, 1)
3:5

AxisIndices.OffsetAxes.IdentityVector — Type

IdentityVector(v::AbstractVector)

Provides an IdentityAxis for indexing v.

Examples

julia> using AxisIndices

julia> IdentityVector(ones(4))[3:4]
2-element AxisArray{Float64,1}
 • dim_1 - 3:4

  3   1.0
  4   1.0

AxisIndices.OffsetAxes.IdentityVector — Method

IdentityVector{T}(init::ArrayInitializer, sz::Integer)

Creates a vector with elements of type T of size sz an IdentityAxis.

Examples

julia> using AxisIndices

julia> IdentityVector{Union{Missing, Int}}(missing, 3)[2:3]
2-element AxisArray{Union{Missing, Int64},1}
 • dim_1 - 2:3

  2   missing
  3   missing

AxisIndices.OffsetAxes.OffsetArray — Type

OffsetArray

An array whose axes are all OffsetAxis

AxisIndices.OffsetAxes.OffsetAxis — Type

OffsetAxis(keys::AbstractUnitRange{<:Integer}, indices::AbstractUnitRange{<:Integer}[, check_length::Bool=true])
OffsetAxis(offset::Integer, indices::AbstractUnitRange{<:Integer})

An axis that has the indexing behavior of an AbstractOffsetAxis and retains an offset from its underlying indices in its keys.

Examples

Users may construct an OffsetAxis by providing an from a set of indices.

julia> using AxisIndices

julia> axis = OffsetAxis(-2, 1:3)
OffsetAxis(-1:1 => 1:3)

In this instance the first index of the wrapped indices is 1 (firstindex(indices(axis))) but adding the offset (-2) moves it to -1.

julia> firstindex(axis)
-1

julia> axis[-1]
-1

Similarly, the last index is move by -2.

julia> lastindex(axis)
1

julia> axis[1]
1

This means that traditional one based indexing no longer applies and may result in errors.

julia> axis[3]
ERROR: BoundsError: attempt to access 3-element OffsetAxis(-1:1 => 1:3) at index [3]
[...]

When an OffsetAxis is reconstructed the offset from indices are presserved.

julia> axis[0:1]  # offset of -2 still applies
OffsetAxis(0:1 => 2:3)

AxisIndices.OffsetAxes.OffsetVector — Type

OffsetVector

A vector whose axis is OffsetAxis

AxisIndices.OffsetAxes.center — Method

center(inds::AbstractUnitRange{<:Integer}) -> CenteredAxis(inds)

Shortcut for creating CenteredAxis.

Examples

julia> using AxisIndices

julia> AxisArray(ones(3), center)
3-element AxisArray{Float64,1}
 • dim_1 - -1:1

  -1   1.0
   0   1.0
   1   1.0

AxisIndices.OffsetAxes.idaxis — Method

idaxis(inds::AbstractUnitRange{<:Integer}) -> IdentityAxis

Shortcut for creating IdentityAxis.

Examples

julia> using AxisIndices

julia> AxisArray(ones(3), idaxis)[2:3]
2-element AxisArray{Float64,1}
 • dim_1 - 2:3

  2   1.0
  3   1.0

AxisIndices.OffsetAxes.offset — Method

offset(x)

Shortcut for creating OffsetAxis where x is the first argument to OffsetAxis.

Examples

julia> using AxisIndices

julia> AxisArray(ones(3), offset(2))
3-element AxisArray{Float64,1}
 • dim_1 - 3:5

  3   1.0
  4   1.0
  5   1.0