Pretty Printing

Warning

Currently pretty printing is an experimental feature that may undergo rapid changes.

It's important that we can view the custom indices that we assign to arrays. Yet a surprising challenge of implementing and using arrays in interactive programming is how complicated printing them can be. Rather than burdening users with cryptic text readouts this package seeks to provide "pretty" printing (quotes because beauty is in the eye of the beholder). This package leans heavily on the PrettyTables.jl package to accomplish this by handing off everything that goes through the show method to pretty_array. pretty_array in turn repeatedly calls PrettyTables.pretty_print along slices of arrays.

AxisIndices.pretty_array — Function

pretty_array([io::IO,] A::AbstractArray[, key_names::Tuple=axes_keys(A)]; kwargs...)

Prints to io the array A with the keys key_names along each dimension of A. Printing of multidimensional arrays is accomplished in a similar manner to Array, where the final two dimensions are sliced producing a series of matrices. kwargs... are passed to pretty_table for 1/2D slice produced.

Examples

julia> using AxisIndices

julia> pretty_array(AxisIndicesArray(ones(2,2,2), (2:3, [:one, :two], ["a", "b"])))
[dim1, dim2, dim3[a]] =
        one     two
  2   1.000   1.000
  3   1.000   1.000

[dim1, dim2, dim3[b]] =
        one     two
  2   1.000   1.000
  3   1.000   1.000