Public API

AxisIndices.AxisCore.AbstractAxisIndices — Type

AbstractAxisIndices

AbstractAxisIndices is a subtype of AbstractArray that offers integration with the AbstractAxis interface. The only methods that absolutely needs to be defined for a subtype of AbstractAxisIndices are axes, parent, similar_type, and similar. Most users should find the provided AxisIndicesArray subtype is sufficient for the majority of use cases. Although custom behavior may be accomplished through a new subtype of AbstractAxisIndices, customizing the behavior of many methods described herein can be accomplished through a unique subtype of AbstractAxis.

This implementation is meant to be basic, well documented, and have sane defaults that can be overridden as necessary. In other words, default methods for manipulating arrays that return an AxisIndicesArray should not cause unexpected downstream behavior for users; and developers should be able to freely customize the behavior of AbstractAxisIndices subtypes with minimal effort.

AxisIndices.AxisCore.AxisIndicesArray — Type

AxisIndicesArray{T,N,P,AI}

An array struct that wraps any parent array and assigns it an AbstractAxis for each dimension. The first argument is the parent array and the second argument is a tuple of subtypes to AbstractAxis or keys that will be converted to subtypes of AbstractAxis with the provided keys.

AxisIndices.AxisCore.AbstractAxis — Type

AbstractAxis

An AbstractVector subtype optimized for indexing.

AxisIndices.AxisCore.AbstractSimpleAxis — Type

AbstractSimpleAxis{V,Vs}

A subtype of AbstractAxis where the keys and values are represented by a single collection.

AxisIndices.AxisCore.Axis — Type

Axis(k[, v=OneTo(length(k))])

Subtypes of AbstractAxis that maps keys to values. The first argument specifies the keys and the second specifies the values. If only one argument is specified then the values span from 1 to the length of k.

Examples

The value for all of these is the same.

julia> using AxisIndices

julia> x = Axis(2.0:11.0, 1:10)
Axis(2.0:1.0:11.0 => 1:10)

julia> y = Axis(2.0:11.0)  # when only one argument is specified assume it's the keys
Axis(2.0:1.0:11.0 => Base.OneTo(10))

julia> z = Axis(1:10)
Axis(1:10 => Base.OneTo(10))

Standard indexing returns the same values

julia> x[2]
2

julia> x[2] == y[2] == z[2]
true

julia> x[1:2]
Axis(2.0:1.0:3.0 => 1:2)

julia> y[1:2]
Axis(2.0:1.0:3.0 => 1:2)

julia> z[1:2]
Axis(1:2 => 1:2)

julia> x[1:2] == y[1:2] == z[1:2]
true

Functions that return true or false may be used to search the keys for their corresponding index. The following is equivalent to the previous example.

julia> x[==(3.0)]
2

julia> x[==(3.0)] ==       # 3.0 is the 2nd key of x
       y[isequal(3.0)] ==  # 3.0 is the 2nd key of y
       z[==(2)]            # 2 is the 2nd key of z
true

julia> x[<(4.0)]  # all keys less than 4.0 are 2.0:3.0 which correspond to values 1:2
Axis(2.0:1.0:3.0 => 1:2)

julia> y[<=(3.0)]  # all keys less than or equal to 3.0 are 2.0:3.0 which correspond to values 1:2
Axis(2.0:1.0:3.0 => 1:2)

julia> z[<(3)]  # all keys less than or equal to 3 are 1:2 which correspond to values 1:2
Axis(1:2 => 1:2)

julia> x[<(4.0)] == y[<=(3.0)] == z[<(3)]
true

Notice that == returns a single value instead of a collection of all elements where the key was found to be true. This is because all keys must be unique so there can only ever be one element returned.

AxisIndices.AxisCore.SimpleAxis — Type

SimpleAxis(v)

Povides an AbstractAxis interface for any AbstractUnitRange, v. v will be considered both the values and keys of the return instance.

Examples

A SimpleAxis is useful for giving a standard set of indices the ability to use the filtering syntax for indexing.

julia> using AxisIndices

julia> x = SimpleAxis(2:10)
SimpleAxis(2:10)

julia> x[2]
2

julia> x[==(2)]
2

julia> x[2] == x[==(2)]  # keys and values are same
true

julia> x[>(2)]
SimpleAxis(3:10)

julia> x[>(2)]
SimpleAxis(3:10)

julia> x[1]
ERROR: BoundsError: attempt to access 9-element SimpleAxis(2:10)) at index [1]
[...]

AxisIndices.Indexing.CartesianAxes — Type

CartesianAxes

Alias for LinearIndices where indices are subtypes of AbstractAxis.

Examples

julia> using AxisIndices

julia> cartaxes = CartesianAxes((Axis(2.0:5.0), Axis(1:4)));

julia> cartinds = CartesianIndices((1:4, 1:4));

julia> cartaxes[2, 2]
CartesianIndex(2, 2)

julia> cartinds[2, 2]
CartesianIndex(2, 2)

AxisIndices.Indexing.LinearAxes — Type

LinearAxes

Alias for LinearIndices where indices are subtypes of AbstractAxis.

Examples

julia> using AxisIndices

julia> linaxes = LinearAxes((Axis(2.0:5.0), Axis(1:4)));

julia> lininds = LinearIndices((1:4, 1:4));

julia> linaxes[2, 2]
6

julia> lininds[2, 2]
6

AxisIndices.Names.NamedIndicesArray — Type

NamedIndicesArray

Type alias for NamedDimsArray whose parent array is a subtype of AxisIndicesArray.

AxisIndices.Names.NIArray — Type

NIArray((parent::AbstractArray; kwargs...) = NIArray(parent, kwargs)
NIArray((parent::AbstractArray, axes::NamedTuple{L,AbstractAxes}))

An abbreviated alias and constructor for NamedIndicesArray. If key word arguments are provided then each key word becomes the name of a dimension and its assigned value is sent to the corresponding axis when constructing the underlying AxisIndicesArray.

Examples

julia> using AxisIndices

julia> A = NIArray(reshape(1:24, 2, 3, 4), x=["a", "b"], y =["one", "two", "three"], z=2:5)
NamedDimsArray{Int64,3,Base.ReshapedArray{Int64,3,UnitRange{Int64},Tuple{}}...}
 • x - Axis(["a", "b"] => Base.OneTo(2))
 • y - Axis(["one", "two", "three"] => Base.OneTo(3))
 • z - Axis(2:5 => Base.OneTo(4))
[x, y, z[2]] =
      one   two   three
  a     1     3       5
  b     2     4       6

[x, y, z[3]] =
      one   two   three
  a     7     9      11
  b     8    10      12

[x, y, z[4]] =
      one   two   three
  a    13    15      17
  b    14    16      18

[x, y, z[5]] =
      one   two   three
  a    19    21      23
  b    20    22      24

julia> dimnames(A)
(:x, :y, :z)

julia> axes_keys(A)
(["a", "b"], ["one", "two", "three"], 2:5)

julia> B = A["a", :, :]
NamedDimsArray{Int64,2,Array{Int64,2}...}
 • y - Axis(["one", "two", "three"] => OneToMRange(3))
 • z - Axis(2:5 => Base.OneTo(4))
          2    3    4    5
    one   1    7   13   19
    two   3    9   15   21
  three   5   11   17   23

julia> C = B["one",:]
NamedDimsArray{Int64,1,Array{Int64,1}...}
 • z - Axis(2:5 => Base.OneTo(4))

  2    1
  3    7
  4   13
  5   19

AxisIndices.AxisCore.indices — Function

indices(x::AbstractAxis)

Returns the indices x.

Examples

julia> using AxisIndices

julia> indices(Axis(["a"], 1:1))
1:1

julia> indices(CartesianIndex(1,1))
(1, 1)

indices(x, i)

Returns the indices corresponding to the i axis

Examples

julia> using AxisIndices

julia> indices(AxisIndicesArray(ones(2,2), (2:3, 3:4)), 1)
Base.OneTo(2)

indices(x)

Returns the indices corresponding to all axes of x.

Examples

julia> using AxisIndices

julia> indices(AxisIndicesArray(ones(2,2), (2:3, 3:4)))
(Base.OneTo(2), Base.OneTo(2))

julia> indices(Axis(["a"], 1:1))
1:1

julia> indices(CartesianIndex(1,1))
(1, 1)

AxisIndices.AxisCore.axes_keys — Function

axes_keys(x::AbstractArray)

Returns the keys corresponding to all axes of x.

Examples

julia> using AxisIndices

julia> axes_keys(AxisIndicesArray(ones(2,2), (2:3, 3:4)))
(2:3, 3:4)

julia> axes_keys(Axis(1:2))
(1:2,)

axes_keys(x, i)

Returns the axis keys corresponding of ith dimension of x.

Examples

julia> using AxisIndices

julia> axes_keys(AxisIndicesArray(ones(2,2), (2:3, 3:4)), 1)
2:3

AxisIndices.PrettyArrays.pretty_array — Function

pretty_array([io::IO,] A::AbstractArray[, axs=axes(A)], dimnames=ntuple(i -> Symbol(:dim_, i), N), backend=:text; kwargs...)

Prints to io the array A with the keys key_names along each dimension of A. Printing of multidimensional arrays is accomplished in a similar manner to Array, where the final two dimensions are sliced producing a series of matrices. kwargs... are passed to pretty_table for 1/2D slice produced.

Examples

julia> using AxisIndices

julia> pretty_array(ones(Int, 2,2,2), (Axis(2:3), Axis([:one, :two]), Axis(["a", "b"])), (:x, :y, :z))
[x, y, z[a]] =
      one   two
  2     1     1
  3     1     1

[x, y, z[b]] =
      one   two
  2     1     1
  3     1     1

AxisIndices.AxisCore.keys_type — Function

keys_type(x)

Retrieves the type of the keys of x.

Examples

julia> using AxisIndices

julia> keys_type(Axis(1:2))
UnitRange{Int64}

julia> keys_type(typeof(Axis(1:2)))
UnitRange{Int64}

julia> keys_type(UnitRange{Int})
Base.OneTo{Int64}

keys_type(x, i)

Retrieves axis keys of the ith dimension of x.

Examples

julia> using AxisIndices

julia> keys_type(AxisIndicesArray([1], ["a"]), 1)
Array{String,1}

AxisIndices.AxisCore.first_key — Function

first_key(x)

Returns the first key of x.

Examples

julia> using AxisIndices

julia> first_key(Axis(2:10))
2

AxisIndices.AxisCore.last_key — Function

last_key(x)

Returns the last key of x.

Examples

julia> using AxisIndices

julia> last_key(Axis(2:10))
10