Resizing Axes

These methods help with operations that need to resize axes, either dynamically or by creating a new instance of an axis. In addition to helping with operations related to array resizing, these may be useful for managing the axis of a vector throughout a push!, pushfirst!, pop, and popfirst! operation.

AxisIndices.resize_first — Function

resize_first(x, n::Integer)

Returns a collection similar to x that grows or shrinks from the first index to be of size n.

Examples

julia> using AxisIndices

julia> x = collect(1:5);

julia> resize_first(x, 2)
2-element Array{Int64,1}:
 4
 5

julia> resize_first(x, 7)
7-element Array{Int64,1}:
 -1
  0
  1
  2
  3
  4
  5

julia> resize_first(x, 5)
5-element Array{Int64,1}:
 1
 2
 3
 4
 5

AxisIndices.resize_first! — Function

resize_first!(x, n::Integer)

Returns the collection x after growing or shrinking the first index to be of size n.

Examples

julia> using AxisIndices

julia> x = collect(1:5);

julia> resize_first!(x, 2);

julia> x
2-element Array{Int64,1}:
 4
 5

julia> resize_first!(x, 6);

julia> x
6-element Array{Int64,1}:
 0
 1
 2
 3
 4
 5

AxisIndices.resize_last — Function

resize_last(x, n::Integer)

Returns a collection similar to x that grows or shrinks from the last index to be of size n.

Examples

julia> using AxisIndices

julia> x = collect(1:5);

julia> resize_last(x, 2)
2-element Array{Int64,1}:
 1
 2

julia> resize_last(x, 7)
7-element Array{Int64,1}:
 1
 2
 3
 4
 5
 6
 7

julia>  resize_last(x, 5)
5-element Array{Int64,1}:
 1
 2
 3
 4
 5

AxisIndices.resize_last! — Function

resize_last!(x, n::Integer)

Returns the collection x after growing or shrinking the last index to be of size n.

Examples

julia> using AxisIndices

julia> x = collect(1:5);

julia> resize_last!(x, 2);

julia> x
2-element Array{Int64,1}:
 1
 2

julia> resize_last!(x, 5);

julia> x
5-element Array{Int64,1}:
 1
 2
 3
 4
 5

AxisIndices.grow_first — Function

grow_first(x, n)

Returns a collection similar to x that grows by n elements from the first index.

Examples

julia> using AxisIndices

julia> mr = UnitMRange(1, 10)
UnitMRange(1:10)

julia> grow_first(mr, 2)
UnitMRange(-1:10)

AxisIndices.grow_first! — Function

grow_first!(x, n)

Returns the collection x after growing from the first index by n elements.

Examples

julia> using AxisIndices

julia> mr = UnitMRange(1, 10)
UnitMRange(1:10)

julia> grow_first!(mr, 2);

julia> mr
UnitMRange(-1:10)

AxisIndices.grow_last — Function

grow_last(x, n)

Returns a collection similar to x that grows by n elements from the last index.

Examples

julia> using AxisIndices

julia> mr = UnitMRange(1, 10)
UnitMRange(1:10)

julia> grow_last(mr, 2)
UnitMRange(1:12)

AxisIndices.grow_last! — Function

grow_last!(x, n)

Returns the collection x after growing from the last index by n elements.

Examples

julia> using AxisIndices

julia> mr = UnitMRange(1, 10)
UnitMRange(1:10)

julia> grow_last!(mr, 2);

julia> mr
UnitMRange(1:12)

AxisIndices.shrink_first — Function

shrink_first(x, n)

Returns a collection similar to x that shrinks by n elements from the first index.

Examples

julia> using AxisIndices

julia> mr = UnitMRange(1, 10)
UnitMRange(1:10)

julia> shrink_first(mr, 2)
UnitMRange(3:10)

AxisIndices.shrink_first! — Function

shrink_first!(x, n)

Returns the collection x after shrinking from the first index by n elements.

AxisIndices.shrink_last — Function

shrink_last(x, n)

Returns a collection similar to x that shrinks by n elements from the last index.

Examples

julia> using AxisIndices

julia> mr = UnitMRange(1, 10)
UnitMRange(1:10)

julia> shrink_last(mr, 2)
UnitMRange(1:8)

AxisIndices.shrink_last! — Function

shrink_last!(x, n)

Returns the collection x after shrinking from the last index by n elements.

Examples

julia> using AxisIndices

julia> mr = UnitMRange(1, 10)
UnitMRange(1:10)

julia> shrink_last!(mr, 2);

julia> mr
UnitMRange(1:8)

AxisIndices.next_type — Function

next_type(x::T)

Returns the immediately greater value of type T.

Examples

julia> using AxisIndices

julia> AxisIndices.next_type("b")
"c"

julia> AxisIndices.next_type(:b)
:c

julia> AxisIndices.next_type('a')
'b': ASCII/Unicode U+0062 (category Ll: Letter, lowercase)

julia> AxisIndices.next_type(1)
2

julia> AxisIndices.next_type(2.0)
2.0000000000000004

julia> AxisIndices.next_type("")
""

AxisIndices.prev_type — Function

prev_type(x::T)

Returns the immediately lesser value of type T.

Examples

julia> using AxisIndices

julia> AxisIndices.prev_type("b")
"a"

julia> AxisIndices.prev_type(:b)
:a

julia> AxisIndices.prev_type('b')
'a': ASCII/Unicode U+0061 (category Ll: Letter, lowercase)

julia> AxisIndices.prev_type(1)
0

julia> AxisIndices.prev_type(1.0)
0.9999999999999999

julia> AxisIndices.prev_type("")
""