Axis Traits

Indexing Traits

At its core, AxisIndices relies on a small change in processing indexing, permitting its unique abilities. Most arrays pass indexing arguments to the to_indices method at some point. Where a user provides performs indexing as A[arg1, arg2], the translation to native indexing space looks a bit like this:

# `axes_of_A = axes(A)` and `indexing_arguments = (arg1, arg2)`
function to_indices(A, axes_of_A::Tuple, indexing_arguments::Tuple)
    return (to_index(A, first(indexing_arguments)), to_indices(A, tail(axes_of_A), tail(indexing_arguments))...)
end

This means that cusotmizing indexing behavior can be accomplished by either

Circumeventing to_index by overloading to_indices

Overloading to_index based on A

Overloading to_index based on each indexing argument (e.g., arg1, arg2)

This package changes this to:

# `axes_of_A = axes(A)` and `indexing_arguments = (arg1, arg2)`
function to_indices(A, axes_of_A::Tuple, indexing_arguments::Tuple)
    return (to_index(first(axes_of_A), first(indexing_arguments)), to_indices(A, tail(axes_of_A), tail(indexing_arguments))...)
end

Now each translation to native indexing space occurs through the direction of each axis of A. This means that customizing indexing behavior can now be accomplished by either

Overloading to_indices based on A.

Overloading to_index based on each axis of A (e.g., axes(A, 1), axes(A, 2))

Overloading to_index based on each indexing argument (e.g., arg1, arg2)

However, this does not mean that overloading to_index is always safe or even advisable. This portion of the indexing pipeline is responsible for mapping arguments to the underlying memory of arrays, so you can cause some significant problems if you don't know what you're doing. It's also easy to inadvertently effect performance when changing how indexing works. Therefore, AxisIndices also small trait system for handling this. This is accomplished through the use of ToIndexStyle, which is injected into the to_index schema as follows:

function to_index(axis, inds)
    return to_index(ToIndexStyle(eltype(inds)), axis, inds)
end

Several subtypes of ToIndexStyle are provided, allowing users to customize indexing behavior without worrying about rewriting performance critical code.

AxisIndices.ToIndexStyle — Type

ToIndexStyle

ToIndexStyle specifies how to_index(axis, inds) should convert a provided argument indices into the native indexing of structure. ToIndexStyle(eltype(inds)) determines whether SearchKeys, SearchIndices, or GetIndices is returned.

AxisIndices.SearchKeys — Type

ToKeysCollection()

Calling to_index(::SearchKeys, axis, inds) -> newinds results in:

Identifying the positions of keys(axis) that equal inds

Retreiving the values(axis) that correspond to the identified positions.

Attempt to reconstruct the axis type (e.g., Axis or SimpleAxis) with the relevant keys and values. This is only possible if newinds isa AbstractUnitRange{Integer} is true

AxisIndices.SearchIndices — Type

SearchIndices()

Calling to_index(::SearchIndices, axis, inds) -> newinds results in:

Identifying the positions of values(axis) that equal inds

Attempt to reconstruct the axis type (e.g., Axis or SimpleAxis) with the relevant keys and values. This is only possible if newinds isa AbstractUnitRange{Integer} is true

AxisIndices.GetIndices — Type

GetIndices()

Calling to_index(::GetIndices, axis, inds) -> newinds results in:

Performs getindex(values(axis), inds)

If the initual output is a subtype of AbstractUnitRange{Integer} then an axis type is returned. Otherwise just the initial output is returned.

Combine Traits

One natural consequence of having axes as independent entities from the actual array data is that it complicates operations such as cat and append!.

How should we propagate characteristics from the values of each axis so that the resulting array is still appropriately memory mapped.

How should we combine keys of different types and that are not necessarily unique?

AxisIndices.CombineStyle — Type

CombineStyle

Abstract type that determines the behavior of broadcast_axis, cat_axis, append_axis!.

AxisIndices.CombineAxis — Type

CombineAxis

Subtype of CombineStyle that informs relevant methods to produce a subtype of AbstractAxis.

AxisIndices.CombineSimpleAxis — Type

CombineSimpleAxis

Subtype of CombineStyle that informs relevant methods to produce a subtype of AbstractSimpleAxis.

AxisIndices.CombineResize — Type

CombineResize

Subtype of CombineStyle that informs relevant methods that axes should be combined by resizing a collection (as opposed to by concatenation or appending).

AxisIndices.CombineStack — Type

CombineStack

Subtype of CombineStyle that informs relevant methods that axes should be combined by stacking elements in some whay (as opposed to resizing a collection).