The Axis Interface

The following describes the components necessary to construct and manipulate existing and new subtypes of AbstractAxis.

Introduction

The supertype to all axis types herein is the AbstractAxis, which is a subtype of AbstractUnitRange{<:Integer}.

If we have a set of keys a b c and a set of indices 1 2 3 then the key a maps to the index 1. Given these definitions, the AbstractAxis differs from the classic dictionary in the following two ways:

The valtype of AbstractAxis is always an integer.

The values are always unique and continuous.

The two main axis types defined here are Axis and SimpleAxis. The standard syntax for indexing doesn't change at all for these types.

julia> using AxisIndices

julia> using Dates

julia> using ChainedFixes  # provides `and`, `or`, `⩓`, `⩔` methods

julia> sa = SimpleAxis(1:10)
SimpleAxis(1:10)

julia> sa[2]
2

julia> sa[>(2)]
SimpleAxis(3:10)

julia> a = Axis(1:10)
Axis(1:10 => Base.OneTo(10))

julia> a[2]
2

julia> a[2:3]
Axis(2:3 => 2:3)

But now we can also use functions to index by the keys of an AbstractAxis.

julia> a = Axis(2.0:11.0, 1:10)
Axis(2.0:1.0:11.0 => 1:10)

julia> a[1]
1

julia> a[isequal(2.0)]
1

julia> a[>(2)]
Axis(3.0:1.0:11.0 => 2:10)

julia> a[>(2.0)]
Axis(3.0:1.0:11.0 => 2:10)

julia> a[and(>(2.0), <(8.0))]
Axis(3.0:1.0:7.0 => 2:6)

julia> sa[in(3:5)]
SimpleAxis(3:5)

This also allows certain syntax special treatment because they are obviously not referring to traditional integer based indexing.

julia> x, y, z = Axis([:one, :two, :three]), Axis(["one", "two", "three"]), Axis(Second(1):Second(1):Second(3));

julia> x[:one]
1

julia> x[:one] == y["one"] == z[Second(1)]
true

julia> x[[:one, :two]]
2-element Array{Int64,1}:
 1
 2

Note in the last example that a vector was returned instead of an AbstractAxis. An AbstractAxis is a subtype of AbstractUnitRange and therefore cannot be reformed after any operation that does not guarantee the return of another unit range. This is similar to the behavior of UnitRange in base.

Indexing an Axis

Setup for running axis examples.

julia> using AxisIndices, Unitful, IntervalSets, ChainedFixes

julia> using Unitful: s

julia> time1 = Axis((1.5:.5:10)s)
Axis((1.5:0.5:10.0) s => Base.OneTo(18))

julia> time2 = Axis((1.5:.5:10)s, 2:19)
Axis((1.5:0.5:10.0) s => 2:19)

Indexing With Integers

Integers will map directly to the indices of an axis.

julia> time1[1]
1

julia> time1[2]
2

julia> time2[2]
2

julia> time2[1]
ERROR: BoundsError: attempt to access 18-element Axis((1.5:0.5:10.0) s => 2:19) at index [1]
[...]

Notice that time2[1] throws an error. This is because the indices of the time2 axis don't contain a 1 and begins at 2. This allows an axis to map to any single dimensional memory mapping, even if it doesn't start at 1.

Indexing an axis with a collection of integers works similarly to indexing any other AbstractUnitRange. That is, using other subtypes of AbstractUnitRange preserves the structure...

julia> time1[1:2]
Axis((1.5:0.5:2.0) s => 1:2)

julia> time2[2:3]
Axis((1.5:0.5:2.0) s => 2:3)

However, we can't ensure that the resulting range will have a step of one in other cases so only the indices are returned.

julia> time1[1:2:4]
1:2:3

julia> time1[[1, 2, 3]]
3-element Array{Int64,1}:
 1
 2
 3

julia> time1[firstindex(time1):end]
Axis((1.5:0.5:10.0) s => 1:18)

Indexing With Keys

julia> time1[1.5s]
1

julia> time2[1.5s]
2

julia> time1[1.5s..3s]
Axis((1.5:0.5:3.0) s => 1:4)

julia> time1[3s..4.5s]
Axis((3.0:0.5:4.5) s => 4:7)

Keys and Indices

If our keys are integers and we want to ensure that we always refer keys we can use Keys

julia> Axis((2:11), 1:10)[Keys(<(5))]
Axis(2:4 => 1:3)

julia> Axis((2:11), 1:10)[Indices(<(5))]
Axis(2:5 => 1:4)

julia> Axis((2:11), 1:10)[Keys(3)]
2

julia> Axis((2:11), 1:10)[Indices(3)]
3

Approximate Indexing

julia> axis = Axis([pi + 0, pi + 1]);

julia> axis[3.141592653589793]
1

julia> axis[3.14159265358979]
ERROR: BoundsError: attempt to access 2-element Axis([3.141592653589793, 4.141592653589793] => OneToMRange(2)) at index [3.14159265358979]
[...]

julia> axis[isapprox(3.14159265358979)]
1

julia> axis[isapprox(3.14, atol=1e-2)]
1

Indexing With Functions

Operators that typically return true or false can often

julia> time1[<(3.0s)]
Axis((1.5:0.5:2.5) s => 1:3)

julia> time1[>(3.0s)]
Axis((3.5:0.5:10.0) s => 5:18)

julia> time1[==(6.0s)]
10

julia> time1[!=(6.0s)] == vcat(1:9, 11:18)
true

These operators can also be combined to get more specific regions of an axis.

julia> time1[and(>(2.5s), <(10.0s))]
Axis((3.0:0.5:9.5) s => 4:17)

julia> time1[>(2.5s) ⩓ <(10.0s)]  # equivalent to `and` you can use \And<TAB>
Axis((3.0:0.5:9.5) s => 4:17)

julia> time1[or(<(2.5s),  >(9.0s))] == vcat(1:2, 17:18)
true

julia> time1[<(2.5s) ⩔ >(9.0s)] == vcat(1:2, 17:18) # equivalent to `or` you can use \Or<TAB>
true

Quick Look at AbstractAxis Types

AbstractAxis Type	Usage

Axis	Attach keys to indices

SimpleAxis	Give standard indices access AxisIndices's syntax

CenteredAxis	Enforce indexing that is centered around zero

OffsetAxis	Enforce indexing with where the first index is offset from 1

MetaAxis	Attach arbitrary metadata to an axis

StructAxis	Map a type's field names and field types to each element along an axis.

Performance

Indexing CartesianAxes is comparable to that of CartesianIndices.

julia> using AxisIndices, BenchmarkTools

julia> cartaxes = CartesianAxes((Axis(2.0:5.0), Axis(1:4)));

julia> cartinds = CartesianIndices((1:4, 1:4));

julia> @btime getindex(cartaxes, 2, 2)
20.848 ns (1 allocation: 32 bytes)
CartesianIndex(2, 2)

julia> @btime getindex(cartinds, 2, 2)
22.317 ns (1 allocation: 32 bytes)
CartesianIndex(2, 2)

julia> @btime getindex(cartaxes, ==(3.0), 2)
444.374 ns (7 allocations: 416 bytes)
CartesianIndex(2, 2)

Indexing LinearAxes is comparable to that of LinearIndices

julia> using AxisIndices, BenchmarkTools

julia> linaxes = LinearAxes((Axis(1.0:4.0), Axis(1:4)));

julia> lininds = LinearIndices((1:4, 1:4));

julia> @btime getindex(linaxes, 2, 2)
18.275 ns (0 allocations: 0 bytes)
6

julia> @btime getindex(lininds, 2, 2)
18.849 ns (0 allocations: 0 bytes)
6

julia> @btime getindex(linaxes, ==(3.0), 2)
381.098 ns (6 allocations: 384 bytes)
7

You may notice there's significant overhead for using the filtering syntax. However, the filtering syntax takes advantage of a special type in base, Fix2. This means that we can take advantage of filtering methods that have been optimized for specific types of keys. Here we do the same thing as above but we create a function that knows it's going to perform filtering.

julia> getindex_filter(a, i1, i2) = a[==(i1), ==(i2)]
getindex_filter (generic function with 1 method)

julia> @btime getindex_filter(linaxes, 3.0, 2)
57.216 ns (0 allocations: 0 bytes)
7

julia> linaxes2 = LinearAxes((Axis(Base.OneTo(4)), Axis(Base.OneTo(4))));

julia> @btime getindex_filter(linaxes2, 3, 2)
22.070 ns (0 allocations: 0 bytes)
7

Indexing linaxes is much faster now that it can be optimized inside of a function call. However, it's still a little over twice as slow as normal indexing. That's largely because of the cost of searching 1.0:4.0 (which is a StepRangeLen type in this case). The second benchmark demonstrates how close we really are to standard indexing given similar range types.

Reference

AxisIndices.Interface.axes_keys

AxisIndices.Interface.axes_keys

AxisIndices.Interface.axis_eltype

AxisIndices.Interface.axis_meta

AxisIndices.Interface.axis_meta

AxisIndices.Interface.colaxis

AxisIndices.Interface.colkeys

AxisIndices.Interface.coltype

AxisIndices.Interface.drop_axes

AxisIndices.Interface.first_key

AxisIndices.Interface.has_dimnames

AxisIndices.Interface.has_metadata

AxisIndices.Interface.indices

AxisIndices.Interface.indices

AxisIndices.Interface.indices

AxisIndices.Interface.indices_type

AxisIndices.Interface.indices_type

AxisIndices.Interface.is_indices_axis

AxisIndices.Interface.keys_type

AxisIndices.Interface.keys_type

AxisIndices.Interface.last_key

AxisIndices.Interface.metadata

AxisIndices.Interface.metadata_type

AxisIndices.Interface.named_axes

AxisIndices.Interface.rowaxis

AxisIndices.Interface.rowkeys

AxisIndices.Interface.rowtype

AxisIndices.Interface.step_key

AxisIndices.Interface.to_index

AxisIndices.Interface.to_keys

AxisIndices.Interface.unsafe_reconstruct

AxisIndices.Interface.unsafe_reconstruct

AxisIndices.Interface.unsafe_reconstruct

AxisIndices.Interface.AxesIterator

AxisIndices.Interface.AxisIterator

AxisIndices.Axes.permute_axes

AxisIndices.Axes.permute_axes

AxisIndices.Axes.permute_axes

AxisIndices.Axes.structview

AxisIndices.Axes.AbstractAxis

AxisIndices.Axes.AbstractOffsetAxis

AxisIndices.Axes.AbstractSimpleAxis

AxisIndices.Axes.Axis

AxisIndices.Axes.CartesianAxes

AxisIndices.Axes.CenteredAxis

AxisIndices.Axes.LinearAxes

AxisIndices.Axes.MetaAxis

AxisIndices.Axes.MetaCartesianAxes

AxisIndices.Axes.MetaLinearAxes

AxisIndices.Axes.NamedCartesianAxes

AxisIndices.Axes.NamedLinearAxes

AxisIndices.Axes.NamedMetaCartesianAxes

AxisIndices.Axes.NamedMetaLinearAxes

AxisIndices.Axes.OffsetAxis

AxisIndices.Axes.SimpleAxis

AxisIndices.Axes.SimpleAxis

AxisIndices.Axes.SimpleAxis

AxisIndices.Axes.StructAxis

AxisIndices.Interface.AxesIterator — Type

AxesIterator

N-dimensional iterator of AxisIterators.

Examples

julia> using AxisIndices

julia> AxesIterator(CartesianAxes((20, 20, 20)), (3,3,3))
AxesIterator:
 • AxisIterator((1:3):3:(16:18))
 • AxisIterator((1:3):3:(16:18))
 • AxisIterator((1:3):3:(16:18))

AxisIndices.Interface.AxisIterator — Type

AxisIterator(axis, window_size[, first_pad=nothing, last_pad=nothing, stride=nothing, dilation=nothing])

Creates an iterator for indexing ranges of elements within axis.

Examples

julia> using AxisIndices

julia> axis = Axis(range(2.0, step=3.0, length=20))
Axis(2.0:3.0:59.0 => Base.OneTo(20))

julia> AxisIterator(axis, 3)
AxisIterator((1:3):3:(16:18))

The final print out indicates that the first window is 1:3 and all subsequent iterations move by 3 until reaching 16:18.

The size of the window may be determined by providing an explicit size or the size in terms of the keys of an axis.

julia> collect(AxisIterator(axis, 3))
6-element Array{Any,1}:
 1:3
 4:6
 7:9
 10:12
 13:15
 16:18

julia> collect(AxisIterator(axis, 9.0))
6-element Array{Any,1}:
 1:3
 4:6
 7:9
 10:12
 13:15
 16:18

The iterator may start with padding from the beginning..

julia> collect(AxisIterator(axis, 3, first_pad=1))
6-element Array{Any,1}:
 2:4
 5:7
 8:10
 11:13
 14:16
 17:19

julia> collect(AxisIterator(axis, 9.0, first_pad=3.0))
6-element Array{Any,1}:
 2:4
 5:7
 8:10
 11:13
 14:16
 17:19

...and the end.

julia> collect(AxisIterator(axis, 3, first_pad=1, last_pad=2))
5-element Array{Any,1}:
 2:4
 5:7
 8:10
 11:13
 14:16

julia> collect(AxisIterator(axis, 9.0, first_pad=3.0, last_pad=6.0))
5-element Array{Any,1}:
 2:4
 5:7
 8:10
 11:13
 14:16

The window can be dilated so that a regular but non-continuous range of elements are indexed.

julia> collect(AxisIterator(axis, 3, first_pad=1, last_pad=2, dilation=2))
5-element Array{Any,1}:
 2:2:4
 5:2:7
 8:2:10
 11:2:13
 14:2:16

julia> collect(AxisIterator(axis, 9.0, first_pad=3.0, last_pad=6.0, dilation=6.0))
5-element Array{Any,1}:
 2:2:4
 5:2:7
 8:2:10
 11:2:13
 14:2:16

Regular strides can be placed between each iteration.

julia> collect(AxisIterator(axis, 3, first_pad=1, last_pad=2, stride=2))
3-element Array{Any,1}:
 2:4
 7:9
 12:14

julia> collect(AxisIterator(axis, 9.0, first_pad=3.0, last_pad=6.0, stride=6.0))
3-element Array{Any,1}:
 2:4
 7:9
 12:14

AxisIndices.Interface.axes_keys — Method

axes_keys(x, i)

Returns the axis keys corresponding of ith dimension of x.

Examples

julia> using AxisIndices

julia> axes_keys(AxisArray(ones(2,2), (2:3, 3:4)), 1)
2:3

AxisIndices.Interface.axes_keys — Method

axes_keys(x) -> Tuple

Returns the keys corresponding to all axes of x.

Examples

julia> using AxisIndices

julia> axes_keys(AxisArray(ones(2,2), (2:3, 3:4)))
(2:3, 3:4)

julia> axes_keys(Axis(1:2))
(1:2,)

AxisIndices.Interface.axis_eltype — Method

axis_eltype(x)

Returns the type corresponds to the type of the ith element returned when slicing along that dimension.

AxisIndices.Interface.axis_meta — Method

axis_meta(x, i)

Returns metadata (i.e. not keys or indices) associated with the ith axis of the array x.

AxisIndices.Interface.axis_meta — Method

axis_meta(x)

Returns metadata (i.e. not keys or indices) associated with each axis of the array x.

AxisIndices.Interface.colaxis — Method

colaxis(x) -> axis

Returns the axis corresponding to the second dimension of x.

Examples

julia> using AxisIndices

julia> colaxis(AxisArray(ones(2,2), ["a", "b"], [:one, :two]))
Axis([:one, :two] => Base.OneTo(2))

AxisIndices.Interface.colkeys — Method

colkeys(x) -> axis

Returns the keys corresponding to the second dimension of x.

Examples

julia> using AxisIndices

julia> colkeys(AxisArray(ones(2,2), ["a", "b"], [:one, :two]))
2-element Array{Symbol,1}:
 :one
 :two

AxisIndices.Interface.coltype — Method

coltype(x)

Returns the type of the axis corresponding to the second dimension of x.

Examples

julia> using AxisIndices

julia> coltype(AxisArray(ones(2,2), ["a", "b"], [:one, :two]))
Axis{Symbol,Int64,Array{Symbol,1},Base.OneTo{Int64}}

AxisIndices.Interface.drop_axes — Method

drop_axes(x, dims)

Returns all axes of x except for those identified by dims. Elements of dims must be unique integers or symbols corresponding to the dimensions or names of dimensions of x.

Examples

julia> using AxisIndices

julia> axs = (Axis(1:5), Axis(1:10));

julia> AxisIndices.drop_axes(axs, 1)
(Axis(1:10 => Base.OneTo(10)),)

julia> AxisIndices.drop_axes(axs, 2)
(Axis(1:5 => Base.OneTo(5)),)

julia> AxisIndices.drop_axes(rand(2, 4), 2)
(Base.OneTo(2),)

AxisIndices.Interface.first_key — Method

first_key(x)

Returns the first key of x.

Examples

julia> using AxisIndices

julia> first_key(Axis(2:10))
2

AxisIndices.Interface.has_dimnames — Method

has_dimnames(x) -> Bool

Returns true if x has names for each dimension.

AxisIndices.Interface.has_metadata — Method

has_metadata(x) -> Bool

Returns true if x contains additional fields besides those for keys or indices

AxisIndices.Interface.indices — Method

indices(x::AbstractUnitRange)

Returns the indices x.

Examples

julia> using AxisIndices

julia> indices(Axis(["a"], 1:1))
1:1

julia> indices(CartesianIndex(1,1))
(1, 1)

AxisIndices.Interface.indices — Method

indices(x, i)

Returns the indices corresponding to the i axis

Examples

julia> using AxisIndices

julia> indices(AxisArray(ones(2,2), (2:3, 3:4)), 1)
Base.OneTo(2)

AxisIndices.Interface.indices — Method

indices(x) -> Tuple

Returns the indices corresponding to all axes of x.

Examples

julia> using AxisIndices

julia> indices(AxisArray(ones(2,2), (2:3, 3:4)))
(Base.OneTo(2), Base.OneTo(2))

julia> indices(Axis(["a"], 1:1))
1:1

julia> indices(CartesianIndex(1,1))
(1, 1)

AxisIndices.Interface.indices_type — Method

indices_type(x, i)

Retrieves axis values of the ith dimension of x.

Examples

julia> using AxisIndices

julia> indices_type([1], 1)
Base.OneTo{Int64}

julia> indices_type(typeof([1]), 1)
Base.OneTo{Int64}

AxisIndices.Interface.indices_type — Method

indices_type(x)

Retrieves the type of the values of x. This should be functionally equivalent to typeof(values(x)).

Examples

julia> using AxisIndices

julia>  indices_type(Axis(1:2))
Base.OneTo{Int64}

julia> indices_type(typeof(Axis(1:2)))
Base.OneTo{Int64}

julia> indices_type(typeof(1:2))
UnitRange{Int64}

AxisIndices.Interface.is_indices_axis — Method

is_indices_axis(x) -> Bool

If true then x is an axis type where the only field parameterizing the axis is a field for the values.

AxisIndices.Interface.keys_type — Method

keys_type(x, i)

Retrieves axis keys of the ith dimension of x.

Examples

julia> using AxisIndices

julia> keys_type(AxisArray([1], ["a"]), 1)
Array{String,1}

AxisIndices.Interface.keys_type — Method

keys_type(x)

Retrieves the type of the keys of x.

Examples

julia> using AxisIndices

julia> keys_type(Axis(1:2))
UnitRange{Int64}

julia> keys_type(typeof(Axis(1:2)))
UnitRange{Int64}

julia> keys_type(UnitRange{Int})
Base.OneTo{Int64}

AxisIndices.Interface.last_key — Method

last_key(x)

Returns the last key of x.

Examples

julia> using AxisIndices

julia> last_key(Axis(2:10))
10

AxisIndices.Interface.metadata — Method

metadata(x)

Returns metadata for x.

AxisIndices.Interface.metadata_type — Method

metadata_type(x)

Returns the type of the metadata of x.

AxisIndices.Interface.named_axes — Method

named_axes(A) -> NamedTuple{names}(axes)

Returns a NamedTuple where the names are the dimension names and each indice is the corresponding dimensions's axis. If dimnesion names are not defined for x default names are returned. x should have an axes method.

julia> using AxisIndices

julia> A = reshape(1:24, 2,3,4);

julia> named_axes(A)
(dim_1 = Base.OneTo(2), dim_2 = Base.OneTo(3), dim_3 = Base.OneTo(4))

julia> named_axes(NamedAxisArray{(:a, :b, :c)}(A))
(a = SimpleAxis(Base.OneTo(2)), b = SimpleAxis(Base.OneTo(3)), c = SimpleAxis(Base.OneTo(4)))

AxisIndices.Interface.rowaxis — Method

rowaxis(x) -> axis

Returns the axis corresponding to the first dimension of x.

Examples

julia> using AxisIndices

julia> rowaxis(AxisArray(ones(2,2), ["a", "b"], [:one, :two]))
Axis(["a", "b"] => Base.OneTo(2))

AxisIndices.Interface.rowkeys — Method

rowkeys(x) -> axis

Returns the keys corresponding to the first dimension of x.

Examples

julia> using AxisIndices

julia> rowkeys(AxisArray(ones(2,2), ["a", "b"], [:one, :two]))
2-element Array{String,1}:
 "a"
 "b"

AxisIndices.Interface.rowtype — Method

rowtype(x)

Returns the type of the axis corresponding to the first dimension of x.

Examples

julia> using AxisIndices

julia> rowtype(AxisArray(ones(2,2), ["a", "b"], [:one, :two]))
Axis{String,Int64,Array{String,1},Base.OneTo{Int64}}

AxisIndices.Interface.step_key — Method

step_key(x)

Returns the step size of the keys of x.

Examples

julia> using AxisIndices

julia> AxisIndices.step_key(Axis(1:2:10))
2

julia> AxisIndices.step_key(rand(2))
1

julia> AxisIndices.step_key([1])  # LinearIndices are treate like unit ranges
1

AxisIndices.Interface.unsafe_reconstruct — Method

unsafe_reconstruct(A::AbstractArray, new_parent, new_axes)

Reconstructs an AbstractArray of the same type as A but with the parent array parent and axes axes. This method depends on an underlying call to similar_types. It is considered unsafe because it bypasses safety checks to ensure the keys of each axis are unique and match the length of each dimension of parent. Therefore, this is not intended for interactive use and should only be used when it is clear all arguments are composed correctly.

AxisIndices.Interface.unsafe_reconstruct — Method

unsafe_reconstruct(axis, keys, indices)

Reconstructs an AbstractAxis of the same type as axis but with keys of type Ks and indices of type Vs. This method is considered unsafe because it bypasses checks to ensure that keys and values have the same length and the all keys are unique.

AxisIndices.Interface.unsafe_reconstruct — Method

unsafe_reconstruct(axis, indices)

Reconstructs an AbstractSimpleAxis of the same type as axis but values of type Vs.

AxisIndices.Interface.@defdim — Macro

@defdim name condition

Produces a series of methods for conveniently manipulating dimensions with specific names. condition is a method that returns true or false when given a name of a dimension. For example, the following would produce methods for manipulating and accessing dimensions with the name :time.

julia> is_time(x::Symbol) = x === :time

julia> @defdim time is_time

name is used to complete the following method names

name_dim(x): returns the dimension number of dimension

nname(x): returns the number of elements stored along the dimension

has_name_dim(x): returns true or false, indicating if the dimension is present

name_axis(x): returns the axis corresponding to the dimension.

name_indices(x): returns the indices corresponding to the dimension.

name_keys(x): returns the keys corresponding to the dimension

name_axis_type(x): returns the type of the axis corresponding to the dimension

select_name_dim(x, i): equivalent to selectdim(x, name_dim(x), i)

each_name(x): equivalent to eachslice(x, name_dim(x))

Warning

@defdim should be considered experimental and subject to change

AxisIndices.Interface.to_index — Method

to_index(axis, arg) -> to_index(AxisIndicesStyle(axis, arg), axis, arg)

Unique implementation of to_index for the AxisIndices package that specializes based on each axis and indexing argument (as opposed to the array and indexing argument).

AxisIndices.Interface.to_keys — Method

to_keys([::AxisIndicesStyle,] axis, arg, index)

This method is the reverse of AxisIndices.to_index. arg refers to an argument originally passed to AxisIndices.to_index and index refers to the index produced by that same call to AxisIndices.to_index.

This method assumes to all arguments have passed through AxisIndices.to_index and have been checked to be in bounds. Therefore, this is unsafe and intended only for internal use.

AxisIndices.Axes.AbstractAxis — Type

AbstractAxis

An AbstractVector subtype optimized for indexing.

AxisIndices.Axes.AbstractSimpleAxis — Type

AbstractSimpleAxis{V,Vs}

A subtype of AbstractAxis where the keys and values are represented by a single collection.

AxisIndices.Axes.Axis — Type

Axis(k[, v=OneTo(length(k))])

Subtypes of AbstractAxis that maps keys to values. The first argument specifies the keys and the second specifies the values. If only one argument is specified then the values span from 1 to the length of k.

Examples

The value for all of these is the same.

julia> using AxisIndices

julia> x = Axis(2.0:11.0, 1:10)
Axis(2.0:1.0:11.0 => 1:10)

julia> y = Axis(2.0:11.0)  # when only one argument is specified assume it's the keys
Axis(2.0:1.0:11.0 => Base.OneTo(10))

julia> z = Axis(1:10)
Axis(1:10 => Base.OneTo(10))

Standard indexing returns the same values

julia> x[2]
2

julia> x[2] == y[2] == z[2]
true

julia> x[1:2]
Axis(2.0:1.0:3.0 => 1:2)

julia> y[1:2]
Axis(2.0:1.0:3.0 => 1:2)

julia> z[1:2]
Axis(1:2 => 1:2)

julia> x[1:2] == y[1:2] == z[1:2]
true

Functions that return true or false may be used to search the keys for their corresponding index. The following is equivalent to the previous example.

julia> x[==(3.0)]
2

julia> x[==(3.0)] ==       # 3.0 is the 2nd key of x
       y[isequal(3.0)] ==  # 3.0 is the 2nd key of y
       z[==(2)]            # 2 is the 2nd key of z
true

julia> x[<(4.0)]  # all keys less than 4.0 are 2.0:3.0 which correspond to values 1:2
Axis(2.0:1.0:3.0 => 1:2)

julia> y[<=(3.0)]  # all keys less than or equal to 3.0 are 2.0:3.0 which correspond to values 1:2
Axis(2.0:1.0:3.0 => 1:2)

julia> z[<(3)]  # all keys less than or equal to 3 are 1:2 which correspond to values 1:2
Axis(1:2 => 1:2)

julia> x[<(4.0)] == y[<=(3.0)] == z[<(3)]
true

Notice that == returns a single value instead of a collection of all elements where the key was found to be true. This is because all keys must be unique so there can only ever be one element returned.

AxisIndices.Axes.CartesianAxes — Type

CartesianAxes

Alias for LinearIndices where indices are subtypes of AbstractAxis.

Examples

julia> using AxisIndices

julia> cartaxes = CartesianAxes((Axis(2.0:5.0), Axis(1:4)));

julia> cartinds = CartesianIndices((1:4, 1:4));

julia> cartaxes[2, 2]
CartesianIndex(2, 2)

julia> cartinds[2, 2]
CartesianIndex(2, 2)

AxisIndices.Axes.CenteredAxis — Type

CenteredAxis(indices)

A CenteredAxis takes indices and provides a user facing set of keys centered around zero. The CenteredAxis is a subtype of AbstractOffsetAxis and its keys are treated as the predominant indexing style. Note that the element type of a CenteredAxis cannot be unsigned because any instance with a length greater than 1 will begin at a negative value.

Examples

A CenteredAxis sends all indexing arguments to the keys and only maps to the indices when to_index is called.

julia> using AxisIndices

julia> axis = CenteredAxis(1:10)
CenteredAxis(-5:4 => 1:10)

julia> axis[10]  # the indexing goes straight to keys and is centered around zero
ERROR: BoundsError: attempt to access 10-element CenteredAxis(-5:4 => 1:10) at index [10]
[...]

julia> axis[-5]
-5

julia> AxisIndices.to_index(axis, -5)
1

AxisIndices.Axes.LinearAxes — Type

LinearAxes

Alias for LinearIndices where indices are subtypes of AbstractAxis.

Examples

julia> using AxisIndices

julia> linaxes = LinearAxes((Axis(2.0:5.0), Axis(1:4)));

julia> lininds = LinearIndices((1:4, 1:4));

julia> linaxes[2, 2]
6

julia> lininds[2, 2]
6

AxisIndices.Axes.MetaAxis — Type

MetaAxis

An axis type that allows storage of arbitraty metadata.

AxisIndices.Axes.MetaCartesianAxes — Type

MetaCartesianAxes

Conveniently construct a CartesianAxes that has metadata.

Examples

julia> using AxisIndices

julia> MetaCartesianAxes(([:a, :b], ["one", "two"]); metadata="some metadata")
2×2 MetaCartesianAxes{CartesianIndex{2},2}
 • dim_1 - [:a, :b]
 • dim_2 - ["one", "two"]
metadata: String
 • some metadata
                       one                    two
  a   CartesianIndex(1, 1)   CartesianIndex(1, 2)
  b   CartesianIndex(2, 1)   CartesianIndex(2, 2)

AxisIndices.Axes.MetaLinearAxes — Type

MetaLinearAxes

Conveniently construct a LinearAxes that has metadata.

Examples

julia> using AxisIndices

julia> MetaLinearAxes(([:a, :b], ["one", "two"]); metadata="some metadata")
2×2 MetaLinearAxes{Int64,2}
 • dim_1 - [:a, :b]
 • dim_2 - ["one", "two"]
metadata: String
 • some metadata
      one   two
  a     1     3
  b     2     4

AxisIndices.Axes.NamedCartesianAxes — Type

NamedCartesianAxes

Conveniently construct a CartesianAxes where each dimension has a name.

Examples

julia> using AxisIndices

julia> x = NamedCartesianAxes{(:dimx, :dimy)}(([:a, :b], ["one", "two"]))
2×2 NamedCartesianAxes{CartesianIndex{2},2}
 • dimx - [:a, :b]
 • dimy - ["one", "two"]
                       one                    two
  a   CartesianIndex(1, 1)   CartesianIndex(1, 2)
  b   CartesianIndex(2, 1)   CartesianIndex(2, 2)

julia> x == NamedCartesianAxes((dimx = [:a, :b], dimy = ["one", "two"]))
true

AxisIndices.Axes.NamedLinearAxes — Type

NamedLinearAxes

Provides LinearAxes where each dimension has a name.

Examples

julia> using AxisIndices

julia> x = NamedLinearAxes{(:dimx,:dimy)}(([:a, :b], ["one", "two"]))
2×2 NamedLinearAxes{Int64,2}
 • dimx - [:a, :b]
 • dimy - ["one", "two"]
      one   two
  a     1     3
  b     2     4

julia> x == NamedLinearAxes((dimx = [:a, :b], dimy = ["one", "two"]))
true

AxisIndices.Axes.NamedMetaCartesianAxes — Type

NamedMetaCartesianAxes

Conveniently construct a CartesianAxes that has metadata and each dimension has a name.

Examples

julia> using AxisIndices

julia> x = NamedMetaCartesianAxes{(:dimx,:dimy)}(([:a, :b], ["one", "two"]); metadata="some metadata")
2×2 NamedMetaCartesianAxes{CartesianIndex{2},2}
 • dimx - [:a, :b]
 • dimy - ["one", "two"]
metadata: String
 • some metadata
                       one                    two
  a   CartesianIndex(1, 1)   CartesianIndex(1, 2)
  b   CartesianIndex(2, 1)   CartesianIndex(2, 2)

julia> x == NamedMetaCartesianAxes((dimx = [:a, :b], dimy = ["one", "two"]); metadata="some metadata")
true

AxisIndices.Axes.NamedMetaLinearAxes — Type

NamedMetaLinearAxes

Conveniently construct a LinearAxes that has metadata and each dimension has a name.

Examples

julia> using AxisIndices

julia> x = NamedMetaLinearAxes{(:dimx,:dimy)}(([:a, :b], ["one", "two"]); metadata="some metadata")
2×2 NamedMetaLinearAxes{Int64,2}
 • dimx - [:a, :b]
 • dimy - ["one", "two"]
metadata: String
 • some metadata
      one   two
  a     1     3
  b     2     4

julia> x == NamedMetaLinearAxes((dimx = [:a, :b], dimy = ["one", "two"]); metadata="some metadata")
true

AxisIndices.Axes.OffsetAxis — Type

OffsetAxis(keys::AbstractUnitRange{<:Integer}, indices::AbstractUnitRange{<:Integer}[, check_length::Bool=true])
OffsetAxis(offset::Integer, indices::AbstractUnitRange{<:Integer})

An axis that has the indexing behavior of an AbstractOffsetAxis and retains an offset from its underlying indices in its keys.

Examples

Users may construct an OffsetAxis by providing an from a set of indices.

julia> using AxisIndices

julia> axis = OffsetAxis(-2, 1:3)
OffsetAxis(-1:1 => 1:3)

In this instance the first index of the wrapped indices is 1 (firstindex(indices(axis))) but adding the offset (-2) moves it to -1.

julia> firstindex(axis)
-1

julia> axis[-1]
-1

Similarly, the last index is move by -2.

julia> lastindex(axis)
1

julia> axis[1]
1

This means that traditional one based indexing no longer applies and may result in errors.

julia> axis[3]
ERROR: BoundsError: attempt to access 3-element OffsetAxis(-1:1 => 1:3) at index [3]
[...]

When an OffsetAxis is reconstructed the offset from indices are presserved.

julia> axis[0:1]  # offset of -2 still applies
OffsetAxis(0:1 => 2:3)

AxisIndices.Axes.SimpleAxis — Type

SimpleAxis(v)

Povides an AbstractAxis interface for any AbstractUnitRange, v. v will be considered both the values and keys of the return instance.

Examples

A SimpleAxis is useful for giving a standard set of indices the ability to use the filtering syntax for indexing.

julia> using AxisIndices, StaticRanges

julia> x = SimpleAxis(2:10)
SimpleAxis(2:10)

julia> x[2]
2

julia> x[==(2)]
2

julia> x[2] == x[==(2)]  # keys and values are same
true

julia> x[>(2)]
SimpleAxis(3:10)

julia> x[>(2)]
SimpleAxis(3:10)

julia> x[1]
ERROR: BoundsError: attempt to access 9-element SimpleAxis(2:10 => 2:10) at index [1]
[...]

AxisIndices.Axes.SimpleAxis — Method

SimpleAxis(start::Integer, stop::Integer) -> SimpleAxis{UnitRange{Integer}}

Passes start and stop arguments to UnitRange to construct the values of SimpleAxis.

Examples

julia> using AxisIndices

julia> SimpleAxis(1, 10)
SimpleAxis(1:10)

AxisIndices.Axes.SimpleAxis — Method

SimpleAxis(stop::Integer) -> SimpleAxis{Base.OneTo{Integer}}

Passes stop Base.OneTo to construct the values of SimpleAxis.

Examples

julia> using AxisIndices

julia> SimpleAxis(10)
SimpleAxis(Base.OneTo(10))

AxisIndices.Axes.StructAxis — Type

StructAxis{T}

An axis that uses a structure T to form its keys. the field names of

AxisIndices.Axes.structview — Method

structview(A)

Creates a MappedArray using the StructAxis of A to identify the dimension that needs to be collapsed into a series of SubArrays as views that composed the MappedArray

AxisIndices.Axes.AbstractOffsetAxis — Type

AbstractOffsetAxis{I,Ks,Inds}

Supertype for axes that begin indexing offset from one. All subtypes of AbstractOffsetAxis use the keys for indexing and only convert to the underlying indices when to_index(::OffsetAxis, ::Integer) is called (i.e. when indexing the an array with an AbstractOffsetAxis. See OffsetAxis, CenteredAxis, and IdentityAxis for more details and examples.

AxisIndices.Axes.permute_axes — Method

permute_axes(x::AbstractVector)

Returns the permuted axes of x as axes of size 1 × length(x)

Examples

julia> using AxisIndices, StaticRanges

julia> length.(AxisIndices.permute_axes(rand(4))) == (1, 4)
true

julia> AxisIndices.permute_axes((Axis(1:4),))
(SimpleAxis(Base.OneTo(1)), Axis(1:4 => Base.OneTo(4)))

julia> AxisIndices.permute_axes((Axis(mrange(1, 4)),))
(SimpleAxis(OneToMRange(1)), Axis(UnitMRange(1:4) => OneToMRange(4)))

julia> AxisIndices.permute_axes((Axis(srange(1, 4)),))
(SimpleAxis(OneToSRange(1)), Axis(UnitSRange(1:4) => OneToSRange(4)))

AxisIndices.Axes.permute_axes — Method

permute_axes(m::AbstractMatrix) -> NTuple{2}

Permute the axes of the matrix m, by flipping the elements across the diagonal of the matrix. Differs from LinearAlgebra's transpose in that the operation is not recursive.

Examples

julia> using AxisIndices

julia> AxisIndices.permute_axes(rand(4, 2))
(Base.OneTo(2), Base.OneTo(4))

julia> AxisIndices.permute_axes((Axis(1:4), Axis(1:2)))
(Axis(1:2 => Base.OneTo(2)), Axis(1:4 => Base.OneTo(4)))

AxisIndices.Axes.permute_axes — Method

permute_axes(x::AbstractArray, perms::Tuple) = permute_axes(axes(x), p)
permute_axes(x::NTuple{N}, perms::NTuple{N}) -> NTuple{N}

Returns axes of x in the order of p.

Examples

julia> using AxisIndices

julia> AxisIndices.permute_axes(rand(2, 4, 6), (1, 3, 2))
(Base.OneTo(2), Base.OneTo(6), Base.OneTo(4))

julia> AxisIndices.permute_axes((Axis(1:2), Axis(1:4), Axis(1:6)), (1, 3, 2))
(Axis(1:2 => Base.OneTo(2)), Axis(1:6 => Base.OneTo(6)), Axis(1:4 => Base.OneTo(4)))

Base.similar — Function

similar(axis::AbstractAxis, new_keys::AbstractVector, new_indices::AbstractUnitRange{Integer} [, check_length::Bool=true] ) -> AbstractAxis

Create a new instance of an axis of the same type as axis but with the keys new_keys and indices new_indices. If check_length is true then the lengths of new_keys and new_indices are checked to ensure they have the same length before construction.

Examples

julia> using AxisIndices

julia> similar(Axis(1.0:10.0, 1:10), [:one, :two], UInt(1):UInt(2))
Axis([:one, :two] => 0x0000000000000001:0x0000000000000002)

julia> similar(Axis(1.0:10.0, 1:10), [:one, :two], UInt(1):UInt(3))
ERROR: DimensionMismatch("keys and indices must have same length, got length(keys) = 2 and length(indices) = 3.")
[...]

Base.similar — Method

similar(axis::AbstractAxis, new_keys::AbstractVector) -> AbstractAxis

Create a new instance of an axis of the same type as axis but with the keys new_keys

Examples

julia> using AxisIndices

julia> similar(Axis(1.0:10.0, 1:10), [:one, :two])
Axis([:one, :two] => 1:2)

Base.similar — Method

similar(axis::AbstractAxis, new_indices::AbstractUnitRange)

Create a new instance of an axis of the same type as axis but with the keys new_keys

Examples

julia> using AxisIndices

julia> similar(SimpleAxis(1:10), 1:3)
SimpleAxis(1:3)