Experimental

Currently there are several experimental modules included. Methods from these modules are not exported by default but may be accessed by using AxisIndices.<NameOfExperimentalModule> These are considered experimental because the are either undergoing rapid change, are likely to end up in a different package entirely, or both.

ObservationDims

AxisIndices.ObservationDims.has_obsdim — Method

has_obsdim(x) -> Bool

Returns true if x has a dimension corresponding to obs.

AxisIndices.ObservationDims.nobs — Method

nobs(x) -> Int

Returns the size along the dimension corresponding to the obs.

AxisIndices.ObservationDims.obs_axis — Method

obs_axis(x)

Returns the axis corresponding to the obs dimension.

AxisIndices.ObservationDims.obs_axis_type — Method

obs_axis_type(x)

Returns the key type corresponding to the obs axis.

AxisIndices.ObservationDims.obs_indices — Method

obs_indices(x)

Returns the indices corresponding to the obs axis

AxisIndices.ObservationDims.obs_keys — Method

obs_keys(x)

Returns the keys corresponding to the obs axis

AxisIndices.ObservationDims.obsdim — Method

obsdim(x) -> Int

Returns the dimension corresponding to the obs.

AxisIndices.ObservationDims.select_obsdim — Method

select_obsdim(x, i)

Return a view of all the data of x where the index for the obs dimension equals i.

AxisIndices.ObservationDims.each_obs — Method

each_obs(x)

Create a generator that iterates over the obs dimensions A, returning views that select all the data from the other dimensions in A.

Named Dimension Methods

AxisIndices.Names.@defdim — Macro

@defdim name condition

Produces a series of methods for conveniently manipulating dimensions with specific names. condition is a method that returns true or false when given a name of a dimension. For example, the following would produce methods for manipulating and accessing dimensions with the name :time.

julia> is_time(x::Symbol) = x === :time

julia> @defdim time is_time

name is used to complete the following method names

name_dim(x): returns the dimension number of dimension

nname(x): returns the number of elements stored along the dimension

has_name_dim(x): returns true or false, indicating if the dimension is present

name_axis(x): returns the axis corresponding to the dimension.

name_indices(x): returns the indices corresponding to the dimension.

name_keys(x): returns the keys corresponding to the dimension

name_axis_type(x): returns the type of the axis corresponding to the dimension

select_name_dim(x, i): equivalent to selectdim(x, name_dim(x), i)

each_name(x): equivalent to eachslice(x, name_dim(x))