Experimental

Currently there are several experimental modules included. Methods from these modules are not exported by default but may be accessed by using AxisIndices.<NameOfExperimentalModule> These are considered experimental because the are either undergoing rapid change, are likely to end up in a different package entirely, or both.

ObservationDims

AxisIndices.ObservationDims.has_obsdim — Method

has_obsdim(x) -> Bool

Returns true if x has a dimension corresponding to obs.

AxisIndices.ObservationDims.nobs — Method

nobs(x) -> Int

Returns the size along the dimension corresponding to the obs.

AxisIndices.ObservationDims.obs_axis — Method

obs_axis(x)

Returns the axis corresponding to the obs dimension.

AxisIndices.ObservationDims.obs_axis_type — Method

obs_axis_type(x)

Returns the key type corresponding to the obs axis.

AxisIndices.ObservationDims.obs_indices — Method

obs_indices(x)

Returns the indices corresponding to the obs axis

AxisIndices.ObservationDims.obs_keys — Method

obs_keys(x)

Returns the keys corresponding to the obs axis

AxisIndices.ObservationDims.obsdim — Method

obsdim(x) -> Int

Returns the dimension corresponding to the obs.

AxisIndices.ObservationDims.select_obsdim — Method

select_obsdim(x, i)

Return x view of all the data of x where the index for the obs dimension equals i.

TimeDims

AxisIndices.TimeDims.duration — Method

duration(x)

Duration of the event along the time axis.

AxisIndices.TimeDims.has_timedim — Method

has_timedim(x) -> Bool

Returns true if x has a dimension corresponding to time.

AxisIndices.TimeDims.ntime — Method

ntime(x) -> Int

Returns the size along the dimension corresponding to the time.

AxisIndices.TimeDims.onset — Method

onset(x)

First time point along the time axis.

AxisIndices.TimeDims.sampling_rate — Method

sampling_rate(x)

Number of samples per second.

AxisIndices.TimeDims.select_timedim — Method

select_timedim(x, i)

Return x view of all the data of x where the index for the time dimension equals i.

AxisIndices.TimeDims.time_axis — Method

time_axis(x)

Returns the axis corresponding to the time dimension.

AxisIndices.TimeDims.time_axis_type — Method

time_axis_type(x)

Returns the key type corresponding to the time axis.

AxisIndices.TimeDims.time_end — Method

time_end(x)

Last time point along the time axis.

AxisIndices.TimeDims.time_indices — Method

time_indices(x)

Returns the indices corresponding to the time axis

AxisIndices.TimeDims.time_keys — Method

time_keys(x)

Returns the keys corresponding to the time axis

AxisIndices.TimeDims.timedim — Method

timedim(x) -> Int

Returns the dimension corresponding to the time.

ColorDims

AxisIndices.ColorDims.color_axis — Method

color_axis(x)

Returns the axis corresponding to the color dimension.

AxisIndices.ColorDims.color_axis_type — Method

color_axis_type(x)

Returns the key type corresponding to the color axis.

AxisIndices.ColorDims.color_indices — Method

color_indices(x)

Returns the indices corresponding to the color axis

AxisIndices.ColorDims.color_keys — Method

color_keys(x)

Returns the keys corresponding to the color axis

AxisIndices.ColorDims.colordim — Method

colordim(x) -> Int

Returns the dimension corresponding to the color.

AxisIndices.ColorDims.has_colordim — Method

has_colordim(x) -> Bool

Returns true if x has a dimension corresponding to color.

AxisIndices.ColorDims.ncolor — Method

ncolor(x) -> Int

Returns the size along the dimension corresponding to the color.

AxisIndices.ColorDims.select_colordim — Method

select_colordim(x, i)

Return x view of all the data of x where the index for the color dimension equals i.

SpatialDims

AxisIndices.SpatialDims.pixel_spacing — Method

pixel_spacing(x)

Return a tuple representing the separation between adjacent pixels along each axis of the image. Derived from the step size of each element of spatial_keys.

AxisIndices.SpatialDims.spatial_axes — Method

spatial_axes(x) -> Tuple

Returns a tuple of each axis corresponding to a spatial dimensions.

AxisIndices.SpatialDims.spatial_indices — Method

spatial_indices(x)

Return a tuple with the indices of the spatial dimensions of the image. Defaults to the same as indices, but using NamedDimsArrah you can mark some axes as being non-spatial.

AxisIndices.SpatialDims.spatial_keys — Method

spatial_keys(x)

AxisIndices.SpatialDims.spatial_offset — Method

spatial_offset(x)

The offset of each dimension (i.e., where each spatial axis starts).

AxisIndices.SpatialDims.spatial_order — Method

spatial_order(x) -> Tuple{Vararg{Symbol}}

Returns the dimnames of x that correspond to spatial dimensions.

AxisIndices.SpatialDims.spatial_size — Method

spatial_size(x) -> Tuple{Vararg{Int}}

Return a tuple listing the sizes of the spatial dimensions of the image.

AxisIndices.SpatialDims.spatialdims — Method

spatialdims(x) -> Tuple{Vararg{Int}}

Return a tuple listing the spatial dimensions of img. Note that a better strategy may be to use ImagesAxes and take slices along the time axis.