References

AxisIndices is composed of several submodules corresponding to each of the following subsections.

<!–TODO separate into public and private API –>

AxisCore

AxisIndices.AxisCore.AbstractAxis — Type

AbstractAxis

An AbstractVector subtype optimized for indexing.

AxisIndices.AxisCore.AbstractAxisIndices — Type

AbstractAxisIndices

AbstractAxisIndices is a subtype of AbstractArray that offers integration with the AbstractAxis interface. The only methods that absolutely needs to be defined for a subtype of AbstractAxisIndices are axes, parent, similar_type, and similar. Most users should find the provided AxisIndicesArray subtype is sufficient for the majority of use cases. Although custom behavior may be accomplished through a new subtype of AbstractAxisIndices, customizing the behavior of many methods described herein can be accomplished through a unique subtype of AbstractAxis.

This implementation is meant to be basic, well documented, and have sane defaults that can be overridden as necessary. In other words, default methods for manipulating arrays that return an AxisIndicesArray should not cause unexpected downstream behavior for users; and developers should be able to freely customize the behavior of AbstractAxisIndices subtypes with minimal effort.

AxisIndices.AxisCore.AbstractSimpleAxis — Type

AbstractSimpleAxis{V,Vs}

A subtype of AbstractAxis where the keys and values are represented by a single collection.

AxisIndices.AxisCore.Axis — Type

Axis(k[, v=OneTo(length(k))])

Subtypes of AbstractAxis that maps keys to values. The first argument specifies the keys and the second specifies the values. If only one argument is specified then the values span from 1 to the length of k.

Examples

The value for all of these is the same.

julia> using AxisIndices

julia> x = Axis(2.0:11.0, 1:10)
Axis(2.0:1.0:11.0 => 1:10)

julia> y = Axis(2.0:11.0)  # when only one argument is specified assume it's the keys
Axis(2.0:1.0:11.0 => Base.OneTo(10))

julia> z = Axis(1:10)
Axis(1:10 => Base.OneTo(10))

Standard indexing returns the same values

julia> x[2]
2

julia> x[2] == y[2] == z[2]
true

julia> x[1:2]
Axis(2.0:1.0:3.0 => 1:2)

julia> y[1:2]
Axis(2.0:1.0:3.0 => 1:2)

julia> z[1:2]
Axis(1:2 => 1:2)

julia> x[1:2] == y[1:2] == z[1:2]
true

Functions that return true or false may be used to search the keys for their corresponding index. The following is equivalent to the previous example.

julia> x[==(3.0)]
2

julia> x[==(3.0)] ==       # 3.0 is the 2nd key of x
       y[isequal(3.0)] ==  # 3.0 is the 2nd key of y
       z[==(2)]            # 2 is the 2nd key of z
true

julia> x[<(4.0)]  # all keys less than 4.0 are 2.0:3.0 which correspond to values 1:2
Axis(2.0:1.0:3.0 => 1:2)

julia> y[<=(3.0)]  # all keys less than or equal to 3.0 are 2.0:3.0 which correspond to values 1:2
Axis(2.0:1.0:3.0 => 1:2)

julia> z[<(3)]  # all keys less than or equal to 3 are 1:2 which correspond to values 1:2
Axis(1:2 => 1:2)

julia> x[<(4.0)] == y[<=(3.0)] == z[<(3)]
true

Notice that == returns a single value instead of a collection of all elements where the key was found to be true. This is because all keys must be unique so there can only ever be one element returned.

AxisIndices.AxisCore.AxisIndicesArray — Type

AxisIndicesArray{T,N,P,AI}

An array struct that wraps any parent array and assigns it an AbstractAxis for each dimension. The first argument is the parent array and the second argument is a tuple of subtypes to AbstractAxis or keys that will be converted to subtypes of AbstractAxis with the provided keys.

AxisIndices.AxisCore.AxisIndicesArray — Method

AxisIndicesArray(parent::AbstractArray, args...) -> AxisIndicesArray(parent, tuple(args))

Passes args to a tuple for constructing an AxisIndicesArray.

Examples

julia> using AxisIndices

julia> A = AxisIndicesArray(reshape(1:9, 3,3), 2:4, 3.0:5.0);

julia> A[1, 1]
1

julia> A[==(2), ==(3.0)]
1

julia> A[1:2, 1:2] == [1 4; 2 5]
true

julia> A[<(4), <(5.0)] == [1 4; 2 5]
true

AxisIndices.AxisCore.AxisIndicesArray — Method

AxisIndicesArray(parent::AbstractArray, axes::Tuple{Vararg{AbstractAxis}}[, check_length=true])

Construct an AxisIndicesArray using parent and explicit subtypes of AbstractAxis. If check_length is true then each dimension of parent's length is checked to match the length of the corresponding axis (e.g., size(parent 1) == length(axes[1]).

Examples

julia> using AxisIndices

julia> AxisIndicesArray(ones(2,2), (SimpleAxis(2), SimpleAxis(2)))
AxisIndicesArray{Float64,2,Array{Float64,2}...}
 • dim_1 - SimpleAxis(Base.OneTo(2))
 • dim_2 - SimpleAxis(Base.OneTo(2))
        1     2
  1   1.0   1.0
  2   1.0   1.0

AxisIndices.AxisCore.AxisIndicesArray — Method

AxisIndicesArray(parent::AbstractArray, keys::Tuple[, values=axes(parent), check_length=true])

Given an the some array parent and a tuple of vectors keys corresponding to each dimension of parent constructs an AxisIndicesArray. Each element of keys is paired with an element of values to compose a subtype of AbstractAxis. values map the keys to the indices of parent.

Examples

julia> using AxisIndices

julia> AxisIndicesArray(ones(2,2), (["a", "b"], ["one", "two"]))
AxisIndicesArray{Float64,2,Array{Float64,2}...}
 • dim_1 - Axis(["a", "b"] => Base.OneTo(2))
 • dim_2 - Axis(["one", "two"] => Base.OneTo(2))
      one   two
  a   1.0   1.0
  b   1.0   1.0

AxisIndices.AxisCore.SimpleAxis — Type

SimpleAxis(v)

Povides an AbstractAxis interface for any AbstractUnitRange, v. v will be considered both the values and keys of the return instance.

Examples

A SimpleAxis is useful for giving a standard set of indices the ability to use the filtering syntax for indexing.

julia> using AxisIndices

julia> x = SimpleAxis(2:10)
SimpleAxis(2:10)

julia> x[2]
2

julia> x[==(2)]
2

julia> x[2] == x[==(2)]  # keys and values are same
true

julia> x[>(2)]
SimpleAxis(3:10)

julia> x[>(2)]
SimpleAxis(3:10)

julia> x[1]
ERROR: BoundsError: attempt to access 9-element SimpleAxis(2:10)) at index [1]
[...]

AxisIndices.AxisCore.SimpleAxis — Method

SimpleAxis(start::Integer, stop::Integer) -> SimpleAxis{UnitRange{Integer}}

Passes start and stop arguments to UnitRange to construct the values of SimpleAxis.

Examples

julia> using AxisIndices

julia> SimpleAxis(1, 10)
SimpleAxis(1:10)

AxisIndices.AxisCore.SimpleAxis — Method

SimpleAxis(stop::Integer) -> SimpleAxis{Base.OneTo{Integer}}

Passes stop Base.OneTo to construct the values of SimpleAxis.

Examples

julia> using AxisIndices

julia> SimpleAxis(10)
SimpleAxis(Base.OneTo(10))

AxisIndices.AxisCore.axes_keys — Method

axes_keys(x, i)

Returns the axis keys corresponding of ith dimension of x.

Examples

julia> using AxisIndices

julia> axes_keys(AxisIndicesArray(ones(2,2), (2:3, 3:4)), 1)
2:3

AxisIndices.AxisCore.axes_keys — Method

axes_keys(x::AbstractArray)

Returns the keys corresponding to all axes of x.

Examples

julia> using AxisIndices

julia> axes_keys(AxisIndicesArray(ones(2,2), (2:3, 3:4)))
(2:3, 3:4)

julia> axes_keys(Axis(1:2))
(1:2,)

AxisIndices.AxisCore.first_key — Method

first_key(x)

Returns the first key of x.

Examples

julia> using AxisIndices

julia> first_key(Axis(2:10))
2

AxisIndices.AxisCore.indices — Method

indices(x::AbstractAxis)

Returns the indices x.

Examples

julia> using AxisIndices

julia> indices(Axis(["a"], 1:1))
1:1

julia> indices(CartesianIndex(1,1))
(1, 1)

AxisIndices.AxisCore.indices — Method

indices(x, i)

Returns the indices corresponding to the i axis

Examples

julia> using AxisIndices

julia> indices(AxisIndicesArray(ones(2,2), (2:3, 3:4)), 1)
Base.OneTo(2)

AxisIndices.AxisCore.indices — Method

indices(x)

Returns the indices corresponding to all axes of x.

Examples

julia> using AxisIndices

julia> indices(AxisIndicesArray(ones(2,2), (2:3, 3:4)))
(Base.OneTo(2), Base.OneTo(2))

julia> indices(Axis(["a"], 1:1))
1:1

julia> indices(CartesianIndex(1,1))
(1, 1)

AxisIndices.AxisCore.keys_type — Method

keys_type(x, i)

Retrieves axis keys of the ith dimension of x.

Examples

julia> using AxisIndices

julia> keys_type(AxisIndicesArray([1], ["a"]), 1)
Array{String,1}

AxisIndices.AxisCore.keys_type — Method

keys_type(x)

Retrieves the type of the keys of x.

Examples

julia> using AxisIndices

julia> keys_type(Axis(1:2))
UnitRange{Int64}

julia> keys_type(typeof(Axis(1:2)))
UnitRange{Int64}

julia> keys_type(UnitRange{Int})
Base.OneTo{Int64}

AxisIndices.AxisCore.last_key — Method

last_key(x)

Returns the last key of x.

Examples

julia> using AxisIndices

julia> last_key(Axis(2:10))
10

AxisIndices.AxisCore.step_key — Method

step_key(x)

Returns the step size of the keys of x.

Examples

julia> using AxisIndices

julia> AxisIndices.step_key(Axis(1:2:10))
2

julia> AxisIndices.step_key(rand(2))
1

julia> AxisIndices.step_key([1])  # LinearIndices are treate like unit ranges
1

AxisIndices.AxisCore.unsafe_reconstruct — Method

unsafe_reconstruct(A::AbstractAxisIndices, parent, axes)

Reconstructs an AbstractAxisIndices of the same type as A but with the parent array parent and axes axes. This method depends on an underlying call to similar_types. It is considered unsafe because it bypasses safety checks to ensure the keys of each axis are unique and match the length of each dimension of parent. Therefore, this is not intended for interactive use and should only be used when it is clear all arguments are composed correctly.

AxisIndices.AxisCore.unsafe_reconstruct — Method

unsafe_reconstruct(axis::AbstractSimpleAxis, values::Vs)

Reconstructs an AbstractSimpleAxis of the same type as axis but values of type Vs.

AxisIndices.AxisCore.unsafe_reconstruct — Method

unsafe_reconstruct(axis::AbstractAxis, keys::Ks, values::Vs)

Reconstructs an AbstractAxis of the same type as axis but with keys of type Ks and values of type Vs. This method is considered unsafe because it bypasses checks to ensure that keys and values have the same length and the all keys are unique.

AxisIndices.AxisCore.values_type — Method

values_type(x, i)

Retrieves axis values of the ith dimension of x.

Examples

julia> using AxisIndices

julia> values_type([1], 1)
Base.OneTo{Int64}

julia> values_type(typeof([1]), 1)
Base.OneTo{Int64}

AxisIndices.AxisCore.values_type — Method

values_type(x)

Retrieves the type of the values of x. This should be functionally equivalent to typeof(values(x)).

Examples

julia> using AxisIndices

julia>  values_type(Axis(1:2))
Base.OneTo{Int64}

julia> values_type(typeof(Axis(1:2)))
Base.OneTo{Int64}

julia> values_type(typeof(1:2))
UnitRange{Int64}

Base.similar — Method

similar(axis::AbstractAxis, new_keys::AbstractVector, new_indices::AbstractUnitRange{Integer} [, check_length::Bool=true] ) -> AbstractAxis

Create a new instance of an axis of the same type as axis but with the keys new_keys and indices new_indices. If check_length is true then the lengths of new_keys and new_indices are checked to ensure they have the same length before construction.

Examples

julia> using AxisIndices

julia> similar(Axis(1.0:10.0, 1:10), [:one, :two], UInt(1):UInt(2))
Axis([:one, :two] => 0x0000000000000001:0x0000000000000002)

julia> similar(Axis(1.0:10.0, 1:10), [:one, :two], UInt(1):UInt(3))
ERROR: keys and indices must have same length, got length(keys) = 2 and length(indices) = 3.
[...]

Base.similar — Method

similar(axis::AbstractAxis, new_keys::AbstractVector) -> AbstractAxis

Create a new instance of an axis of the same type as axis but with the keys new_keys

Examples

julia> using AxisIndices

julia> similar(Axis(1.0:10.0, 1:10), [:one, :two])
Axis([:one, :two] => 1:2)

Base.similar — Method

similar(axis::AbstractSimpleAxis, new_indices::AbstractUnitRange{Integer}) -> AbstractSimpleAxis

Create a new instance of an axis of the same type as axis but with the keys new_keys

Examples

julia> using AxisIndices

julia> similar(SimpleAxis(1:10), 1:3)
SimpleAxis(1:3)

AxisIndicesStyles

AxisIndices.AxisIndicesStyles.AxisIndicesStyle — Type

AxisIndicesStyle

Supertype for traits that control the behavior of AxisIndices.to_index and AxisIndices.to_key.

AxisIndices.AxisIndicesStyles.BoolElement — Type

BoolElement

A subtype of AxisIndicesStyle for mapping an argument that refers to a single Bool to a single index.

AxisIndices.AxisIndicesStyles.BoolsCollection — Type

BoolsCollection

A subtype of AxisIndicesStyle for mapping a collection of Bools to a collection of indices.

AxisIndices.AxisIndicesStyles.CartesianElement — Type

CartesianElement

A subtype of AxisIndicesStyle for mapping an argument that refers to a CartesianIndex to a single index.

AxisIndices.AxisIndicesStyles.IndexElement — Type

IndexElement

A subtype of AxisIndicesStyle for propagating an argument to a single index.

AxisIndices.AxisIndicesStyles.IndicesCollection — Type

IndicesCollection

A subtype of AxisIndicesStyle for propagating an argument to a collection of indices.

AxisIndices.AxisIndicesStyles.IntervalCollection — Type

IntervalCollection

A subtype of AxisIndicesStyle for mapping an interval argument (start..stop) from keys within said interval to a collection of indices.

AxisIndices.AxisIndicesStyles.KeyElement — Type

KeyElement

A subtype of AxisIndicesStyle for mapping an argument that refers to a single key to a single index.

AxisIndices.AxisIndicesStyles.KeyEquals — Type

KeyEquals

A subtype of AxisIndicesStyle for mapping a single key in a isequal(key) argument to a single index.

AxisIndices.AxisIndicesStyles.KeysCollection — Type

KeysCollection

A subtype of AxisIndicesStyle for mapping a collection of keys a collection of indices.

AxisIndices.AxisIndicesStyles.KeysFix2 — Type

KeysFix2

A subtype of AxisIndicesStyle for mapping all keys from fixed argument (e.g., >(key)) to the corresponding collection of indices.

AxisIndices.AxisIndicesStyles.KeysIn — Type

KeysIn

A subtype of AxisIndicesStyle for mapping all keys given in(keys) to a collection of indices.

AxisIndices.AxisIndicesStyles.SliceCollection — Type

SliceCollection

A subtype of AxisIndicesStyle indicating that the entire axis should be propagated.

AxisIndices.AxisIndicesStyles.is_collection — Method

is_collection(x) -> Bool

Whether x is an AxisIndicesStyle, returns true if it's used to index a collection of indices.

AxisIndices.AxisIndicesStyles.is_element — Method

is_element(x) -> Bool

Whether x is an AxisIndicesStyle, returns true if it's used to index a single element.

AxisIndices.AxisIndicesStyles.is_index — Method

is_index(x) -> Bool

Whether x is an AxisIndicesStyle, returns true if it's used to search the indices space.

AxisIndices.AxisIndicesStyles.is_key — Method

is_key(x) -> Bool

Whether x is an AxisIndicesStyle, returns true if it's used to search the keys space.

AxisIndices.AxisIndicesStyles.to_index — Method

to_index(axis, arg) -> to_index(AxisIndicesStyle(axis, arg), axis, arg)

Unique implementation of to_index for the AxisIndices package that specializes based on each axis and indexing argument (as opposed to the array and indexing argument).

AxisIndices.AxisIndicesStyles.to_keys — Method

to_keys([::AxisIndicesStyle,] axis, arg, index)

This method is the reverse of AxisIndices.to_index. arg refers to an argument originally passed to AxisIndices.to_index and index refers to the index produced by that same call to AxisIndices.to_index.

This method assumes to all arguments have passed through AxisIndices.to_index and have been checked to be in bounds. Therefore, this is unsafe and intended only for internal use.

Indexing

AxisIndices.Indexing.CartesianAxes — Type

CartesianAxes

Alias for LinearIndices where indices are subtypes of AbstractAxis.

Examples

julia> using AxisIndices

julia> cartaxes = CartesianAxes((Axis(2.0:5.0), Axis(1:4)));

julia> cartinds = CartesianIndices((1:4, 1:4));

julia> cartaxes[2, 2]
CartesianIndex(2, 2)

julia> cartinds[2, 2]
CartesianIndex(2, 2)

AxisIndices.Indexing.LinearAxes — Type

LinearAxes

Alias for LinearIndices where indices are subtypes of AbstractAxis.

Examples

julia> using AxisIndices

julia> linaxes = LinearAxes((Axis(2.0:5.0), Axis(1:4)));

julia> lininds = LinearIndices((1:4, 1:4));

julia> linaxes[2, 2]
6

julia> lininds[2, 2]
6

Basics

AxisIndices.Basics.CombineAxis — Type

CombineAxis

Subtype of CombineStyle that informs relevant methods to produce a subtype of AbstractAxis.

AxisIndices.Basics.CombineResize — Type

CombineResize

Subtype of CombineStyle that informs relevant methods that axes should be combined by resizing a collection (as opposed to by concatenation or appending).

AxisIndices.Basics.CombineSimpleAxis — Type

CombineSimpleAxis

Subtype of CombineStyle that informs relevant methods to produce a subtype of AbstractSimpleAxis.

AxisIndices.Basics.CombineStack — Type

CombineStack

Subtype of CombineStyle that informs relevant methods that axes should be combined by stacking elements in some whay (as opposed to resizing a collection).

AxisIndices.Basics.CombineStyle — Type

CombineStyle

Abstract type that determines the behavior of broadcast_axis, cat_axis, append_axis!.

AxisIndices.Basics.append_axis! — Method

append_axis!(x, y)
append_axis!(::CombineStyle, x, y)

Returns the appended axes x and y.

Examples

julia> using AxisIndices

julia> x, y = Axis(UnitMRange(1, 10)), SimpleAxis(UnitMRange(1, 10));

julia> AxisIndices.append_axis!(x, y);

julia> length(x)
20

julia> AxisIndices.append_axis!(y, x);

julia> length(y)
30

AxisIndices.Basics.broadcast_axis — Method

broadcast_axis(x, y) -> collection
broadcast_axis(::PromoteStyle, ::CombineStyle, x, y)

Returns an axis given the axes x and y that are appropriate for a broadcasting operation.

Examples

julia> using AxisIndices

julia> a, b = 1:10, Base.OneTo(10);

julia> AxisIndices.broadcast_axis(a, SimpleAxis(b))
SimpleAxis(1:10)

julia> AxisIndices.broadcast_axis(a, b)
1:10

julia> AxisIndices.broadcast_axis(b, a)
1:10

julia> c = AxisIndices.broadcast_axis(a, SimpleAxis(b))
SimpleAxis(1:10)

julia> d = AxisIndices.broadcast_axis(SimpleAxis(b), a)
SimpleAxis(1:10)

julia> e = AxisIndices.broadcast_axis(c, d)
SimpleAxis(1:10)

julia> f = AxisIndices.broadcast_axis(Axis(a), a)
Axis(1:10 => 1:10)

julia> g = AxisIndices.broadcast_axis(a, Axis(b))
Axis(Base.OneTo(10) => 1:10)

julia> AxisIndices.broadcast_axis(f, g)
Axis(1:10 => 1:10)

julia> AxisIndices.broadcast_axis(Base.OneTo(10), Base.OneTo(10))
Base.OneTo(10)

julia> AxisIndices.broadcast_axis(1:10, Base.OneTo(10))
1:10

Using combine allows combining axes that aren't appropriate for conventional promotion.

julia> AxisIndices.broadcast_axis(1:2, string.(1:2))
2-element Array{String,1}:
 "1"
 "2"

julia> AxisIndices.broadcast_axis(1:2, Symbol.(1:2))
2-element Array{Symbol,1}:
 Symbol("1")
 Symbol("2")

AxisIndices.Basics.cat_axes — Method

cat_axes(x::AbstractArray, y::AbstractArray, xy::AbstractArray, dims)

Produces the appropriate set of axes where x and y are the arrays that were concatenated over dims to produce xy. The appropriate indices of each axis are derived from from xy.

AxisIndices.Basics.cat_axis — Method

cat_axis(x, y) -> cat_axis(CombineStyle(x, y), x, y)
cat_axis(::CombineStyle, x, y) -> collection

Returns the concatenated axes x and y.

Examples

julia> using AxisIndices

julia> AxisIndices.cat_axis(Axis(UnitMRange(1, 10)), SimpleAxis(UnitMRange(1, 10)))
Axis(UnitMRange(1:20) => UnitMRange(1:20))

julia> AxisIndices.cat_axis(SimpleAxis(UnitMRange(1, 10)), SimpleAxis(UnitMRange(1, 10)))
SimpleAxis(UnitMRange(1:20))

AxisIndices.Basics.drop_axes — Method

drop_axes(x, dims)

Returns all axes of x except for those identified by dims. Elements of dims must be unique integers or symbols corresponding to the dimensions or names of dimensions of x.

Examples

julia> using AxisIndices

julia> axs = (Axis(1:5), Axis(1:10));

julia> AxisIndices.drop_axes(axs, 1)
(Axis(1:10 => Base.OneTo(10)),)

julia> AxisIndices.drop_axes(axs, 2)
(Axis(1:5 => Base.OneTo(5)),)

julia> AxisIndices.drop_axes(rand(2, 4), 2)
(Base.OneTo(2),)

AxisIndices.Basics.permute_axes — Method

permute_axes(x::AbstractVector)

Returns the permuted axes of x as axes of size 1 × length(x)

Examples

julia> using AxisIndices

julia> length.(AxisIndices.permute_axes(rand(4))) == (1, 4)
true

julia> AxisIndices.permute_axes((Axis(1:4),))
(SimpleAxis(Base.OneTo(1)), Axis(1:4 => Base.OneTo(4)))

julia> AxisIndices.permute_axes((Axis(mrange(1, 4)),))
(SimpleAxis(OneToMRange(1)), Axis(UnitMRange(1:4) => OneToMRange(4)))

julia> AxisIndices.permute_axes((Axis(srange(1, 4)),))
(SimpleAxis(OneToSRange(1)), Axis(UnitSRange(1:4) => OneToSRange(4)))

AxisIndices.Basics.permute_axes — Method

permute_axes(m::AbstractMatrix) -> NTuple{2}

Permute the axes of the matrix m, by flipping the elements across the diagonal of the matrix. Differs from LinearAlgebra's transpose in that the operation is not recursive.

Examples

julia> using AxisIndices

julia> AxisIndices.permute_axes(rand(4, 2))
(Base.OneTo(2), Base.OneTo(4))

julia> AxisIndices.permute_axes((Axis(1:4), Axis(1:2)))
(Axis(1:2 => Base.OneTo(2)), Axis(1:4 => Base.OneTo(4)))

AxisIndices.Basics.permute_axes — Method

permute_axes(x::AbstractArray, perms::Tuple) = permute_axes(axes(x), p)
permute_axes(x::NTuple{N}, perms::NTuple{N}) -> NTuple{N}

Returns axes of x in the order of p.

Examples

julia> using AxisIndices

julia> AxisIndices.permute_axes(rand(2, 4, 6), (1, 3, 2))
(Base.OneTo(2), Base.OneTo(6), Base.OneTo(4))

julia> AxisIndices.permute_axes((Axis(1:2), Axis(1:4), Axis(1:6)), (1, 3, 2))
(Axis(1:2 => Base.OneTo(2)), Axis(1:6 => Base.OneTo(6)), Axis(1:4 => Base.OneTo(4)))

AxisIndices.Basics.unwrap_broadcasted — Method

unwrap_broadcasted

Recursively unwraps AbstractAxisIndicess and AxisIndicesArrayStyles. replacing the AbstractAxisIndicess with the wrapped array, and AxisIndicesArrayStyle with the wrapped BroadcastStyle.

AxisIndices.Basics.AxisIndicesArrayStyle — Type

AxisIndicesArrayStyle{S}

This is a BroadcastStyle for AxisIndicesArray's It preserves the dimension names. S should be the BroadcastStyle of the wrapped type.

Base.deleteat! — Method

deleteat!(a::AbstractAxisIndicesVector, arg)

Remove the items corresponding to A[arg], and return the modified a. Subsequent items are shifted to fill the resulting gap. If the axis of a is an AbstractSimpleAxis then it is shortened to match the length of a. If the

inds can be either an iterator or a collection of sorted and unique integer indices, or a boolean vector of the same length as a with true indicating entries to delete.

Examples

julia> using AxisIndices

julia> x = AxisIndicesArray([1, 2, 3, 4]);

julia> deleteat!(x, 3)
AxisIndicesArray{Int64,1,Array{Int64,1}...}
 • dim_1 - SimpleAxis(OneToMRange(3))

  1   1
  2   2
  3   4

julia> x = AxisIndicesArray([1, 2, 3, 4], ["a", "b", "c", "d"]);

julia> deleteat!(x, "c")
AxisIndicesArray{Int64,1,Array{Int64,1}...}
 • dim_1 - Axis(["a", "b", "d"] => OneToMRange(3))

  a   1
  b   2
  d   4

Base.rot180 — Method

rot180(A::AbstractAxisIndices)

Rotate A 180 degrees, along with its axes keys.

Examples

julia> using AxisIndices

julia> a = AxisIndicesArray([1 2; 3 4], ["a", "b"], ["one", "two"])
AxisIndicesArray{Int64,2,Array{Int64,2}...}
 • dim_1 - Axis(["a", "b"] => Base.OneTo(2))
 • dim_2 - Axis(["one", "two"] => Base.OneTo(2))
      one   two
  a     1     2
  b     3     4

julia> b = rot180(a)
AxisIndicesArray{Int64,2,Array{Int64,2}...}
 • dim_1 - Axis(["b", "a"] => OneToMRange(2))
 • dim_2 - Axis(["two", "one"] => OneToMRange(2))
      two   one
  b     4     3
  a     2     1

julia> c = rotr90(rotr90(a))
AxisIndicesArray{Int64,2,Array{Int64,2}...}
 • dim_1 - Axis(["b", "a"] => OneToMRange(2))
 • dim_2 - Axis(["two", "one"] => OneToMRange(2))
      two   one
  b     4     3
  a     2     1

julia> a["a", "one"] == b["a", "one"] == c["a", "one"]
true

Base.rotl90 — Method

rotl90(A::AbstractAxisIndices)

Rotate A left 90 degrees, along with its axes keys.

Examples

julia> using AxisIndices

julia> a = AxisIndicesArray([1 2; 3 4], ["a", "b"], ["one", "two"])
AxisIndicesArray{Int64,2,Array{Int64,2}...}
 • dim_1 - Axis(["a", "b"] => Base.OneTo(2))
 • dim_2 - Axis(["one", "two"] => Base.OneTo(2))
      one   two
  a     1     2
  b     3     4

julia> b = rotl90(a)
AxisIndicesArray{Int64,2,Array{Int64,2}...}
 • dim_1 - Axis(["two", "one"] => OneToMRange(2))
 • dim_2 - Axis(["a", "b"] => OneToMRange(2))
        a   b
  two   2   4
  one   1   3

julia> a["a", "one"] == b["one", "a"]
true

Base.rotr90 — Method

rotr90(A::AbstractAxisIndices)

Rotate A right 90 degrees, along with its axes keys.

Examples

julia> using AxisIndices

julia> a = AxisIndicesArray([1 2; 3 4], ["a", "b"], ["one", "two"])
AxisIndicesArray{Int64,2,Array{Int64,2}...}
 • dim_1 - Axis(["a", "b"] => Base.OneTo(2))
 • dim_2 - Axis(["one", "two"] => Base.OneTo(2))
      one   two
  a     1     2
  b     3     4

julia> b = rotr90(a)
AxisIndicesArray{Int64,2,Array{Int64,2}...}
 • dim_1 - Axis(["one", "two"] => OneToMRange(2))
 • dim_2 - Axis(["b", "a"] => OneToMRange(2))
        b   a
  one   3   1
  two   4   2

julia> a["a", "one"] == b["one", "a"]
true

Math

AxisIndices.Math.covcor_axes — Method

covcor_axes(x, dim) -> NTuple{2}

Returns appropriate axes for a cov or var method on array x.

Examples

julia> using AxisIndices

julia> AxisIndices.covcor_axes(rand(2,4), 1)
(Base.OneTo(4), Base.OneTo(4))

julia> AxisIndices.covcor_axes((Axis(1:4), Axis(1:6)), 2)
(Axis(1:4 => Base.OneTo(4)), Axis(1:4 => Base.OneTo(4)))

julia> AxisIndices.covcor_axes((Axis(1:4), Axis(1:4)), 1)
(Axis(1:4 => Base.OneTo(4)), Axis(1:4 => Base.OneTo(4)))

Each axis is resized to equal to the smallest sized dimension if given a dimensional argument greater than 2.

julia> AxisIndices.covcor_axes((Axis(2:4), Axis(3:4)), 3)
(Axis(3:4 => Base.OneTo(2)), Axis(3:4 => Base.OneTo(2)))

AxisIndices.Math.diagonal_axes — Method

diagonal_axes(x::Tuple{<:AbstractAxis,<:AbstractAxis}) -> collection

Determines the appropriate axis for the resulting vector from a call to diag(::AxisIndicesMatrix). The default behavior is to place the smallest axis at the beginning of a call to combine (e.g., broadcast_axis(small_axis, big_axis)).

Examples

julia> using AxisIndices

julia> AxisIndices.diagonal_axes((Axis(string.(2:5)), SimpleAxis(1:2)))
Axis(["1", "2"] => UnitMRange(1:2))

julia> AxisIndices.diagonal_axes((SimpleAxis(1:3), Axis(string.(2:5))))
Axis(["1", "2", "3"] => UnitMRange(1:3))

AxisIndices.Math.get_factorization — Function

get_factorization(F::Factorization, A::AbstractArray, d::Symbol)

Used internally to compose an AxisIndicesArray for each component of a factor decomposition. F is the result of decomposition, A is an arry (likely a subtype of AbstractAxisIndices), and d is a symbol referring to a component of the factorization.

AxisIndices.Math.get_factorization — Method

get_factorization(F::LQ, A::AbstractArray, d::Symbol)

Returns a component of the LQ decomposition F with the appropriate axes given A.

LQ Factorization

julia> using AxisIndices, LinearAlgebra

julia> m = AxisIndicesArray([1.0 2; 3 4], (Axis(2:3 => Base.OneTo(2)), Axis(3:4 => Base.OneTo(2))));

julia> F = lq(m);

julia> keys.(axes(F.L))
(2:3, Base.OneTo(2))

julia> keys.(axes(F.Q))
(Base.OneTo(2), 3:4)

julia> keys.(axes(F.L * F.Q))
(2:3, 3:4)

AxisIndices.Math.get_factorization — Method

get_factorization(F::LU, A::AbstractArray, d::Symbol)

Returns a component of the LU decomposition F with the appropriate axes given A.

Examples

julia> using AxisIndices, LinearAlgebra

julia> m = AxisIndicesArray([1.0 2; 3 4], (Axis(2:3 => Base.OneTo(2)), Axis(3:4 => Base.OneTo(2))));

julia> F = lu(m);

julia> keys.(axes(F.L))
(2:3, Base.OneTo(2))

julia> keys.(axes(F.U))
(Base.OneTo(2), 3:4)

julia> keys.(axes(F.p))
(2:3,)

julia> keys.(axes(F.P))
(2:3, 2:3)

julia> keys.(axes(F.P * m))
(2:3, 3:4)

julia> keys.(axes(F.L * F.U))
(2:3, 3:4)

AxisIndices.Math.get_factorization — Method

get_factorization(F::SVD, A::AbstractArray, d::Symbol)

Returns a component of the SVD decomposition F with the appropriate axes given A.

Examples

julia> using AxisIndices, LinearAlgebra

julia> m = AxisIndicesArray([1.0 2; 3 4], (Axis(2:3 => Base.OneTo(2)), Axis(3:4 => Base.OneTo(2))));

julia> F = svd(m);

julia> axes(F.U)
(Axis(2:3 => Base.OneTo(2)), SimpleAxis(Base.OneTo(2)))

julia> axes(F.V)
(Axis(3:4 => Base.OneTo(2)), SimpleAxis(Base.OneTo(2)))

julia> axes(F.Vt)
(SimpleAxis(Base.OneTo(2)), Axis(3:4 => Base.OneTo(2)))

julia> axes(F.U * Diagonal(F.S) * F.Vt)
(Axis(2:3 => Base.OneTo(2)), Axis(3:4 => Base.OneTo(2)))

AxisIndices.Math.get_factorization — Method

get_factorization(F::Union{LinearAlgebra.QRCompactWY,QRPivoted,QR}, A::AbstractArray, d::Symbol)

Returns a component of the QR decomposition F with the appropriate axes given A.

Examples

julia> using AxisIndices, LinearAlgebra

julia> m = AxisIndicesArray([1.0 2; 3 4], (Axis(2:3 => Base.OneTo(2)), Axis(3:4 => Base.OneTo(2))));

julia> F = qr(m, Val(true));

julia> keys.(axes(F.Q))
(2:3, Base.OneTo(2))

julia> keys.(axes(F.R))
(Base.OneTo(2), 3:4)

julia> keys.(axes(F.Q * F.R))
(2:3, 3:4)

julia> keys.(axes(F.p))
(2:3,)

julia> keys.(axes(F.P))
(2:3, 2:3)

julia> keys.(axes(F.P * AxisIndicesArray([1.0 2; 3 4], (2:3, 3:4))))
(2:3, 3:4)

AxisIndices.Math.matmul_axes — Method

matmul_axes(a, b) -> Tuple

Returns the appropriate axes for the return of a * b where a and b are a vector or matrix.

Examples

julia> using AxisIndices

julia> axs2, axs1 = (Axis(1:2), Axis(1:4)), (Axis(1:6),);

julia> AxisIndices.matmul_axes(axs2, axs2)
(Axis(1:2 => Base.OneTo(2)), Axis(1:4 => Base.OneTo(4)))

julia> AxisIndices.matmul_axes(axs1, axs2)
(Axis(1:6 => Base.OneTo(6)), Axis(1:4 => Base.OneTo(4)))

julia> AxisIndices.matmul_axes(axs2, axs1)
(Axis(1:2 => Base.OneTo(2)),)

julia> AxisIndices.matmul_axes(axs1, axs1)
()

julia> AxisIndices.matmul_axes(rand(2, 4), rand(4, 2))
(SimpleAxis(Base.OneTo(2)), SimpleAxis(Base.OneTo(2)))

julia> AxisIndices.matmul_axes(CartesianAxes((2,4)), CartesianAxes((4, 2))) == AxisIndices.matmul_axes(rand(2, 4), rand(4, 2))
true

Mapped

Names

AxisIndices.Names.NIArray — Type

NIArray((parent::AbstractArray; kwargs...) = NIArray(parent, kwargs)
NIArray((parent::AbstractArray, axes::NamedTuple{L,AbstractAxes}))

An abbreviated alias and constructor for NamedIndicesArray. If key word arguments are provided then each key word becomes the name of a dimension and its assigned value is sent to the corresponding axis when constructing the underlying AxisIndicesArray.

Examples

julia> using AxisIndices

julia> A = NIArray(reshape(1:24, 2, 3, 4), x=["a", "b"], y =["one", "two", "three"], z=2:5)
NamedDimsArray{Int64,3,Base.ReshapedArray{Int64,3,UnitRange{Int64},Tuple{}}...}
 • x - Axis(["a", "b"] => Base.OneTo(2))
 • y - Axis(["one", "two", "three"] => Base.OneTo(3))
 • z - Axis(2:5 => Base.OneTo(4))
[x, y, z[2]] =
      one   two   three
  a     1     3       5
  b     2     4       6

[x, y, z[3]] =
      one   two   three
  a     7     9      11
  b     8    10      12

[x, y, z[4]] =
      one   two   three
  a    13    15      17
  b    14    16      18

[x, y, z[5]] =
      one   two   three
  a    19    21      23
  b    20    22      24

julia> dimnames(A)
(:x, :y, :z)

julia> axes_keys(A)
(["a", "b"], ["one", "two", "three"], 2:5)

julia> B = A["a", :, :]
NamedDimsArray{Int64,2,Array{Int64,2}...}
 • y - Axis(["one", "two", "three"] => OneToMRange(3))
 • z - Axis(2:5 => Base.OneTo(4))
          2    3    4    5
    one   1    7   13   19
    two   3    9   15   21
  three   5   11   17   23

julia> C = B["one",:]
NamedDimsArray{Int64,1,Array{Int64,1}...}
 • z - Axis(2:5 => Base.OneTo(4))

  2    1
  3    7
  4   13
  5   19

AxisIndices.Names.NamedIndicesArray — Type

NamedIndicesArray

Type alias for NamedDimsArray whose parent array is a subtype of AxisIndicesArray.

PrettyArrays

AxisIndices.PrettyArrays.get_formatters — Method

get_formatters(x)

Returns an argument for the formatters argument of a pretty_table method.

AxisIndices.PrettyArrays.pretty_array — Method

pretty_array([io::IO,] A::AbstractArray[, axs=axes(A)], dimnames=ntuple(i -> Symbol(:dim_, i), N), backend=:text; kwargs...)

Prints to io the array A with the keys key_names along each dimension of A. Printing of multidimensional arrays is accomplished in a similar manner to Array, where the final two dimensions are sliced producing a series of matrices. kwargs... are passed to pretty_table for 1/2D slice produced.

Examples

julia> using AxisIndices

julia> pretty_array(ones(Int, 2,2,2), (Axis(2:3), Axis([:one, :two]), Axis(["a", "b"])), (:x, :y, :z))
[x, y, z[a]] =
      one   two
  2     1     1
  3     1     1

[x, y, z[b]] =
      one   two
  2     1     1
  3     1     1