AxisIndices Styles

TODO

Documentation

AxisIndices.Styles.AxisIndicesStyle — Type

AxisIndicesStyle

Supertype for traits that control the behavior of AxisIndices.to_index and AxisIndices.to_key.

AxisIndices.Styles.BoolElement — Type

BoolElement

A subtype of AxisIndicesStyle for mapping an argument that refers to a single Bool to a single index.

AxisIndices.Styles.BoolsCollection — Type

BoolsCollection

A subtype of AxisIndicesStyle for mapping a collection of Bools to a collection of indices.

AxisIndices.Styles.CartesianElement — Type

CartesianElement

A subtype of AxisIndicesStyle for mapping an argument that refers to a CartesianIndex to a single index.

AxisIndices.Styles.CartesianIndexCollection — Type

CartesianIndexCollection

A subtype of AxisIndicesStyle for propagating an argument to a collection of indices.

AxisIndices.Styles.IndexElement — Type

IndexElement

A subtype of AxisIndicesStyle for propagating an argument to a single index.

AxisIndices.Styles.IndexEquals — Type

IndexEquals

A subtype of AxisIndicesStyle for mapping a single index in a isequal(index) argument to a single index.

AxisIndices.Styles.Indices — Type

Indices(arg)

Forces arg to refer to indices when indexing.

AxisIndices.Styles.IndicesCollection — Type

IndicesCollection

A subtype of AxisIndicesStyle for propagating an argument to a collection of indices.

AxisIndices.Styles.IndicesFix2 — Type

IndicesFix2

A subtype of AxisIndicesStyle for mapping all indices from fixed argument (e.g., >(indices)) to the corresponding collection of indices.

AxisIndices.Styles.IndicesIn — Type

IndicesIn

A subtype of AxisIndicesStyle for mapping all keys given in(keys) to a collection of indices.

AxisIndices.Styles.IntervalCollection — Type

IntervalCollection

A subtype of AxisIndicesStyle for mapping an interval argument (start..stop) from keys within said interval to a collection of indices.

AxisIndices.Styles.KeyElement — Type

KeyElement

A subtype of AxisIndicesStyle for mapping an argument that refers to a single key to a single index.

AxisIndices.Styles.KeyEquals — Type

KeyEquals

A subtype of AxisIndicesStyle for mapping a single key in a isequal(key) argument to a single index.

AxisIndices.Styles.KeyedStyle — Type

KeyedStyle{S}

A subtype of AxisIndicesStyle indicating that the axis is a always defaults to key based indexing.

AxisIndices.Styles.Keys — Type

Keys(arg)

Forces arg to refer to keys when indexing.

AxisIndices.Styles.KeysCollection — Type

KeysCollection

A subtype of AxisIndicesStyle for mapping a collection of keys a collection of indices.

AxisIndices.Styles.KeysFix2 — Type

KeysFix2

A subtype of AxisIndicesStyle for mapping all keys from fixed argument (e.g., >(key)) to the corresponding collection of indices.

AxisIndices.Styles.KeysIn — Type

KeysIn

A subtype of AxisIndicesStyle for mapping all keys given in(keys) to a collection of indices.

AxisIndices.Styles.SliceCollection — Type

SliceCollection

A subtype of AxisIndicesStyle indicating that the entire axis should be propagated.

AxisIndices.Styles.is_collection — Method

is_collection(x) -> Bool

Whether x is an AxisIndicesStyle, returns true if it's used to index a collection of indices.

AxisIndices.Styles.is_element — Method

is_element(x) -> Bool

Whether x is an AxisIndicesStyle, returns true if it's used to index a single element.

AxisIndices.Styles.is_index — Method

is_index(x) -> Bool

Whether x is an AxisIndicesStyle, returns true if it's used to search the indices space.

AxisIndices.Styles.is_key — Method

is_key(x) -> Bool

Whether x is an AxisIndicesStyle, returns true if it's used to search the keys space.